### 2023-11-02 #6012 @editorialbot
[PRE REVIEW]: Catalyst: a Python JIT compiler for auto-differentiable hybrid quantum programs
**Submitting author:** @josh146 (<a href="http://orcid.org/0000-0003-2640-0734">Joshua Izaac</a>)
**Repository:** https://github.com/PennyLaneAI/catalyst
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.3.1
