### 2023-11-02 #6012 @danielskatz
@josh146 - thanks for your submission. I'll look for an editor next.

While I do so, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, please proofread how the references look in the pdf. I see at least one where the authors are not formatted correctly, which can be fixed by protecting a string in the .bib file using {}s.
