### 2023-11-02 80c0bed
Only dump if keep_intrmediate is set
