+++ paper/paper.bib
@@ -82,7 +82,8 @@
     publisher = {APS},
     volume = 104,
     number = 5,
-    pages = {052402}
+    pages = {052402},
+    doi = {10.1103/physreva.104.052402}
 }
 
 @software{jax,
@@ -189,7 +190,8 @@
     publisher = {APS},
     volume = 99,
     number = 3,
-    pages = {032331}
+    pages = {032331},
+    doi = {10.1103/physreva.99.032331}
 }
 
 @article{moldovan2018autograph,
