### 2023-11-02 a593e95
Address review suggestions: assert on the matching return value shapes
