+++ paper/paper.bib
@@ -43,7 +43,7 @@
 
 @software{cuquantum,
     title = {NVIDIA/cuQuantum: cuQuantum v22.03.0},
-    author = {NVIDIA cuQuantum team},
+    author = {{NVIDIA cuQuantum team}},
     year = 2022,
     month = mar,
     publisher = {Zenodo},
