### 2023-11-02 3a6971c
fix footnote type
