### 2023-11-02 b3787a3
[Frontend] Implement an if-elif-else result type unification (#333)
Return values of conditional functions no longer need to be of exactly
the same type. Catalyst would apply type promotion to branch return
values if their types don't match.

[sc-41332]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
