### 2023-11-02 6c7919e
[MLIR] Improve the compiler driver failure reporting (#349)
In this PR we add the following functionality to the CompilerDriver:
* [x] Print the name of failed pass along with the name of the pipeline
* [x] Dump the failing IR into the workspace with the "_FAILED" suffix
