+++ runtime/Makefile
@@ -50,7 +50,7 @@ configure:
 		-DCMAKE_CXX_COMPILER=$(CXX_COMPILER) \
 		-DCMAKE_C_COMPILER_LAUNCHER=$(COMPILER_LAUNCHER) \
 		-DCMAKE_CXX_COMPILER_LAUNCHER=$(COMPILER_LAUNCHER) \
-		-DENABLE_WARNINGS=$(ENABLE_WARNINGS) \
+		-DRUNTIME_ENABLE_WARNINGS=$(ENABLE_WARNINGS) \
 		-DENABLE_CODE_COVERAGE=$(CODE_COVERAGE) \
 		-DBUILD_QIR_STDLIB_FROM_SRC=$(BUILD_QIR_STDLIB_FROM_SRC) \
 		$(CMAKE_ARGS)
