+++ runtime/CMakeLists.txt
@@ -1,6 +1,7 @@
 cmake_minimum_required(VERSION 3.20)
 
 project(catalyst_runtime)
+include(FetchContent)
 
 set(CMAKE_CXX_STANDARD  20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -26,9 +27,33 @@ set(backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/common")
 set(QIRSTDLIB_VERSION "0.4.0")
 set(QIRSTDLIB_REPO "https://github.com/qir-alliance/qir-runner")
 
+# Get LLVM hash to target from source tree.
+file(READ ../.dep-versions DEPENDENCY_VERSIONS)
+string(REGEX MATCH "llvm=([0-9a-f]+)" _ ${DEPENDENCY_VERSIONS})
+set(LLVM_HASH ${CMAKE_MATCH_1})
+message(STATUS "Detected LLVM version - ${LLVM_HASH}")
+
+FetchContent_Declare(
+    MLIRRunnerUtils
+    URL                 https://raw.githubusercontent.com/llvm/llvm-project/${LLVM_HASH}/mlir/include/mlir/ExecutionEngine/RunnerUtils.h
+    DOWNLOAD_NO_EXTRACT True
+    SOURCE_DIR        mlir/ExecutionEngine
+)
+
+FetchContent_Declare(
+    MLIRCRunnerUtils
+    URL                 https://raw.githubusercontent.com/llvm/llvm-project/${LLVM_HASH}/mlir/include/mlir/ExecutionEngine/CRunnerUtils.h
+    DOWNLOAD_NO_EXTRACT True
+    SOURCE_DIR          mlir/ExecutionEngine
+)
+
+if(RUNTIME_ENABLE_WARNINGS)
+    message(STATUS "Building with compiler warnings as errors enabled.")
+    add_compile_options(-Werror -Wextra -Wall)
+endif()
+
 if(BUILD_QIR_STDLIB_FROM_SRC)
     message(STATUS "Building qir-stdlib from source.")
-    include(FetchContent)
 
     FetchContent_Declare(
         qir_runner
