### 2023-11-03 a54ed5f
Update Python versions for macOS arm64
