### 2023-11-03 45248d8
Draft catalyst.empty
