+++ mlir/include/Catalyst/IR/CatalystOps.h
@@ -21,6 +21,8 @@
 #include "mlir/Interfaces/InferTypeOpInterface.h"
 
 #include "Catalyst/IR/CatalystDialect.h"
+#include "mlir/Dialect/Quant/QuantTypes.h"
+#include "stablehlo/dialect/Base.h"
 
 #define GET_OP_CLASSES
 #include "Catalyst/IR/CatalystOps.h.inc"
