### 2023-11-03 f0aaff7
Add TensorInit patterns file
