+++ runtime/lib/capi/MemRefUtils.hpp
@@ -15,14 +15,6 @@
 #pragma once
 
 #include <cstddef>
-#include <sstream>
-#include <type_traits>
-#include <variant>
-#include <vector>
-
-#include "DataView.hpp"
-#include "Exception.hpp"
-#include "Types.h"
 
 extern "C" {
 void *_mlir_memref_to_llvm_alloc(size_t size);
@@ -39,137 +31,3 @@ template <typename T, size_t R> struct MemRefT {
     size_t sizes[R];
     size_t strides[R];
 };
-
-typedef std::variant<bool *, int16_t *, int32_t *, int64_t *, float *, double *, CplxT_float *,
-                     CplxT_double *>
-    NumericPtr;
-
-struct DynamicMemRefT {
-    NumericPtr data_allocated;
-    NumericPtr data_aligned;
-    size_t offset;
-    std::vector<size_t> sizes;
-    std::vector<size_t> strides;
-    std::ostringstream in_str;
-};
-
-template <typename T, typename TPtr>
-static inline void _gen_oss_dyn_memref(DynamicMemRefT &newMemref, int64_t rank)
-{
-    const auto data = std::get<TPtr>(newMemref.data_aligned);
-    const auto sizes = newMemref.sizes;
-
-    if (rank > 0) {
-        newMemref.in_str << "[ ";
-    }
-
-    if (rank == 0) {
-        if constexpr (std::is_same_v<T, CplxT_float> || std::is_same_v<T, CplxT_double>) {
-            newMemref.in_str << data[0].real << "+" << data[0].imag << "j";
-        }
-        else {
-            newMemref.in_str << data[0];
-        }
-    }
-    else if (rank == 1) {
-        DataView<T, 1> view(data, newMemref.offset, newMemref.sizes.data(),
-                            newMemref.strides.data());
-        for (const auto elem : view) {
-            if constexpr (std::is_same_v<T, CplxT_float> || std::is_same_v<T, CplxT_double>) {
-                newMemref.in_str << elem.real << "+" << elem.imag << "j";
-            }
-            else {
-                newMemref.in_str << elem << " ";
-            }
-        }
-    }
-    else if (rank == 2) {
-        DataView<T, 2> view(data, newMemref.offset, newMemref.sizes.data(),
-                            newMemref.strides.data());
-        size_t col_index = 0;
-        newMemref.in_str << "[ ";
-        for (const auto elem : view) {
-            if (col_index == sizes[1]) {
-                newMemref.in_str << "]";
-                col_index = 0;
-                newMemref.in_str << ", [ ";
-            }
-            if constexpr (std::is_same_v<T, CplxT_float> || std::is_same_v<T, CplxT_double>) {
-                newMemref.in_str << elem.real << "+" << elem.imag << "j";
-            }
-            else {
-                newMemref.in_str << elem << " ";
-            }
-            col_index++;
-        }
-        newMemref.in_str << "] ";
-    }
-    else {
-        newMemref.in_str << "MemRef of Rank > 2 ";
-    }
-
-    if (rank > 0) {
-        newMemref.in_str << "]";
-    }
-}
-
-inline DynamicMemRefT get_dynamic_memref(OpaqueMemRefT memref)
-{
-    DynamicMemRefT newMemref;
-    const auto rank = memref.rank;
-
-    newMemref.offset = *(size_t *)((char *)memref.descriptor + 2 * sizeof(void *));
-    size_t *sizesPtr = (size_t *)((char *)memref.descriptor + 2 * sizeof(void *) + sizeof(size_t));
-    size_t *stridesPtr = (size_t *)((char *)memref.descriptor + 2 * sizeof(void *) +
[…]
