### 2023-11-03 217c8ee
[Frontend] Add additional type unification test (#351)
Add additional frontend test checking the if-else type unification
against mismatching pytrees
