### 2023-11-03 f7263c7
Update docstring and changelog
