+++ mlir/lib/Mitigation/Transforms/MitigationMethods/Zne.cpp
@@ -16,6 +16,7 @@
 
 #include <algorithm>
 #include <deque>
+#include <iostream>
 #include <sstream>
 #include <vector>
 
@@ -26,8 +27,7 @@
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Index/IR/IndexOps.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
-
-#include "Quantum/Utils/RemoveQuantumMeasurements.h"
+#include "mlir/IR/IRMapping.h"
 
 namespace catalyst {
 namespace mitigation {
@@ -66,9 +66,8 @@ void ZneLowering::rewrite(mitigation::ZneOp op, PatternRewriter &rewriter) const
     //                                                  builder.create<scf::YieldOp>(loc, iterArgs);
     //                                              })
     //                          .getResult(0);
-    func::FuncOp circuit =
-        SymbolTable::lookupNearestSymbolFrom<func::FuncOp>(op, op.getCalleeAttr());
-    auto resultsValues = rewriter.create<func::CallOp>(loc, circuit, op.getArgs());
+
+    auto resultsValues = rewriter.create<func::CallOp>(loc, foldedCircuit, op.getArgs());
     // Call the folded circuit
     rewriter.replaceOp(op, resultsValues);
 }
@@ -83,78 +82,234 @@ FlatSymbolRefAttr ZneLowering::getOrInsertFoldedCircuit(Location loc, PatternRew
     rewriter.setInsertionPointToStart(moduleOp.getBody());
     // Original function
     func::FuncOp fnOp = SymbolTable::lookupNearestSymbolFrom<func::FuncOp>(op, op.getCalleeAttr());
+    TypeRange originalTypes = op.getArgs().getTypes();
 
-    // Function Name
-    std::string fnName = op.getCallee().str() + ".folded";
+    Type qregType = quantum::QuregType::get(rewriter.getContext());
+    // Alloc function
+    std::string fnNameAlloc = op.getCallee().str() + ".alloc";
+    Type indexType = rewriter.getI64Type();
+    FunctionType fnAllocType = FunctionType::get(ctx, /*inputs=*/
+                                                 indexType,
+                                                 /*outputs=*/qregType);
 
-    // Get callee arg types
-    auto originalTypes = op.getArgs().getTypes();
-    SmallVector<Type> typesFolded = {originalTypes.begin(), originalTypes.end()};
-    typesFolded.push_back(scalarType);
+    func::FuncOp fnAlloc = rewriter.create<func::FuncOp>(loc, fnNameAlloc, fnAllocType);
+    fnAlloc.setPrivate();
 
-    // Get the scalar factor type
+    Block *allocBloc = fnAlloc.addEntryBlock();
+    rewriter.setInsertionPointToStart(allocBloc);
+    Value nQubits = allocBloc->getArgument(0);
+    IntegerAttr intAttr;
+    auto qreg = rewriter.create<quantum::AllocOp>(loc, qregType, nQubits, intAttr);
+    rewriter.create<func::ReturnOp>(loc, qreg.getResult());
 
-    // Get the arguments and outputs types from the original block
-    FunctionType updateFnType = FunctionType::get(ctx, /*inputs=*/
-                                                  typesFolded,
-                                                  /*outputs=*/op.getResultTypes());
+    // Modify the function to take a qreg
+    std::string fnNameWithQregArg = op.getCallee().str() + ".qregarg";
[…]
