### 2023-11-03 01ab781
For loop for the index (fn, adjoint)**n
