+++ mlir/lib/Mitigation/Transforms/MitigationMethods/Zne.cpp
@@ -27,6 +27,7 @@
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/Index/IR/IndexOps.h"
 #include "mlir/Dialect/SCF/IR/SCF.h"
+#include "mlir/Dialect/Tensor/IR/Tensor.h"
 #include "mlir/IR/IRMapping.h"
 
 namespace catalyst {
@@ -38,8 +39,8 @@ void ZneLowering::rewrite(mitigation::ZneOp op, PatternRewriter &rewriter) const
 {
     Location loc = op.getLoc();
 
-    auto scalarFactor = op.getScalarFactors();
-    RankedTensorType scalarFactorType = scalarFactor.getType().cast<RankedTensorType>();
+    auto scalarFactors = op.getScalarFactors();
+    RankedTensorType scalarFactorType = scalarFactors.getType().cast<RankedTensorType>();
     auto sizeInt = scalarFactorType.getDimSize(0);
 
     // Create the folded circuit function
@@ -66,8 +67,11 @@ void ZneLowering::rewrite(mitigation::ZneOp op, PatternRewriter &rewriter) const
     //                                                  builder.create<scf::YieldOp>(loc, iterArgs);
     //                                              })
     //                          .getResult(0);
-
-    auto resultsValues = rewriter.create<func::CallOp>(loc, foldedCircuit, op.getArgs());
+    std::vector<Value> newArgs{op.getArgs().begin(), op.getArgs().end()};
+    Value indexValue = rewriter.create<index::ConstantOp>(loc, 0);
+    auto scalarFactor = rewriter.create<tensor::ExtractOp>(loc, scalarFactors, indexValue);
+    newArgs.push_back(scalarFactor);
+    auto resultsValues = rewriter.create<func::CallOp>(loc, foldedCircuit, newArgs);
     // Call the folded circuit
     rewriter.replaceOp(op, resultsValues);
 }
@@ -87,9 +91,9 @@ FlatSymbolRefAttr ZneLowering::getOrInsertFoldedCircuit(Location loc, PatternRew
     Type qregType = quantum::QuregType::get(rewriter.getContext());
     // Alloc function
     std::string fnNameAlloc = op.getCallee().str() + ".alloc";
-    Type indexType = rewriter.getI64Type();
+    Type i64Type = rewriter.getI64Type();
     FunctionType fnAllocType = FunctionType::get(ctx, /*inputs=*/
-                                                 indexType,
+                                                 i64Type,
                                                  /*outputs=*/qregType);
 
     func::FuncOp fnAlloc = rewriter.create<func::FuncOp>(loc, fnNameAlloc, fnAllocType);
@@ -161,8 +165,11 @@ FlatSymbolRefAttr ZneLowering::getOrInsertFoldedCircuit(Location loc, PatternRew
     // Function folded
     rewriter.setInsertionPointToStart(moduleOp.getBody());
     std::string fnNameFolded = op.getCallee().str() + ".folded";
+    SmallVector<Type> typesFolded = {originalTypes.begin(), originalTypes.end()};
+    Type indexType = rewriter.getIndexType();
+    typesFolded.push_back(indexType);
     FunctionType fnFoldedType = FunctionType::get(ctx, /*inputs=*/
-                                                  originalTypes,
+                                                  typesFolded,
                                                   /*outputs=*/fnOp.getResultTypes());
 
     func::FuncOp fnFoldedOp = rewriter.create<func::FuncOp>(loc, fnNameFolded, fnFoldedType);
@@ -174,22 +181,42 @@ FlatSymbolRefAttr ZneLowering::getOrInsertFoldedCircuit(Location loc, PatternRew
[…]
