### 2023-11-03 8f71da8
Build qir-stdlib from src on macOS for version compatibilities
