+++ .github/workflows/build-wheel-macos-arm64.yaml
@@ -8,7 +8,7 @@ on:
   workflow_dispatch:
 
 env:
-  MACOSX_DEPLOYMENT_TARGET: 11.0
+  MACOSX_DEPLOYMENT_TARGET: 13.0
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -253,6 +253,7 @@ jobs:
 
     - name: Test Catalyst-Runtime
       run: |
+        python${{ matrix.python_version }} -m pip install amazon-braket-pennylane-plugin
         cmake --build runtime-build --target runner_tests
         ./runtime-build/tests/runner_tests
 
