+++ frontend/catalyst/pennylane_extensions.py
@@ -897,6 +897,7 @@ class ForLoop(HybridOp):
                 body_jaxpr=ClosedJaxpr(convert_constvars_jaxpr(jaxpr), ()),
                 body_nconsts=len(consts),
                 apply_reverse_transform=apply_reverse_transform,
+                _id=self._id,
             )
         )
         return qrp2
@@ -946,6 +947,7 @@ class Cond(HybridOp):
                 trace,
                 *(op.in_classical_tracers + combined_consts + [qreg]),
                 branch_jaxprs=unify_result_types(jaxprs2),
+                _id=self._id,
             )
         )
         return qrp2
@@ -984,6 +986,7 @@ class WhileLoop(HybridOp):
                 body_jaxpr=ClosedJaxpr(convert_constvars_jaxpr(body_jaxpr), ()),
                 cond_nconsts=len(cond_consts),
                 body_nconsts=len(body_consts),
+                _id=self._id,
             )
         )
         return qrp2
@@ -1113,6 +1116,7 @@ class CondCallable:
         self.preds = [pred]
         self.branch_fns = [true_fn]
         self.otherwise_fn = lambda: None
+        self._id = 0
 
     def else_if(self, pred):
         """
@@ -1171,7 +1175,7 @@ class CondCallable:
         _check_cond_same_shapes(out_trees, out_avals)
         res_avals = list(map(shaped_abstractify, res_classical_tracers))
         out_classical_tracers = [new_inner_tracer(outer_trace, aval) for aval in res_avals]
-        Cond(in_classical_tracers, out_classical_tracers, regions)
+        Cond(in_classical_tracers, out_classical_tracers, regions, _id=self._id)
         return tree_unflatten(out_tree(), out_classical_tracers)
 
     def _call_with_classical_ctx(self):
@@ -1182,7 +1186,11 @@ class CondCallable:
         )
         _check_cond_same_shapes(out_trees, [j.out_avals for j in branch_jaxprs])
         branch_jaxprs = unify_result_types(branch_jaxprs)
-        out_classical_tracers = qcond_p.bind(*(self.preds + consts), branch_jaxprs=branch_jaxprs)
+        out_classical_tracers = qcond_p.bind(
+            *(self.preds + consts),
+            branch_jaxprs=branch_jaxprs,
+            _id=self._id,
+        )
         return tree_unflatten(out_trees[0], out_classical_tracers)
 
     def _call_during_interpretation(self):
@@ -1200,6 +1208,9 @@ class CondCallable:
         else:
             assert mode == EvaluationMode.INTERPRETATION, f"Unsupported evaluation mode {mode}"
             return self._call_during_interpretation()
+
+    def _set_id(self, x):
+        self._id = x
 
 
 def cond(pred: DynamicJaxprTracer):
@@ -1379,6 +1390,94 @@ def cond(pred: DynamicJaxprTracer):
     return _decorator
 
 
+class ForLoopCallable:
+    """A callable wrapper for functional for loops."""
+
+    def __init__(self, lower_bound, upper_bound, step, body_fn):
+        self.lower_bound = lower_bound
+        self.upper_bound = upper_bound
+        self.step = step
+        self.body_fn = body_fn
+        self._id = 0
+
+    def __call__(self, *init_state):
+        def _call_with_quantum_ctx(ctx: JaxTracingContext):
+            quantum_tape = QuantumTape()
+            outer_trace = ctx.trace
+            with EvaluationContext.frame_tracing_context(ctx) as inner_trace:
+                in_classical_tracers = [
+                    self.lower_bound,
+                    self.upper_bound,
+                    self.step,
+                    self.lower_bound,
+                ] + tree_flatten(init_state)[0]
[…]
