### 2023-11-06 fcd876d
Add ability to discover current JAXPR frame
