### 2023-11-06 e1e0548
Implement zeros, empty and ones
