+++ mlir/include/Catalyst/IR/CatalystOps.h
@@ -21,6 +21,7 @@
 #include "mlir/Interfaces/InferTypeOpInterface.h"
 
 #include "Catalyst/IR/CatalystDialect.h"
+#include "mlir/Dialect/Complex/IR/Complex.h"
 #include "mlir/Dialect/Quant/QuantTypes.h"
 #include "stablehlo/dialect/Base.h"
 
