+++ mlir/python/dialects/catalyst.py
@@ -1,0 +1,18 @@
+# Copyright 2022-2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""MLIR Dialect for Catalyst dialect."""
+
+# from .._mlir_libs._quantumDialects.catalyst import *  # noqa: F401
+from ._catalyst_ops_gen import *  # noqa: F401
