### 2023-11-06 649d046
Add input/output checks for while loops
