+++ mlir/lib/Catalyst/Transforms/CMakeLists.txt
@@ -3,6 +3,7 @@ add_mlir_library(MLIRCatalystTransforms
     scatter_lowering.cpp
     ScatterPatterns.cpp
     RegisterAllPasses.cpp
+    tensor_init_lowering.cpp
     TensorInitPatterns.cpp
 
     DEPENDS
