### 2023-11-06 c18ac31
Revert ScatterLowering changes
