+++ mlir/lib/Catalyst/Transforms/scatter_lowering.cpp
@@ -26,10 +26,6 @@
 #include "mlir/Pass/Pass.h"
 #include "mlir/Transforms/GreedyPatternRewriteDriver.h"
 
-#include "mlir/Dialect/Arith/IR/Arith.h"
-#include "mlir/Dialect/Tensor/IR/Tensor.h"
-#include "mlir/Transforms/DialectConversion.h"
-
 #include "Catalyst/Transforms/Patterns.h"
 
 using namespace llvm;
