### 2023-11-06 13eff5b
Revert ScatterLowering changes
