+++ mlir/lib/Catalyst/Transforms/scatter_lowering.cpp
@@ -58,3 +58,4 @@ std::unique_ptr<Pass> createScatterLoweringPass()
 }
 
 } // namespace catalyst
+
