+++ mlir/python/dialects/catalyst.py
@@ -14,5 +14,5 @@
 
 """MLIR Dialect for Catalyst dialect."""
 
-# from .._mlir_libs._quantumDialects.catalyst import *  # noqa: F401
+from .._mlir_libs._quantumDialects.catalyst import *  # noqa: F401
 from ._catalyst_ops_gen import *  # noqa: F401
