+++ doc/conf.py
@@ -88,6 +88,7 @@ MOCK_MODULES = [
     "mlir_quantum.dialects.arith",
     "mlir_quantum.dialects.tensor",
     "mlir_quantum.dialects.scf",
+    "mlir_quantum.dialects.catalyst",
     "mlir_quantum.dialects.quantum",
     "mlir_quantum.dialects.gradient",
     "mlir_quantum.compiler_driver",
