### 2023-11-06 916a019
Fix malformed JAXPR after Python fallback
