+++ frontend/test/pytest/test_autograph.py
@@ -1389,6 +1389,109 @@ class TestWhileLoops:
 
 
 @pytest.mark.tf
+class TestFallback:
+    """Test that Python fallbacks still produce correct results."""
+
+    def test_postbinding_errors_for(self):
+        """Test that errors are handled correctly if they trigger after the JAX primitive binding
+        step (e.g. during result verification), and that no errors occur after the AG tracing step
+        (e.g. during lowering) because of malformed primitives in the JAXPR.
+        This test ensures the primitive identification and removal works correctly on fallback. In
+        this case, the loop primitive should be removed since the exception happens after binding.
+        """
+
+        def f():
+            arr = jnp.array([1, 2])
+            for _ in range(2):
+                # fails result verification, will trigger fallback
+                # would raise an error during lowering if left in the JAXPR
+                arr = jnp.kron(arr, arr)
+            return arr
+
+        with pytest.warns(
+            UserWarning, match="Tracing of an AutoGraph converted for loop failed with an exception"
+        ):
+            f_jit = qjit(autograph=True)(f)
+
+        arr = jnp.array([1, 2])
+        expected = jnp.kron(*([jnp.kron(arr, arr)] * 2))
+        assert np.allclose(f_jit(), expected)
+
+    def test_prebinding_errors_for(self):
+        """Test that errors are handled correctly if they trigger before the JAX primitive binding
+        step (e.g. during argument verification).
+        This test ensures the primitive identification and removal works correctly on fallback. In
+        this case no primitive should be removed since the exception happens before binding.
+        """
+
+        def f():
+            string = "hi"
+            arr = jnp.array([1, 2])
+            for _ in range(2):
+                arr = arr + 3
+            for i in range(1, 4):
+                string = string * i  # fails return type verification, triggers fallback
+            return arr, len(string)
+
+        with pytest.warns(
+            UserWarning, match="Tracing of an AutoGraph converted for loop failed with an exception"
+        ):
+            f_jit = qjit(autograph=True)(f)
+
+        results = f_jit()
+        assert np.allclose(results[0], [7, 8])
+        assert results[1] == (2) * 1 * 2 * 3  # i = range(1, 4)
+
+    def test_postbinding_errors_while(self):
+        """Test that errors are handled correctly if they trigger after the JAX primitive binding
+        step (e.g. during result verification), and that no errors occur after the AG tracing step
+        (e.g. during lowering) because of malformed primitives in the JAXPR.
+        This test ensures the primitive identification and removal works correctly on fallback. In
+        this case, the loop primitive should be removed since the exception happens after binding.
+        """
+
+        @qjit(autograph=True)
+        def f():
+            arr = jnp.array([1, 2])
+            while len(arr) < 16:
+                # fails result verification, will trigger fallback
+                # would raise an error during lowering if left in the JAXPR
+                arr = jnp.kron(arr, arr)
+            return arr
+
+        arr = jnp.array([1, 2])
+        result = f()
+        expected = jnp.kron(*([jnp.kron(arr, arr)] * 2))
+        assert np.allclose(result, expected)
+
+    def test_prebinding_errors_while(self):
+        """Test that errors are handled correctly if they trigger before the JAX primitive binding
+        step (e.g. during argument verification).
+        This test ensures the primitive identification and removal works correctly on fallback. In
+        this case no primitive should be removed since the exception happens before binding.
+        """
+
+        @qjit(autograph=True)
+        def f():
+            string = "hi"
+            arr = jnp.array([1, 2])
+            i = 1
+
+            while arr[0] < 7:
+                arr = arr + 3
+
+            while len(string) < 12:
[…]
