### 2023-11-06 ee0bcf9
review: Improve wording
Co-authored-by: Josh Izaac <josh146@gmail.com>
