### 2023-11-06 15029fd
review: copyright header & obsolete import
Co-authored-by: Ali Asadi <ali@xanadu.ai>
