+++ runtime/include/Types.h
@@ -19,7 +19,6 @@
 #include <cmath>
 #include <cstdint>
 #include <limits>
-#include <ostream>
 
 #ifdef __cplusplus
 extern "C" {
