### 2023-11-06 a571389
review: add note about f-strings
