+++ frontend/catalyst/pennylane_extensions.py
@@ -1742,8 +1742,8 @@ def debug_print(x):
     Enables printing of numeric values at runtime. Can also print objects or strings as constants.
 
     Args:
-        x (jax.Arrar, Any): A single jax array whose numeric values are printed at runtime, or any object whose
-           string representation will be treated as a constant and printed at runtime.
+        x (jax.Arrar, Any): A single jax array whose numeric values are printed at runtime, or any
+            object whose string representation will be treated as a constant and printed at runtime.
 
     **Example**
 
