+++ mlir/Makefile
@@ -1,4 +1,4 @@
-PYTHON?=python3
+PYTHON?=$(shell which python3)
 C_COMPILER?=clang
 CXX_COMPILER?=clang++
 COMPILER_LAUNCHER?=$(shell which ccache)
