+++ Makefile
@@ -1,4 +1,4 @@
-PYTHON ?= python3
+PYTHON ?= $(shell which python3)
 C_COMPILER ?= clang
 CXX_COMPILER ?= clang++
 BLACKVERSIONMAJOR := $(shell black --version 2> /dev/null | head -n1 | awk '{ print $$2 }' | cut -d. -f1)
