### 2023-11-07 232de5f
Make memref descriptor printing optional
