+++ frontend/catalyst/pennylane_extensions.py
@@ -1736,7 +1736,7 @@ def measure(wires) -> DynamicJaxprTracer:
 
 
 # Note that "print" would alias the built-in print function.
-def debug_print(x):
+def debug_print(x, memref=False):
     """A :func:`qjit` compatible print function for printing values at runtime.
 
     Enables printing of numeric values at runtime. Can also print objects or strings as constants.
@@ -1744,6 +1744,10 @@ def debug_print(x):
     Args:
         x (jax.Arrar, Any): A single jax array whose numeric values are printed at runtime, or any
             object whose string representation will be treated as a constant and printed at runtime.
+        memref (Optional[bool]): When set to ``True``, additional information about how the array is
+            stored in memory is printed, via the so-called "memref" descriptor. This includes the
+            base memory address of the data buffer, as well as the rank of the array, the size of
+            each dimension, and the strides between elements.
 
     **Example**
 
@@ -1751,18 +1755,13 @@ def debug_print(x):
 
         @qjit
         def func(x: float):
-            debug_print(x)
+            debug_print(x, memref=True)
             debug_print("exit")
 
     >>> func(jnp.array(0.43))
     Unranked Memref base@ = 0x5629ff2b6680 rank = 0 offset = 0 sizes = [] strides = [] data =
     [0.43]
     exit
-
-    Note that the output for arrays also includes information about how the array is stored in
-    memory, via the so-called "memref" descriptor. This includes the base memory address of the data
-    buffer, as well as the rank of the array, the size of each dimension, and the strides between
-    elements.
 
     Outside a :func:`qjit` compiled function the operation falls back to the Python print statement.
 
@@ -1774,7 +1773,7 @@ def debug_print(x):
     """
     if EvaluationContext.is_tracing():
         if isinstance(x, jax.core.Tracer):
-            print_p.bind(x)
+            print_p.bind(x, memref=memref)
         else:
             print_p.bind(string=str(x))
     else:
