### 2023-11-07 595a519
Do not use -Wextra with Lightning
Co-authored-by: Ali Asadi <ali@xanadu.ai>
