### 2023-11-07 2460cf1
Update C++ tests
