### 2023-11-07 4aa4f30
Update aligned_alloc total-size
