+++ frontend/test/pytest/test_tensor_init.py
@@ -15,14 +15,16 @@
 """Test integration for the lowering of catalyst.tensor_init."""
 
 import numpy as np
+import pennylane as qml
 import pytest
 from jax import numpy as jnp
 from numpy import array_equal
+from numpy.testing import assert_allclose
 
 from catalyst import empty, ones, qjit, zeros
 
-DTYPES = [float, int, jnp.float32, jnp.float64, jnp.int8, jnp.int16, np.float32, np.float64]
-SHAPES = [(2, 3, 1), ()]
+DTYPES = [float, int, jnp.float32, jnp.float64, jnp.int8, jnp.int16, "float32", np.float64]
+SHAPES = [3, (2, 3, 1), (), jnp.array([2, 1], dtype=int)]
 
 
 def _assert_equal(a, b):
@@ -34,47 +36,65 @@ def _assert_equal(a, b):
 
 @pytest.mark.parametrize("dtype", DTYPES)
 @pytest.mark.parametrize("shape", SHAPES)
-def test_ones(dtype, shape):
-    """Test catalyst.ones primitive"""
+def test_interpretation(shape, dtype):
+    """Test that tensor primitive work in the interpretation mode"""
 
-    @qjit
+    _assert_equal((lambda: zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal((lambda: ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
+
     def f(s):
-        return ones(shape=s, dtype=dtype)
+        return empty(shape=s, dtype=dtype)
 
-    _assert_equal(f(shape), jnp.ones(shape, dtype=dtype))
+    res = f(shape)
+    assert_allclose(res.shape, shape)
+    assert res.dtype == dtype
 
 
 @pytest.mark.parametrize("dtype", DTYPES)
 @pytest.mark.parametrize("shape", SHAPES)
-def test_zeros(dtype, shape):
-    """Test catalyst.zeros primitive"""
+def test_classical_tracing(shape, dtype):
+    """Test that tensor primitive work in the classical tracing mode"""
 
-    @qjit
-    def f(s):
-        return zeros(shape=s, dtype=dtype)
-
-    _assert_equal(f(shape), jnp.zeros(shape, dtype=dtype))
-
-
-@pytest.mark.parametrize("dtype", DTYPES)
-@pytest.mark.parametrize("shape", SHAPES)
-def test_empty(dtype, shape):
-    """Test catalyst.empty primitive"""
+    _assert_equal(qjit(lambda: zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal(qjit(lambda: ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
 
     @qjit
     def f(s):
         return empty(shape=s, dtype=dtype)
 
     res = f(shape)
-    assert res.shape == shape
+    assert_allclose(res.shape, shape)
     assert res.dtype == dtype
 
 
+def test_quantum_tracing():
+    """Test that catalyst tensor primitive is compatible with quantum tracing mode"""
+
+    @qjit(autograph=True)
+    @qml.qnode(qml.device("lightning.qubit", wires=4))
+    def f(shape):
+        i = 0
+        a = ones(shape, dtype=float)
[…]
