+++ runtime/tests/Test_LightningCoreQIS.cpp
@@ -40,10 +40,7 @@ TEST_CASE("Test __quantum__rt__print_tensor i64 1-dim", "[qir_lightning_core]")
     OpaqueMemRefT omr_i64_1d{1, (void *)(&mr_i64_1d), NumericType::i64};
     CHECK(omr_i64_1d.rank == 1);
 
-    DynamicMemRefT dmr_i64_1d = get_dynamic_memref(omr_i64_1d);
-    CHECK(dmr_i64_1d.sizes[0] == 100);
-
-    __quantum__rt__print_tensor(&omr_i64_1d);
+    __quantum__rt__print_tensor(&omr_i64_1d, false);
 }
 
 TEST_CASE("Test __quantum__rt__print_tensor dbl 2-dim", "[qir_lightning_core]")
@@ -52,21 +49,13 @@ TEST_CASE("Test __quantum__rt__print_tensor dbl 2-dim", "[qir_lightning_core]")
     buffer[0] = 1.0;
 
     MemRefT_double_2d mr_dbl_2d{buffer.data(), buffer.data(), 0, {100, 2}, {2, 1}};
+    CHECK(mr_dbl_2d.sizes[0] == 100);
+    CHECK(mr_dbl_2d.sizes[1] == 2);
+
     OpaqueMemRefT omr_dbl_2d{2, (void *)(&mr_dbl_2d), NumericType::f64};
     CHECK(omr_dbl_2d.rank == 2);
 
-    DynamicMemRefT dmr_dbl_2d = get_dynamic_memref(omr_dbl_2d);
-    CHECK(dmr_dbl_2d.sizes[0] == 100);
-    CHECK(dmr_dbl_2d.sizes[1] == 2);
-    CHECK(dmr_dbl_2d.strides[0] == 2);
-    CHECK(dmr_dbl_2d.strides[1] == 1);
-
-    auto data = dmr_dbl_2d.data_aligned;
-    CHECK(std::get<double *>(data)[0] == 1.0);
-    CHECK(std::get<double *>(data)[1] == 0.0);
-    CHECK(std::get<double *>(data)[50] == 0.0);
-
-    __quantum__rt__print_tensor(&omr_dbl_2d);
+    __quantum__rt__print_tensor(&omr_dbl_2d, true);
 }
 
 TEST_CASE("Test __quantum__rt__print_tensor cplx 2-dim", "[qir_lightning_core]")
@@ -79,20 +68,13 @@ TEST_CASE("Test __quantum__rt__print_tensor cplx 2-dim", "[qir_lightning_core]")
     };
 
     MemRefT_CplxT_double_2d mr_cplx_2d{matrix_data, matrix_data, 0, {2, 2}, {1, 0}};
+    CHECK(mr_cplx_2d.sizes[0] == 2);
+    CHECK(mr_cplx_2d.sizes[1] == 2);
+
     OpaqueMemRefT omr_cplx_2d{2, (void *)(&mr_cplx_2d), NumericType::c128};
     CHECK(omr_cplx_2d.rank == 2);
 
-    DynamicMemRefT dmr_cplx_2d = get_dynamic_memref(omr_cplx_2d);
-    CHECK(dmr_cplx_2d.sizes[0] == 2);
-    CHECK(dmr_cplx_2d.sizes[1] == 2);
-    CHECK(dmr_cplx_2d.strides[0] == 1);
-    CHECK(dmr_cplx_2d.strides[1] == 0);
-
-    auto data = dmr_cplx_2d.data_aligned;
-    CHECK(std::get<CplxT_double *>(data)[0].real == -0.67);
-    CHECK(std::get<CplxT_double *>(data)[0].imag == -0.63);
-
[…]
