### 2023-11-07 3f7b2f2
Move print function to debug module
