### 2023-11-07 d82f997
Don't enable warnings from the lightning repo
These interfere with our build flags as they appear to be appended last.
