+++ frontend/test/pytest/test_loops.py
@@ -31,7 +31,6 @@ class TestLoopToJaxpr:
         expected = """\
 { lambda ; a:f64[]. let
     b:i64[] c:f64[] = qwhile[
-      _id=0
       body_jaxpr={ lambda ; d:i64[] e:f64[]. let f:i64[] = add d 1 in (f, e) }
       body_nconsts=0
       cond_jaxpr={ lambda ; g:i64[] h:f64[]. let i:bool[] = lt g 10 in (i,) }
@@ -56,7 +55,6 @@ class TestLoopToJaxpr:
         expected = """\
 { lambda ; a:f64[] b:i64[]. let
     c:i64[] d:f64[] = qfor[
-      _id=0
       apply_reverse_transform=False
       body_jaxpr={ lambda ; e:i64[] f:i64[] g:f64[]. let
           h:i64[] = add f 1
