### 2023-11-07 1105543
Revert "Fix malformed JAXPR after Python fallback"
This reverts commit 916a01908612f2fc33a91a12a723123baa4aa29d.
