+++ frontend/catalyst/ag_primitives.py
@@ -43,8 +43,6 @@ from tensorflow.python.autograph.pyct.origin_info import LineLocation
 
 import catalyst
 from catalyst.ag_utils import AutoGraphError
-from catalyst.jax_primitives import qfor_p, qwhile_p
-from catalyst.utils.contexts import EvaluationContext
 from catalyst.utils.jax_extras import DynamicJaxprTracer, ShapedArray
 from catalyst.utils.patching import Patcher
 
@@ -159,8 +157,7 @@ def assert_iteration_results(inputs, outputs, symbol_names):
         inp_t, out_t = jax.api_util.shaped_abstractify(inp), jax.api_util.shaped_abstractify(out)
         if inp_t.dtype != out_t.dtype or inp_t.shape != out_t.shape:
             raise AutoGraphError(
-                f"The variable '{symbol_names[i]}' was initialized with the wrong type, or you may "
-                f"be trying to change its type from one iteration to the next. "
+                f"The variable '{symbol_names[i]}' was initialized with the wrong type. "
                 f"Expected: {out_t}, Got: {inp_t}"
             )
 
@@ -204,10 +201,6 @@ def _call_catalyst_for(
 
         return get_state()
 
-    # Needed in case we need to uniquely identify this for loop instance in the JAXPR,
-    # for example on recovery from tracing errors.
-    functional_for._set_id(id(body_fn))  # pylint: disable=protected-access
-
     final_iter_args = functional_for(*init_iter_args)
     assert_iteration_results(init_iter_args, final_iter_args, symbol_names)
     return final_iter_args
@@ -222,7 +215,6 @@ def _call_python_for(body_fn, get_state, non_array_iterable):
     return get_state()
 
 
-# pylint: disable=too-many-statements,too-many-branches
 def for_stmt(
     iteration_target: Any,
     _extra_test: Union[Callable[[], bool], None],
@@ -305,19 +297,9 @@ def for_stmt(
                 enum_start,
                 iteration_array,
             )
-
         except Exception as e:  # pylint: disable=broad-exception-caught
             if catalyst.autograph_strict_conversion:
                 raise e
-
-            # On fallback to Python we need to remove the jax primitive for this loop, if it was
-            # queued. The reason is that it may be malformed which can cause issues during lowering.
-            if EvaluationContext.is_tracing():  # pragma: no branch
-                jaxpr_frame = EvaluationContext.find_jaxpr_frame((start, stop, step, *init_state))
-                last_prim_is_for = jaxpr_frame.eqns and jaxpr_frame.eqns[-1].primitive is qfor_p
-                last_prim_id = jaxpr_frame.eqns[-1].params["_id"] if last_prim_is_for else None
-                if last_prim_id == id(body_fn):
-                    jaxpr_frame.eqns.pop()
 
             # pylint: disable=import-outside-toplevel
             import inspect
@@ -373,10 +355,6 @@ def _call_catalyst_while(loop_test, loop_body, get_state, set_state, symbol_name
         loop_body()
         return get_state()
 
-    # Needed in case we need to uniquely identify this for loop instance in the JAXPR,
-    # for example on recovery from tracing errors.
-    functional_while._set_id(id(loop_body))  # pylint: disable=protected-access
-
     final_iter_args = functional_while(init_iter_args)
     assert_iteration_results(init_iter_args, final_iter_args, symbol_names)
     return final_iter_args
@@ -404,16 +382,6 @@ def while_stmt(loop_test, loop_body, get_state, set_state, symbol_names, _opts):
     except Exception as e:  # pylint: disable=broad-exception-caught
         if catalyst.autograph_strict_conversion:
             raise e
-
[…]
