### 2023-11-07 847fb70
Update changelog for 0.3.2
