+++ doc/changelog.md
@@ -1,144 +1,132 @@
-# Release 0.3.2-dev
+# Release 0.3.2
 
 <h3>New features</h3>
 
-<h3>Improvements</h3>
-
-* Improve the compiler driver diagnostic output even more. Now it dumps the failing IR as well as
-  prints names of failing passes.
-  [(#349)](https://github.com/PennyLaneAI/catalyst/pull/349)
-
-* Return values of conditional functions no longer need to be of exactly the same type. Catalyst
-  would apply type promotion to branch return values if their types don't match.
-  [(#333)](https://github.com/PennyLaneAI/catalyst/pull/333)
-
-  ```python
-  @qjit
-  def func(i: int, f: float):
-      @cond(i < 3)
-      def cond_fn():
-          return i
-      @cond_fn.otherwise
-      def otherwise():
-          return f
-      return cond_fn()
-  ```
-
-  ```pycon
-  >>> func(1, 4.0)
-  array(1.0)
-  ```
-
-* Improve the `CopyGlobalMemRefPass` of our MLIR processing pipeline by adding the support of
-  dynamically shaped arrays.
-  [(#348)](https://github.com/PennyLaneAI/catalyst/pull/348)
-
-* Improve the compiler driver diagnostic output. The driver now provides more context for error
-  messages and includes a verbose trace if verbose mode is enabled.
-  [(#303)](https://github.com/PennyLaneAI/catalyst/pull/303)
-
-* Return values of conditional functions no longer need to be of exactly the same type. Catalyst
-  would apply type promotion to branch return values if their types don't match.
-  [(#333)](https://github.com/PennyLaneAI/catalyst/pull/333)
-
-  ```python
-  @qjit
-  def func(i: int, f: float):
-      @cond(i < 3)
-      def cond_fn():
-          return i
-      @cond_fn.otherwise
-      def otherwise():
-          return f
-      return cond_fn()
-  ```
-
-  ```pycon
-  >>> func(1, 4.0)
-  array(1.0)
-  ```
-
-* The AutoGraph feature, still experimental, now supports native Python `while` loops as well.
+* The experimental AutoGraph feature now supports Python `while` loops, allowing native Python loops
+  to be captured and compiled with Catalyst.
   [(#318)](https://github.com/PennyLaneAI/catalyst/pull/318)
 
   ```python
+  dev = qml.device("lightning.qubit", wires=4)
+
   @qjit(autograph=True)
-  @qml.qnode(qml.device("lightning.qubit", wires=4))
-  def circuit(n:int):
+  @qml.qnode(dev)
+  def circuit(n: int, x: float):
       i = 0
+
       while i < n:
-          qml.RX(jnp.pi/2, wires=i)
+          qml.RX(x, wires=i)
           i += 1
+
       return qml.expval(qml.PauliZ(0))
   ```
 
-* The AutoGraph feature now also supports native Python `and`, `or` and `not` operators in Boolean
-  expressions.
+  ```pycon
+  >>> circuit(4, 0.32)
+  array(0.94923542)
+  ```
+
+  This feature extends the existing AutoGraph support for Python `for` loops and if` statements
+  introduced in v0.3. Note that TensorFlow must be installed for AutoGraph support.
+
+  For more details, please see the
+  [AutoGraph guide](https://docs.pennylane.ai/projects/catalyst/en/stable/dev/autograph.html).
+
+* In addition to loops and conditional branches, AutoGraph now supports native Python `and`, `or`
+  and `not` operators in Boolean expressions.
   [(#325)](https://github.com/PennyLaneAI/catalyst/pull/325)
 
   ```python
+  dev = qml.device("lightning.qubit", wires=1)
+
   @qjit(autograph=True)
-  @qml.qnode(qml.device("lightning.qubit", wires=1))
-  def circuit(param:float):
[…]
