+++ runtime/Makefile
@@ -8,7 +8,7 @@ MK_DIR := $(dir $(MK_ABSPATH))
 RT_BUILD_DIR?=$(MK_DIR)/build
 CMAKE_ARGS?=""
 CODE_COVERAGE?=OFF
-BUILD_TYPE?=RelWithDebInfo
+BUILD_TYPE?=Release
 ENABLE_WARNINGS?=ON
 ENABLE_LIGHTNING_KOKKOS?=OFF
 ENABLE_OPENQASM?=OFF
@@ -27,7 +27,7 @@ ENABLE_OPENMP?=$(DEFAULT_ENABLE_OPENMP)
 coverage: CODE_COVERAGE=ON
 coverage: BUILD_TYPE=Debug
 test: CODE_COVERAGE=OFF
-test: BUILD_TYPE?=RelWithDebInfo
+test: BUILD_TYPE?=Release
 
 .PHONY: help
 help:
