+++ doc/dev/sharp_bits.rst
@@ -606,23 +606,10 @@ Conditional debugging
 There are various constraints and restrictions that should be kept in mind
 when working with classical control in Catalyst.
 
-- The return values of all branches of :func:`~.cond` must be the same type.
-  Returning different types, or ommitting a return value in one branch (e.g.,
-  returning ``None``) but not in others will result in an error.
+- The return values of all branches of :func:`~.cond` do not have to be the same type;
+  Catalyst will perform automatic type promotion (for example, converting integers)
+  to floats) where possible.
 
-  >>> @qjit
-  ... def f(x: float):
-  ...     @cond(x > 1.5)
-  ...     def cond_fn():
-  ...         return x ** 2  # float
-  ...     @cond_fn.otherwise
-  ...     def else_branch():
-  ...         return 6  # int
-  ...     return cond_fn()
-  TypeError: Conditional requires consistent return types across all branches, got:
-  - Branch at index 0: [ShapedArray(float64[], weak_type=True)]
-  - Branch at index 1: [ShapedArray(int64[], weak_type=True)]
-  Please specify an else branch if none was specified.
   >>> @qjit
   ... def f(x: float):
   ...     @cond(x > 1.5)
@@ -635,9 +622,10 @@ when working with classical control in Catalyst.
   >>> f(1.5)
   array(6.)
 
-- Similarly, the else (``my_cond_fn.otherwise``) may be omitted **as long as
-  other branches do not return any values**. If other branches do return values,
-  the else branch must be specified.
+- There may be some cases where automatic type promotion cannot be applied; for example,
+  ommitting a return value in one branch (e.g., which by default in Python is equivalent
+  to returning ``None``) but not in others. This will result in an error ---
+  if other branches do return values, the else branch must be specified.
 
   >>> @qjit
   ... def f(x: float):
