### 2023-11-07 c31d6af
Reimplenent program state reset without IDs
Instead, simply use program length as a marker, since instructions are
inserted in an ordered fashion.

Also implement state reset for quantum tracing.
