+++ frontend/test/pytest/test_autograph.py
@@ -16,7 +16,6 @@
 
 import sys
 import traceback
-import warnings
 from collections import defaultdict
 
 import jax
@@ -1264,7 +1263,6 @@ class TestWhileLoops:
 
     def test_whileloop_temporary_variable(self, monkeypatch):
         """Test that temporary (local) variables can be initialized inside a while loop."""
-
         monkeypatch.setattr("catalyst.autograph_strict_conversion", True)
 
         @qjit(autograph=True)
@@ -1280,7 +1278,6 @@ class TestWhileLoops:
 
     def test_whileloop_forloop_interop(self, monkeypatch):
         """Test for-loop co-existing with while loop."""
-
         monkeypatch.setattr("catalyst.autograph_strict_conversion", True)
 
         @qjit(autograph=True)
@@ -1296,7 +1293,6 @@ class TestWhileLoops:
 
     def test_whileloop_cond_interop(self, monkeypatch):
         """Test for-loop co-existing with while loop."""
-
         monkeypatch.setattr("catalyst.autograph_strict_conversion", True)
 
         @qjit(autograph=True)
@@ -1311,27 +1307,23 @@ class TestWhileLoops:
 
         assert f1() == sum([1, 1, 2, 2])
 
+    @pytest.mark.xfail(reason="this won't run warning-free until we fix the resource warning issue")
+    @pytest.mark.filterwarnings("error")
     def test_whileloop_no_warning(self, monkeypatch):
         """Test the absence of warnings if fallbacks are ignored."""
-        # pylint: disable=anomalous-backslash-in-string
-
-        monkeypatch.setattr("catalyst.autograph_strict_conversion", False)
         monkeypatch.setattr("catalyst.autograph_ignore_fallbacks", True)
 
-        def f1():
+        @qjit(autograph=True)
+        def f():
             acc = 0
             while Failing(acc).val < 5:
                 acc = acc + 1
             return acc
 
-        with warnings.catch_warnings():
-            warnings.simplefilter("error")
-            assert 5 == qjit(autograph=True)(f1)()
+        assert f() == 5
 
     def test_whileloop_exception(self, monkeypatch):
         """Test for-loop error if strict-conversion is enabled."""
-        # pylint: disable=anomalous-backslash-in-string
-
         monkeypatch.setattr("catalyst.autograph_strict_conversion", True)
 
         def f1():
@@ -1386,6 +1378,116 @@ class TestWhileLoops:
 
         with pytest.raises(AutoGraphError, match="'x' was initialized with the wrong type"):
             qjit(autograph=True)(f)
+
+
+@pytest.mark.tf
+@pytest.mark.parametrize(
+    "execution_context", (lambda fn: fn, qml.qnode(qml.device("lightning.qubit", wires=1)))
+)
+class TestFallback:
+    """Test that Python fallbacks still produce correct results."""
+
+    def test_postbinding_errors_for(self, execution_context):
+        """Test that errors are handled correctly if they trigger after the JAX primitive binding
+        step (e.g. during result verification), and that no errors occur after the AG tracing step
+        (e.g. during lowering) because of malformed primitives in the JAXPR.
+        This test ensures the primitive identification and removal works correctly on fallback. In
+        this case, the loop primitive should be removed since the exception happens after binding.
+        """
+
+        @execution_context
+        def f():
+            arr = jnp.array([1, 2])
+            for _ in range(2):
+                # fails result verification, will trigger fallback
+                # would raise an error during lowering if left in the JAXPR
+                arr = jnp.kron(arr, arr)
+            return arr
+
+        with pytest.warns(
[…]
