### 2023-11-07 64880cf
Apply suggestions from code review
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
