### 2023-11-07 a848332
[CI] Binaries are available for macOS aarch64 and x86_64 (#347)
- [X] Add wheels recipes for macOS `aarch64` and `x86_64`
- [X] Update Make commands to support building from src on both Linux
and macOS
- [X] Fix a few minor issues in the build system and typos

[sc-46388]

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
