### 2023-11-07 12349f2
starting logical section
