### 2023-11-07 98b0ac3
Apply suggestions from code review
Co-authored-by: Ali Asadi <ali@xanadu.ai>
