### 2023-11-07 5f73839
Small fix to print output in changelog
