### 2023-11-07 1d1af72
Add tests for rt_print_tensor
