### 2023-11-07 be489a5
review: wording & typo
Co-authored-by: Josh Izaac <josh146@gmail.com>
