### 2023-11-07 38fd1cb
Include the `debug` module in the docs
