### 2023-11-07 95bb0df
Add Catalyst print function for debugging (#279)
Add support for "printf"-style debugging to catalyst. A new operator
`catalyst.debug.print` allows users to print arbitrary values from their
program at runtime.

Example:
```python
@qjit
def func(x: float):
    debug.print(x, memref=True)
    debug.print("exit")
```

```pycon
>>> func(jnp.array(0.43))
MemRef: base@ = 0x5629ff2b6680 rank = 0 offset = 0 sizes = [] strides = [] data =
[0.43]
exit
```

[sc-41312]

---------

Co-authored-by: Ali Asadi <ali@xanadu.ai>
Co-authored-by: Josh Izaac <josh146@gmail.com>
