+++ frontend/catalyst/pennylane_extensions.py
@@ -41,16 +41,16 @@ from catalyst.jax_primitives import (
     AbstractQreg,
     GradParams,
     adjoint_p,
+    cond_p,
     expval_p,
+    for_p,
     func_p,
     grad_p,
     jvp_p,
     probs_p,
-    qcond_p,
-    qfor_p,
     qmeasure_p,
-    qwhile_p,
     vjp_p,
+    while_p,
 )
 from catalyst.jax_tracer import (
     Function,
@@ -869,7 +869,7 @@ def _check_cond_same_shapes(trees: List[PyTreeDef], avals: List[List[Any]]) -> N
 class ForLoop(HybridOp):
     """PennyLane ForLoop Operation."""
 
-    binder = qfor_p.bind
+    binder = for_p.bind
 
     def trace_quantum(self, ctx, device, trace, qrp) -> QRegPromise:
         op = self
@@ -919,7 +919,7 @@ class MidCircuitMeasure(HybridOp):
 class Cond(HybridOp):
     """PennyLane's conditional operation."""
 
-    binder = qcond_p.bind
+    binder = cond_p.bind
 
     def trace_quantum(self, ctx, device, trace, qrp) -> QRegPromise:
         jaxprs, consts = [], []
@@ -954,7 +954,7 @@ class Cond(HybridOp):
 class WhileLoop(HybridOp):
     """PennyLane's while loop operation."""
 
-    binder = qwhile_p.bind
+    binder = while_p.bind
 
     def trace_quantum(self, ctx, device, trace, qrp) -> QRegPromise:
         cond_trace = self.regions[0].trace
@@ -1182,7 +1182,7 @@ class CondCallable:
         )
         _check_cond_same_shapes(out_trees, [j.out_avals for j in branch_jaxprs])
         branch_jaxprs = unify_result_types(branch_jaxprs)
-        out_classical_tracers = qcond_p.bind(*(self.preds + consts), branch_jaxprs=branch_jaxprs)
+        out_classical_tracers = cond_p.bind(*(self.preds + consts), branch_jaxprs=branch_jaxprs)
         return tree_unflatten(out_trees[0], out_classical_tracers)
 
     def _call_during_interpretation(self):
@@ -1500,7 +1500,7 @@ def for_loop(lower_bound, upper_bound, step):
                 )
 
                 apply_reverse_transform = isinstance(step, int) and step < 0
-                out_classical_tracers = qfor_p.bind(
+                out_classical_tracers = for_p.bind(
                     lower_bound,
                     upper_bound,
                     step,
@@ -1647,7 +1647,7 @@ def while_loop(cond_fn):
                     body_fn, in_tree, init_avals, "while_loop"
                 )
                 _check_single_bool_value(cond_tree, cond_jaxpr.out_avals)
-                out_classical_tracers = qwhile_p.bind(
+                out_classical_tracers = while_p.bind(
                     *(cond_consts + body_consts + init_vals),
                     cond_jaxpr=cond_jaxpr,
                     body_jaxpr=body_jaxpr,

[…]
