### 2023-11-07 2ae74b9
Update rt_print_string to skip segfault on macOS
