### 2023-11-07 7ebc80a
Add the newly added catalyst module to the Make wheel command
