### 2023-11-07 ab60358
Update __quantum__rt__print_string test
