### 2023-11-08 7dcfbb4
Add the missing return to __quantum__rt__print_string
