### 2023-11-08 48d917b
Update macos dev version to suppress ld warnings for the compiler-built obj files
