### 2023-11-08 7c92934
Apply suggestions from code review
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
