### 2023-11-08 2737a84
Update wheels: Addressing changes in PR #279 (#356)
With Changes in PR #279, 
- [X] Disable the warning-as-error build option for 3rd-party backend
devices on wheels recipes.
- [X] Add the new `catalyst` module to the wheels
- [x] Fix segfaults with debug_print tests on macOS

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
