### 2023-11-08 cdff829
[Frontend] Fix Python fallback errors with AutoGraph in certain instances (#352)
Improve robustness of AutoGraph conversion. 

List of changes:
- fix bug when extracting the variable name in question for certain
error messages
- add input/output checking for while loops & add tests
- setup fix for leftover JAX primitive when fallback is triggered, by
adding access to the current JAXPR frame and quantum queuing context
- remove for/while loop primitives from the JAXPR when a fallback to
Python has occurred
 
Enables to following to succeed for example:
```py
@qjit(autograph=True)
def f():
    l
[…]
