+++ frontend/catalyst/utils/contexts.py
@@ -29,6 +29,8 @@ from jax._src.interpreters.partial_eval import (
     extend_jaxpr_stack,
 )
 from jax._src.source_info_util import reset_name_stack
+from jax.core import find_top_trace
+from pennylane.queuing import QueuingManager
 
 from catalyst.utils.exceptions import CompileError
 from catalyst.utils.jax_extras import new_dynamic_main2
@@ -171,6 +173,13 @@ class EvaluationContext:
         ]
 
     @classmethod
+    def is_quantum_tracing(cls):
+        """Returns true or false depending on whether the execution is currently being
+        traced.
+        """
+        return cls.get_mode() == EvaluationMode.QUANTUM_COMPILATION
+
+    @classmethod
     def check_modes(cls, modes, msg):
         """Asserts if the execution mode is not among the expected ``modes``.
 
@@ -213,3 +222,25 @@ class EvaluationContext:
         """
         if cls.is_tracing():
             raise CompileError(msg)
+
+    @classmethod
+    def find_jaxpr_frame(cls, *args):
+        """Obtain the current JAXPR frame, in which primitives are being inserted.
+
+        Raises: CompileError
+        """
+        cls.check_is_tracing("No JAXPR frames exist outside a tracing context.")
+        return find_top_trace(args).frame
+
+    @classmethod
+    def find_quantum_queue(cls):
+        """Obtain the current quantum queuing context, in which operations are being inserted.
+
+        Raises: CompileError
+        """
+        cls.check_is_quantum_tracing("No quantum queueing context found.")
+
+        queuing_context = QueuingManager.active_context()
+        assert queuing_context is not None
+
+        return queuing_context
