+++ frontend/catalyst/jax_primitives.py
@@ -62,6 +62,7 @@ from pennylane import QNode as pennylane_QNode
 
 from catalyst.utils.calculate_grad_shape import Signature, calculate_grad_shape
 from catalyst.utils.extra_bindings import FromElementsOp, TensorExtractOp
+from catalyst.utils.jax_extras import finfo_or_none, iinfo_or_none
 
 # pylint: disable=unused-argument,too-many-lines
 
@@ -1444,52 +1445,35 @@ def _adjoint_lowering(
 #
 
 
-def _tryinfo(f, dtype) -> Optional[Any]:
-    """Wrapper around jnp.iinfo/jnp.finfo functions"""
-    try:
-        return f(dtype)
-    except ValueError:
-        return None
-
-
-def _iinfo(dtype) -> Optional[Any]:
-    """Return integer-type information or None"""
-    return _tryinfo(jnp.iinfo, dtype)
-
-
-def _finfo(dtype) -> Optional[Any]:
-    """Return float-type information or None"""
-    return _tryinfo(jnp.finfo, dtype)
-
-
 @tensor_init_p.def_impl
-def _tensor_init_def_impl(ctx, shape, *, initializer, dtype):  # pragma: no cover
+def _tensor_init_def_impl(
+    ctx, dynamic_shape, *, static_shape, initializer, dtype
+):  # pragma: no cover
     raise NotImplementedError()
 
 
 @tensor_init_p.def_abstract_eval
-def _tensor_init_abstract(shape, *, initializer, dtype):
-    if len(shape.shape) != 1 or _iinfo(shape.dtype) is None:
-        raise ValueError("The shape is expected to have rank one and contain integers")
-    return core.ShapedArray([ir.ShapedType.get_dynamic_size()] * shape.shape[0], dtype=dtype)
+def _tensor_init_abstract(dynamic_shape, *, static_shape, initializer, dtype):
+    static_shape = [dim if dim >= 0 else ir.ShapedType.get_dynamic_size() for dim in static_shape]
+    return core.ShapedArray(static_shape, dtype=dtype)
 
 
 def _tensor_init_lowering(
-    jax_ctx: mlir.LoweringRuleContext, shape: ir.Value, *, initializer, dtype
+    jax_ctx: mlir.LoweringRuleContext, dynamic_shape: ir.Value, *, static_shape, initializer, dtype
 ) -> ir.Value:
     empty = mlir.ir_constants(True) if initializer is None else None  # do not initialize
     output_type = [mlir.aval_to_ir_types(a)[0] for a in jax_ctx.avals_out][0]
 
     init_value = initializer if not empty else 0
     initializer_attr = None
-    if _finfo(dtype) is not None:
+    if finfo_or_none(dtype) is not None:
         initializer_attr = ir.FloatAttr.get(output_type.element_type, float(init_value))
-    elif _iinfo(dtype) is not None:
+    elif iinfo_or_none(dtype) is not None:
         initializer_attr = ir.IntegerAttr.get(output_type.element_type, int(init_value))
     else:
         raise ValueError(f"Unsupported initializer {initializer} of type {dtype}")
 
-    op = TensorInitOp(output_type, shape=shape, empty=empty, initializer=initializer_attr)
+    op = TensorInitOp(output_type, shape=dynamic_shape, empty=empty, initializer=initializer_attr)
     return op.results
 
 

[…]
