### 2023-11-08 a793c11
Fix lowering pass; Address codefactor issues
