### 2023-11-08 8704922
Update changelog for 0.3.2 (#355)
Co-authored-by: Ali Asadi <ali@xanadu.ai>
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
