+++ doc/changelog.md
@@ -1,6 +1,66 @@
-# Release 0.3.2-dev
+# Release 0.3.2
 
 <h3>New features</h3>
+
+* The experimental AutoGraph feature now supports Python `while` loops, allowing native Python loops
+  to be captured and compiled with Catalyst.
+  [(#318)](https://github.com/PennyLaneAI/catalyst/pull/318)
+
+  ```python
+  dev = qml.device("lightning.qubit", wires=4)
+
+  @qjit(autograph=True)
+  @qml.qnode(dev)
+  def circuit(n: int, x: float):
+      i = 0
+
+      while i < n:
+          qml.RX(x, wires=i)
+          i += 1
+
+      return qml.expval(qml.PauliZ(0))
+  ```
+
+  ```pycon
+  >>> circuit(4, 0.32)
+  array(0.94923542)
+  ```
+
+  This feature extends the existing AutoGraph support for Python `for` loops and `if` statements
+  introduced in v0.3. Note that TensorFlow must be installed for AutoGraph support.
+
+  For more details, please see the
+  [AutoGraph guide](https://docs.pennylane.ai/projects/catalyst/en/stable/dev/autograph.html).
+
+* In addition to loops and conditional branches, AutoGraph now supports native Python `and`, `or`
+  and `not` operators in Boolean expressions.
+  [(#325)](https://github.com/PennyLaneAI/catalyst/pull/325)
+
+  ```python
+  dev = qml.device("lightning.qubit", wires=1)
+
+  @qjit(autograph=True)
+  @qml.qnode(dev)
+  def circuit(x: float):
+
+      if x >= 0 and x < jnp.pi:
+          qml.RX(x, wires=0)
+
+      return qml.probs()
+  ```
+
+  ```pycon
+  >>> circuit(0.43)
+  array([0.95448287, 0.04551713])
+  >>> circuit(4.54)
+  array([1., 0.])
+  ```
+
+  Note that logical Boolean operators will only be captured by AutoGraph if all
+  operands are dynamic variables (that is, a value known only at runtime, such
+  as a measurement result or function argument). For other use
+  cases, it is recommended to use the `jax.numpy.logical_*` set of functions where
+  appropriate.
 
 * Debug compiled programs and print dynamic values at runtime with ``debug.print``
   [(#279)](https://github.com/PennyLaneAI/catalyst/pull/279)
@@ -28,121 +88,100 @@
   ```
 
 * Catalyst now officially supports macOS X86_64 devices, with macOS binary wheels
-  available for both AARCH64 and X86_64 on PyPI.
+  available for both AARCH64 and X86_64.
   [(#347)](https://github.com/PennyLaneAI/catalyst/pull/347)
   [(#313)](https://github.com/PennyLaneAI/catalyst/pull/313)
 
-<h3>Improvements</h3>
-
-* Improve the compiler driver diagnostic output even more. Now it dumps the failing IR as well as
-  prints names of failing passes.
-  [(#349)](https://github.com/PennyLaneAI/catalyst/pull/349)
-
-* Return values of conditional functions no longer need to be of exactly the same type. Catalyst
-  would apply type promotion to branch return values if their types don't match.
-  [(#333)](https://github.com/PennyLaneAI/catalyst/pull/333)
-
-  ```python
-  @qjit
-  def func(i: int, f: float):
-      @cond(i < 3)
-      def cond_fn():
-          return i
-      @cond_fn.otherwise
-      def otherwise():
-          return f
-      return cond_fn()
-  ```
-
-  ```pycon
-  >>> func(1, 4.0)
-  array(1.0)
-  ```
-
[…]
