### 2023-11-08 #6012 @danielskatz
👋 @jarvist - a ping to see if you would be able to edit this, as [previously invited](https://github.com/openjournals/joss-reviews/issues/6012#issuecomment-1790759685)
