+++ mlir/lib/CAPI/Dialects.cpp
@@ -24,4 +24,5 @@
 
 MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Quantum, quantum, catalyst::quantum::QuantumDialect)
 MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Gradient, gradient, catalyst::gradient::GradientDialect)
+MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Catalyst, catalyst, catalyst::CatalystDialect)
 MLIR_DEFINE_CAPI_DIALECT_REGISTRATION(Mitigation, mitigation, catalyst::mitigation::MitigationDialect)
