### 2023-11-09 e1a2155
Remove memory semantics
