### 2023-11-09 7c00ec4
Merge `v0.3.2-rc` into `main` (#359)
Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Ali Asadi <ali@xanadu.ai>
