### 2023-11-09 0ef5d9a
Raise error if transformed and returns something other than measurement processes.
