+++ frontend/test/pytest/test_transform.py
@@ -409,8 +409,10 @@ class TestTransformValidity:
         )
         H4 += qml.PauliZ(0) @ qml.PauliX(1) @ qml.PauliY(2)
 
-        msg = "A transformed quantum function must return either a single measurement, "\
-                "or a nonempty sequence of measurements."
+        msg = (
+            "A transformed quantum function must return either a single measurement, "
+            "or a nonempty sequence of measurements."
+        )
         with pytest.raises(CompileError, match=msg):
 
             @qjit
