### 2023-11-09 79a9861
Split backend devices to separate shared libs
