+++ runtime/CMakeLists.txt
@@ -20,7 +20,6 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 set(runtime_includes "${PROJECT_SOURCE_DIR}/include")
 set(capi_utils_includes "${PROJECT_SOURCE_DIR}/lib/capi")
-set(extensions_includes "${PROJECT_SOURCE_DIR}/extensions")
 set(backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/common")
 
 set(QIRSTDLIB_VERSION "0.4.0")
@@ -70,24 +69,30 @@ endif()
 
 message(STATUS "QIR_STDLIB: ${QIR_STDLIB_PATH}")
 
-if(ENABLE_LIGHTNING)
-    message(STATUS "ENABLE_LIGHTNING is ${ENABLE_LIGHTNING}.")
-    list(APPEND backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/lightning")
-endif()
+message(STATUS "ENABLE_LIGHTNING is ${ENABLE_LIGHTNING}.")
+message(STATUS "ENABLE_LIGHTNING_KOKKOS is ${ENABLE_LIGHTNING_KOKKOS}.")
+message(STATUS "ENABLE_OPENQASM is ${ENABLE_OPENQASM}.")
 
-if(ENABLE_LIGHTNING_KOKKOS)
-    message(STATUS "ENABLE_LIGHTNING_KOKKOS is ${ENABLE_LIGHTNING_KOKKOS}.")
-    list(APPEND backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/lightning-kokkos")
+set(devices_list)
+
+if(ENABLE_LIGHTNING OR ENABLE_LIGHTNING_KOKKOS)
+    list(APPEND rtd_lightning)
+    if(ENABLE_LIGHTNING)
+        list(APPEND backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/lightning/lightning_dynamic")
+    endif()
+    if(ENABLE_LIGHTNING_KOKKOS)
+        list(APPEND backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/lightning/lightning_kokkos")
+    endif()
 endif()
 
 if(ENABLE_OPENQASM)
-    message(STATUS "ENABLE_OPENQASM is ${ENABLE_OPENQASM}.")
     list(APPEND backend_includes "${PROJECT_SOURCE_DIR}/lib/backend/openqasm")
+    list(APPEND devices_list rtd_openqasm)
 endif()
 
 add_library(catalyst_qir_runtime INTERFACE)
 
-target_link_libraries(catalyst_qir_runtime INTERFACE rt_backend rt_capi)
+target_link_libraries(catalyst_qir_runtime INTERFACE ${devices_list} rt_capi)
 
 target_include_directories(catalyst_qir_runtime INTERFACE runtime_includes ${QIR_STDLIB_PATH})
 
@@ -110,7 +115,5 @@ if(ENABLE_SANITIZER)
     set(CMAKE_LINKER_FLAGS_DEBUG "${CMAKE_LINKER_FLAGS_DEBUG} -fno-omit-frame-pointer -fsanitize=address")
 ENDIF()
 
-add_subdirectory(extensions)
 add_subdirectory(lib)
-
 add_subdirectory(tests)
