### 2023-11-09 0e347fc
Fix lit tests
