+++ frontend/test/lit/test_if_else.py
@@ -29,7 +29,7 @@ def circuit(n: int):
     # CHECK-DAG:   [[qreg_0:%[a-zA-Z0-9_]+]] = quantum.alloc
     # CHECK:       [[b:%[a-zA-Z0-9_]+]] = tensor.extract [[b_t]]
     @cond(n <= 5)
-    # CHECK:       scf.if [[b]]
+    # CHECK:       [[qreg_2:%.+]]:2 = scf.if [[b]]
     def cond_fn():
         # CHECK-DAG:   [[q0:%[a-zA-Z0-9_]+]] = quantum.extract
         # CHECK-DAG:   [[q1:%[a-zA-Z0-9_]+]] = quantum.custom "PauliX"() [[q0]]
@@ -47,7 +47,8 @@ def circuit(n: int):
         return n**3
 
     out = cond_fn()
-    # CHECK:       quantum.dealloc [[qreg_0]]
+    # CHECK:       [[qreg_3:%.+]] = quantum.insert [[qreg_2]]#1[ 0], %out_qubit : !quantum.reg, !quantum.bit
+    # CHECK:       quantum.dealloc [[qreg_3]]
     # CHECK:       return
     return out, measure(wires=0)
 
