### 2023-11-09 cd1e9f7
Fix insert canonicalize issue
