### 2023-11-09 28828ad
Re-design the custom device interface
