+++ frontend/test/pytest/test_compiler.py
@@ -272,8 +272,8 @@ module @workflow {
     return %0 : tensor<f64>
   }
   func.func private @workflow(%arg0: tensor<f64>) -> tensor<f64> attributes {diff_method = "finite-diff", llvm.linkage = #llvm.linkage<internal>, qnode} {
-    quantum.device ["kwargs", "{'shots': 0}"]
-    quantum.device ["backend", "lightning.qubit"]
+    quantum.device ["rtd_kwargs", "{'shots': 0}"]
+    quantum.device ["rtd_lib", "lightning.qubit"]
     %0 = stablehlo.constant dense<4> : tensor<i64>
     %1 = quantum.alloc( 4) : !quantum.reg
     %2 = stablehlo.constant dense<0> : tensor<i64>
