### 2023-11-10 #6012 @jarvist
@josh146 - do you have any suggestions for reviewers? This style of software is really quite specialist. I was wondering if some people from the slightly more general area of quantum computing / quantum information might have a view.
