### 2023-11-10 20088c8
[Frontend] Support for transforms (#280)
**Context:** Allow PL transforms to be traced in Catalyst

**Description of the Change:** These changes are implemented on top of
the `tape-transform` PR in Catalyst and the development version of PL.
These changes apply transforms that are stored on the device into the
tape, potentially generating many tapes. Then each tape is traced
individually. At last, the classical co-transform is also traced.

**Benefits:** Catalyst support for PL transforms.

[sc-41344]
