+++ frontend/catalyst/pennylane_extensions.py
@@ -123,7 +123,9 @@ class QFunc:
         update_wrapper(self, fn)
 
     def __call__(self, *args, **kwargs):
+        qnode = None
         if isinstance(self, qml.QNode):
+            qnode = self
             if isinstance(self.device, qml.Device):
                 name = self.device.short_name
             else:
@@ -165,7 +167,7 @@ class QFunc:
             device = self.device
 
         with EvaluationContext(EvaluationMode.QUANTUM_COMPILATION):
-            jaxpr, shape = trace_quantum_function(self.func, device, args, kwargs)
+            jaxpr, shape = trace_quantum_function(self.func, device, args, kwargs, qnode)
 
         retval_tree = tree_structure(shape)
 
