### 2023-11-10 80ea302
Change tensorflow requirement for macOS support
