### 2023-11-10 1b8d447
Add section on how to build docs to the docs
