### 2023-11-10 c6bd229
Fix default build directory for mhlo
The directory should be include in the .gitignore of the submodule,
so that it is properly ignore with various tools (git status, vscode).

`build` cannot be chosen for mhlo, to avoid conflict with file named
`BUILD` on non-case-sensitive file systems.
