### 2023-11-10 78d46c5
Do not overwrite static decomposition with instance method
