### 2023-11-10 1868bb2
Fix self.decomposition signature
