### 2023-11-10 2719303
Cleanup intermediate files in tests
