### 2023-11-10 c2343fe
Do not check unused instance attributes
