### 2023-11-13 6b50dc4
[Frontend] Fix custom `Operation` overwriting static decomposition with instance method (#362)
