+++ frontend/catalyst/utils/jax_extras.py
@@ -15,23 +15,29 @@
 """
 from __future__ import annotations
 
-from contextlib import contextmanager
+from contextlib import ExitStack, contextmanager
 from typing import Any, Callable, Dict, Generator, List, Optional, Sequence, Set, Type
 
+import jax._src.core as jax_core
 from jax import ShapeDtypeStruct
 from jax._src import state, util
+from jax._src.api import _flat_axes_specs
 from jax._src.core import _update_thread_local_jit_state, eval_jaxpr
 from jax._src.dispatch import jaxpr_replicas
 from jax._src.effects import ordered_effects as jax_ordered_effects
 from jax._src.interpreters.mlir import _module_name_regex
-from jax._src.interpreters.partial_eval import _input_type_to_tracers
+from jax._src.interpreters.partial_eval import (
+    _input_type_to_tracers,
+    infer_lambda_input_type,
+    trace_to_jaxpr_dynamic2,
+)
 from jax._src.lax.control_flow import _initial_style_jaxpr, _initial_style_open_jaxpr
 from jax._src.lax.lax import _abstractify, xla
 from jax._src.linear_util import annotate
 from jax._src.sharding_impls import ReplicaAxisContext
 from jax._src.source_info_util import current as jax_current
 from jax._src.source_info_util import new_name_stack
-from jax._src.util import partition_list, safe_map, unzip2, unzip3, wrap_name
+from jax._src.util import partition_list, safe_map, unzip2, unzip3, wrap_name, wraps
 from jax.api_util import flatten_fun, shaped_abstractify
 from jax.core import ClosedJaxpr, Jaxpr, JaxprEqn, MainTrace
 from jax.core import Primitive as JaxprPrimitive
@@ -91,6 +97,8 @@ __all__ = (
     "wrap_init",
     "iinfo_or_none",
     "finfo_or_none",
+    "make_jaxpr_pytree",
+    "jaxpr_filter_outputs",
 )
 
 map, unsafe_map = safe_map, map  # pylint: disable=redefined-builtin
@@ -277,7 +285,7 @@ def deduce_avals(f: Callable, args, kwargs):
     return wffa, in_avals, out_tree_promise
 
 
-def jaxpr_to_mlir(func_name, jaxpr, shape):
+def jaxpr_to_mlir(func_name, jaxpr):
     """Lower a Python function into an MLIR module.
 
     Args:
@@ -289,24 +297,32 @@ def jaxpr_to_mlir(func_name, jaxpr, shape):
         module: the MLIR module corresponding to ``func``
         context: the MLIR context corresponding
         jaxpr: the jaxpr corresponding to ``func``
-        shape: the shape of the return values in ``PyTreeDef``
     """
 
-    nrep = jaxpr_replicas(jaxpr)
-    effects = jax_ordered_effects.filter_in(jaxpr.effects)
-    axis_context = ReplicaAxisContext(xla.AxisEnv(nrep, (), ()))
-    name_stack = new_name_stack(wrap_name("ok", "jit"))
-    module, context = custom_lower_jaxpr_to_module(
-        func_name="jit_" + func_name,
-        module_name=func_name,
-        jaxpr=jaxpr,
-        effects=effects,
-        platform="cpu",
-        axis_context=axis_context,
-        name_stack=name_stack,
-    )
+    def _stub(eqn, env, last_used):
+        pass
 
-    return module, context, jaxpr, tree_structure(shape)
+    old = jax_core.clean_up_dead_vars
+    jax_core.clean_up_dead_vars = _stub
+
+    try:
+        nrep = jaxpr_replicas(jaxpr)
[…]
