### 2023-11-13 1a1b465
Update tensor_init tests to use jax dynamic api
