+++ frontend/test/pytest/test_tensor_init.py
@@ -21,10 +21,10 @@ from jax import numpy as jnp
 from numpy import array_equal
 from numpy.testing import assert_allclose
 
-from catalyst import empty, ones, qjit, zeros
+from catalyst import qjit
 
 DTYPES = [float, int, jnp.float32, jnp.float64, jnp.int8, jnp.int16, "float32", np.float64]
-SHAPES = [3, (2, 3, 1), (), jnp.array([2, 1], dtype=int)]
+SHAPES = [3, (2, 3, 1), (), jnp.array([2, 1, 3], dtype=int)]
 
 
 def _assert_equal(a, b):
@@ -40,13 +40,13 @@ def test_interpretation(shape, dtype):
     """Test that tensor primitive work in the interpretation mode"""
     # pylint: disable=unnecessary-direct-lambda-call
 
-    _assert_equal((lambda: zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
-    _assert_equal((lambda: ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
-    _assert_equal((lambda s: ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
-    _assert_equal((lambda s: zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
+    _assert_equal((lambda: jnp.zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal((lambda: jnp.ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: jnp.ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: jnp.zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
 
     def f(s):
-        return empty(shape=s, dtype=dtype)
+        return jnp.empty(shape=s, dtype=dtype)
 
     res = f(shape)
     assert_allclose(res.shape, shape)
@@ -58,14 +58,15 @@ def test_interpretation(shape, dtype):
 def test_classical_tracing(shape, dtype):
     """Test that tensor primitive work in the classical tracing mode"""
 
-    _assert_equal(qjit(lambda: zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
-    _assert_equal(qjit(lambda: ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
-    _assert_equal(qjit(lambda s: ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
-    _assert_equal(qjit(lambda s: zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
+    _assert_equal(qjit(lambda: jnp.zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal(qjit(lambda: jnp.ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: jnp.ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: jnp.zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
 
     @qjit
     def f(s):
-        return empty(shape=s, dtype=dtype)
+        res = jnp.empty(shape=s, dtype=dtype)
+        return res
 
     res = f(shape)
     assert_allclose(res.shape, shape)
@@ -77,7 +78,7 @@ def test_classical_tracing_2():
 
     @qjit
     def f(x):
-        return ones(shape=[1, x], dtype=int)
[…]
