### 2023-11-13 fa14d02
Add abstracted_axes to compiled options.
