### 2023-11-13 002a4ed
Avoid recomputing runtime_signature.
