### 2023-11-13 8cd46fe
Update the installation guide w.r.t. backends
