+++ doc/dev/installation.rst
@@ -144,27 +144,36 @@ directory:
 
   make all
 
-To build each component one by one starting from the runtime, you can follow the instructions below.
+To build each component one by one starting from the runtime, or to build additional backend devices
+beyond ``lightning.qubit``, please follow the instructions below.
 
 Runtime
 """""""
 
-By default, the runtime is backed by `PennyLane-Lightning
-<https://github.com/PennyLaneAI/pennylane-lightning>`_
-requiring the use of C++20 standard library headers, and leverages the `QIR
-standard library <https://github.com/qir-alliance/qir-runner>`_. Assuming
-``libomp-dev`` is available, you can build the runtime from the top level
-directory:
+By default, the runtime builds and installs the `PennyLane-Lightning
+<https://github.com/PennyLaneAI/pennylane-lightning>`_ simulator device, which requires C++20
+standard library features. Older C++ compilers may not support this, so it is recommended to use a
+modern compiler with these features. An additional dependency, the `QIR
+standard library <https://github.com/qir-alliance/qir-runner>`_, is automatically fetched and
+built on supported platforms.
+
+From the root project directory, the runtime can then be built as follows:
 
 .. code-block:: console
 
   make runtime
 
-The runtime supports multiple backend devices, enabling the execution of quantum
-circuits locally on CPUs and GPUs, and remotely on Amazon Braket NISQ hardware.
-A list of supported backends, along with Make arguments for each device, is available in the
+Additional devices are constantly added, enabling the execution of quantum circuits on CPUs, GPUs,
+and remote services, such as Amazon Braket. The full list of supported backends, and additional
+configuration options, are available in the
 `Catalyst Runtime <https://docs.pennylane.ai/projects/catalyst/en/latest/modules/runtime.html>`_
 page.
+
+To install Catalyst with all available backends, simply run:
+
+.. code-block:: console
+
+  make runtime ENABLE_LIGHTNING_KOKKOS=ON ENABLE_OPENQASM=ON
 
 MLIR Dialects
 """""""""""""
