### 2023-11-13 dd1225f
Make OpenMP build options explicit in wheel recipe
