+++ .github/workflows/build-wheel-macos-arm64.yaml
@@ -244,9 +244,11 @@ jobs:
               -DENABLE_LIGHTNING_KOKKOS=ON \
               -DLIGHTNING_GIT_TAG="latest_release" \
               -DKokkos_ENABLE_SERIAL=ON \
+              -DKokkos_ENABLE_OPENMP=OFF \
               -DKokkos_ENABLE_COMPLEX_ALIGN=OFF \
               -DENABLE_WARNINGS=OFF \
-              -DENABLE_OPENQASM=ON
+              -DENABLE_OPENQASM=ON \
+              -ENABLE_OPENMP=OFF
 
         cmake --build runtime-build --target rt_capi
 
