### 2023-11-13 b659c6b
Update installation guide with extra test options
