+++ requirements.txt
@@ -19,5 +19,7 @@ pytest-xdist
 pytest-cov
 nbmake
 
-# optional rt dependencies
+# optional rt/test dependencies
 tensorflow
+amazon-braket-pennylane-plugin>=1.23.0
+pennylane-lightning[kokkos]
