+++ frontend/test/pytest/test_jax_dynamic_api.py
@@ -1,0 +1,150 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""This test fixes our expectations regarding the JAX dynamic API."""
+
+import numpy as np
+import pennylane as qml
+import pytest
+from jax import numpy as jnp
+from numpy import array_equal
+from numpy.testing import assert_allclose
+
+from catalyst import qjit
+
+DTYPES = [float, int, jnp.float32, jnp.float64, jnp.int8, jnp.int16, "float32", np.float64]
+SHAPES = [3, (2, 3, 1), (), jnp.array([2, 1, 3], dtype=int)]
+
+
+def _assert_equal(a, b):
+    """Check that two arrays have exactly the same values and types"""
+
+    assert array_equal(a, b)
+    assert a.dtype == b.dtype
+
+
+@pytest.mark.parametrize("dtype", DTYPES)
+@pytest.mark.parametrize("shape", SHAPES)
+def test_interpretation(shape, dtype):
+    """Test that tensor primitive work in the interpretation mode"""
+    # pylint: disable=unnecessary-direct-lambda-call
+
+    _assert_equal((lambda: jnp.zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal((lambda: jnp.ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: jnp.ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal((lambda s: jnp.zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
+
+    def f(s):
+        return jnp.empty(shape=s, dtype=dtype)
+
+    res = f(shape)
+    assert_allclose(res.shape, shape)
+    assert res.dtype == dtype
+
+
+@pytest.mark.parametrize("dtype", DTYPES)
+@pytest.mark.parametrize("shape", SHAPES)
+def test_classical_tracing(shape, dtype):
+    """Test that tensor primitive work in the classical tracing mode"""
+
+    _assert_equal(qjit(lambda: jnp.zeros(shape, dtype))(), jnp.zeros(shape, dtype=dtype))
+    _assert_equal(qjit(lambda: jnp.ones(shape, dtype))(), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: jnp.ones(s, dtype))(shape), jnp.ones(shape, dtype=dtype))
+    _assert_equal(qjit(lambda s: jnp.zeros(s, dtype))(shape), jnp.zeros(shape, dtype=dtype))
+
+    @qjit
+    def f(s):
+        res = jnp.empty(shape=s, dtype=dtype)
+        return res
+
+    res = f(shape)
+    assert_allclose(res.shape, shape)
+    assert res.dtype == dtype
+
+
+def test_classical_tracing_2():
+    """Test that tensor primitive work in the classical tracing mode, the traced dimention case"""
+
+    @qjit
+    def f(x):
+        return jnp.ones(shape=[1, x], dtype=int)
+
+    _assert_equal(f(3), jnp.ones((1, 3), dtype=int))
+
+
[…]
