+++ frontend/catalyst/pennylane_extensions.py
@@ -78,6 +78,7 @@ from catalyst.utils.jax_extras import (
     new_inner_tracer,
     tree_structure,
     wrap_init,
+    abstractify,
 )
 from catalyst.utils.patching import Patcher
 
@@ -1530,6 +1531,7 @@ def for_loop(lower_bound, upper_bound, step):
 
     return _body_query
 
+from jax._src.util import (unzip2)
 
 def while_loop(cond_fn):
     """A :func:`~.qjit` compatible while-loop decorator for PennyLane/Catalyst.
@@ -1605,6 +1607,8 @@ def while_loop(cond_fn):
                     arg_classical_tracers = _input_type_to_tracers(
                         cond_trace.new_arg, cond_in_avals
                     )
+                    _, keep_inputs = unzip2(cond_wffa.in_type)
+                    arg_classical_tracers = [t for t, keep in zip(arg_classical_tracers, keep_inputs) if keep]
                     res_classical_tracers = [
                         cond_trace.full_raise(t)
                         for t in cond_wffa.call_wrapped(*arg_classical_tracers)
@@ -1618,6 +1622,8 @@ def while_loop(cond_fn):
                 with EvaluationContext.frame_tracing_context(ctx) as body_trace:
                     wffa, in_avals, body_tree = deduce_avals(body_fn, init_state, {})
                     arg_classical_tracers = _input_type_to_tracers(body_trace.new_arg, in_avals)
+                    _, keep_inputs = unzip2(wffa.in_type)
+                    arg_classical_tracers = [t for t, keep in zip(arg_classical_tracers, keep_inputs) if keep]
                     quantum_tape = QuantumTape()
                     with QueuingManager.stop_recording(), quantum_tape:
                         res_classical_tracers = [
@@ -1629,7 +1635,7 @@ def while_loop(cond_fn):
                     )
 
                 res_avals = list(map(shaped_abstractify, res_classical_tracers))
-                out_classical_tracers = [new_inner_tracer(outer_trace, aval) for aval in res_avals]
+                out_classical_tracers = [outer_trace.full_raise(aval) for aval in res_avals]
 
                 WhileLoop(in_classical_tracers, out_classical_tracers, [cond_region, body_region])
                 return tree_unflatten(body_tree(), out_classical_tracers)
