+++ doc/changelog.md
@@ -13,12 +13,17 @@
 
 <h3>Breaking changes</h3>
 
+* ``AllocOp``, ``DeallocOp`` have now (only) value semantics. In the frontend, the last 
+  quantum register is deallocated instead of the first one.
+  [(#360)](https://github.com/PennyLaneAI/catalyst/pull/360)
+
 <h3>Bug fixes</h3>
 
 <h3>Contributors</h3>
 
 This release contains contributions from (in alphabetical order):
 
+Romain Moyard,
 Erick Ochoa Lopez.
 
 # Release 0.3.2
