+++ mlir/lib/Quantum/Transforms/AdjointPatterns.cpp
@@ -15,7 +15,6 @@
 #define DEBUG_TYPE "adjoint"
 
 #include <algorithm>
-#include <iostream>
 #include <iterator>
 #include <string>
 #include <unordered_map>
@@ -269,6 +268,96 @@ class AdjointGenerator {
         }
     }
 
+    void visitOperation(func::CallOp callOp, OpBuilder &builder)
+    {
+        // Get the the original function
+        SymbolRefAttr symbol = dyn_cast_if_present<SymbolRefAttr>(callOp.getCallableForCallee());
+        func::FuncOp funcOp =
+            dyn_cast_or_null<func::FuncOp>(SymbolTable::lookupNearestSymbolFrom(callOp, symbol));
+
+        std::optional<Value> returnedQureg =
+            getQuantumReg(funcOp.front().getTerminator()->getOperands());
+        if (!returnedQureg.has_value()) {
+            // This operation is purely classical
+            return;
+        }
+
+        // Save the insertion point to come back to it after creating the adjoint of the function
+        auto insertionSaved = builder.saveInsertionPoint();
+        MLIRContext *ctx = builder.getContext();
+
+        // Create the adjoint of the original function at the module level
+        ModuleOp moduleOp = funcOp->getParentOfType<ModuleOp>();
+        builder.setInsertionPointToStart(moduleOp.getBody());
+
+        Block *originalBlock = &funcOp.front();
+        Operation *originalTerminator = originalBlock->getTerminator();
+        ValueRange originalArguments = originalBlock->getArguments();
+
+        // Get the arguments and outputs types from the original function (same signature)
+        FunctionType adjointFnType =
+            FunctionType::get(ctx, /*inputs=*/
+                              originalArguments.getTypes(),
+                              /*outputs=*/originalTerminator->getOperandTypes());
+        std::string adjointName = funcOp.getName().str() + ".adjoint";
+        Location loc = funcOp.getLoc();
+        func::FuncOp adjointFnOp = builder.create<func::FuncOp>(loc, adjointName, adjointFnType);
+        adjointFnOp.setPrivate();
+
+        // Create the block of the adjoint function
+        Block *adjointFnBlock = adjointFnOp.addEntryBlock();
+        builder.setInsertionPointToStart(adjointFnBlock);
+        Type qregType = quantum::QuregType::get(builder.getContext());
+        auto argumentsSize = adjointFnOp.getArguments().size();
+
+        // The last argument is the quantum register
+        Value lastArg = adjointFnOp.getArgument(argumentsSize - 1);
+        assert(isa<QuregType>(lastArg.getType()) &&
+               "The last argument of the function must be the quantum register.");
+        quantum::AdjointOp adjointOp = builder.create<quantum::AdjointOp>(loc, qregType, lastArg);
+
+        Region *adjointRegion = &adjointOp.getRegion();
+        Region &originalRegion = funcOp.getRegion();
+
+        // Map the arguments with the original function
+        IRMapping map;
+        for (size_t i = 0; i < funcOp.getArguments().size() - 1; i++) {
+            Value arg = adjointFnOp.front().getArgument(i);
+            Value argBlock = originalRegion.front().getArgument(i);
+            map.map(argBlock, arg);
+        }
+
+        // Copy the original region in the adjoint region
+        originalRegion.cloneInto(adjointRegion, map);
+
+        // Replace the return of the quantum register with a quantum yield
+        auto terminator = adjointRegion->front().getTerminator();
+        ValueRange res = terminator->getOperands();
+        TypeRange resTypes = terminator->getResultTypes();
[…]
