### 2023-11-14 173db7f
Adjoint update
