+++ mlir/lib/OpenQASM/CMakeLists.txt
@@ -1,0 +1,6 @@
+add_mlir_library(MLIROpenQASM
+    OpenQASM.cpp
+
+    DEPENDS
+    MLIROpenQASMIncGen
+)
