+++ mlir/include/OpenQASM/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_mlir_dialect(OpenQASM oq)
