### 2023-11-14 5817973
Add warning for XCode dependency on macOS
