### 2023-11-15 2669dd0
Improve scripts and instructions for macOS & various minor fixes (#363)
List of changes:
- `tensorflow-cpu` doesn't exist for macOS, use regular `tensorflow`
instead in `requirements.txt`
- add instructions to the documentation for building the docs
- fix the default mhlo build directory not being ignored by git (now
`bazel-build`, as it has to be a directory already present in the
upstream `.gitignore`)
- fix an AutoGraph lit test which produces slightly different output
with different TF versions when assigning tuples
- update formatting exclusions (e.g. with Enzyme directories)
- fix test t
[…]
