+++ runtime/lib/capi/ExecutionContext.hpp
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <dlfcn.h>
 #include <functional>
 #include <memory>
 #include <string>
@@ -20,25 +19,36 @@
 #include <unordered_map>
 #include <unordered_set>
 
+#include <pybind11/embed.h>
+
 #include "Exception.hpp"
 #include "QuantumDevice.hpp"
 
-#if __has_include("LightningSimulator.hpp")
-// device: lightning.qubit
-#include "LightningSimulator.hpp"
-#endif
-
-#if __has_include("LightningKokkosSimulator.hpp")
-// device: lightning.kokkos
-#include "LightningKokkosSimulator.hpp"
-#endif
-
-#if __has_include("OpenQasmDevice.hpp")
-// device: openqasm
-#include "OpenQasmDevice.hpp"
+#ifdef __linux__
+#include <dlfcn.h>
+#elif defined(__APPLE__)
+#include <mach-o/dyld.h>
 #endif
 
 namespace Catalyst::Runtime {
+
+/**
+ * A (RAII) class for `pybind11::initialize_interpreter` and `pybind11::finalize_interpreter`.
+ *
+ * @note This is not copiable or movable and used in C++ tests and the ExecutionContext manager
+ * of the runtime to solve the issue with re-initialization of the Python interpreter in `catch2`
+ * tests which also enables the runtime to reuse the same interpreter in the scope of the global
+ * quantum device unique pointer.
+ */
+struct PythonInterpreterGuard {
+    PythonInterpreterGuard() { pybind11::initialize_interpreter(); }
+    ~PythonInterpreterGuard() { pybind11::finalize_interpreter(); }
+
+    PythonInterpreterGuard(const PythonInterpreterGuard &) = delete;
+    PythonInterpreterGuard(PythonInterpreterGuard &&) = delete;
+    PythonInterpreterGuard &operator=(const PythonInterpreterGuard &) = delete;
+    PythonInterpreterGuard &operator=(PythonInterpreterGuard &&) = delete;
+};
 
 class MemoryManager final {
   private:
@@ -71,12 +81,19 @@ class SharedLibraryManager final {
         // If you have compiled this file with sanitizers and you reach this line
         // you will get an error.
         // Please re-compile without sanitizers.
+        // std::cerr << "filename: " << filename << std::endl;
+
 #ifdef __APPLE__
         // macOS doesn't support RTLD_DEEPBIND
         _handler = dlopen(filename.c_str(), RTLD_LAZY);
 #else
-        _handler = dlopen(filename.c_str(), RTLD_LAZY | RTLD_DEEPBIND);
+        // Closing the dynamic library of Lightning simulators with dlclose() where
+        // OpenMP directives are in use would raise memory segfaults
+        // Ali: Adding RTLD_NODELETE would fix the issue.
+
+        _handler = dlopen(filename.c_str(), RTLD_LAZY | RTLD_NODELETE);
 #endif
+        // std::cerr << "_handler: " << _handler << std::endl;
         RT_FAIL_IF(!_handler, dlerror());
     }
 
@@ -106,6 +123,7 @@ class SharedLibraryManager final {
         // SharedLibraryManager to manage shared libraries.
         //
         // Exercise for the reader, how could one trigger the "cannot create scope list" error?
+
         dlclose(_handler);
     }
 
@@ -124,17 +142,9 @@ class ExecutionContext final {
   private:
     using DeviceInitializer =
         std::function<std::unique_ptr<QuantumDevice>(bool, const std::string &)>;
-    std::unordered_map<std::string_view, DeviceInitializer> _device_map{
-        {"lightning.qubit",
-         [](bool tape_recording, const std::string &device_kwargs) {
[…]
