+++ runtime/lib/backend/lightning/lightning_dynamic/LightningSimulator.cpp
@@ -519,3 +519,9 @@ void LightningSimulator::Gradient(std::vector<DataView<double, 1>> &gradients,
 }
 
 } // namespace Catalyst::Runtime::Simulator
+
+extern "C" Catalyst::Runtime::QuantumDevice *LightningSimulatorFactory(bool status,
+                                                                       const std::string &kwargs)
+{
+    return new Catalyst::Runtime::Simulator::LightningSimulator(status, kwargs);
+}
