+++ frontend/test/pytest/test_decomposition.py
@@ -53,6 +53,7 @@ class CustomDevice(qml.QubitDevice):
 dev = CustomDevice(wires=2)
 
 
+@pytest.mark.skip(reason="no way to test this with the plugin system")
 @pytest.mark.parametrize("param,expected", [(0.0, True), (jnp.pi, False)])
 def test_decomposition(param, expected):
     @qjit()

[…]
