+++ frontend/test/lit/test_custom_ops.py
@@ -49,6 +49,7 @@ class CustomDeviceWithoutSupport(qml.QubitDevice):
     operations = lightning.operations.copy()
     observables = lightning.observables.copy()
 
+    # pylint: disable=too-many-arguments
     def __init__(
         self, shots=None, wires=None, backend_name=None, backend_lib=None, backend_kwargs=None
     ):
