+++ frontend/test/lit/test_decomposition.py
@@ -39,6 +39,7 @@ def get_custom_device_without(num_wires, discards):
         operations = copy
         observables = discard
 
+        # pylint: disable=too-many-arguments
         def __init__(
             self, shots=None, wires=None, backend_name=None, backend_lib=None, backend_kwargs=None
         ):
