+++ frontend/catalyst/utils/runtime.py
@@ -1,0 +1,95 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""
+Runtime utility methods.
+"""
+
+import os
+import pathlib
+import platform
+
+import pennylane as qml
+
+from catalyst._configuration import INSTALLED
+from catalyst.utils.exceptions import CompileError
+
+package_root = os.path.dirname(__file__)
+
+
+# Default paths to dep libraries
+DEFAULT_LIB_PATHS = {
+    "llvm": os.path.join(package_root, "../../../mlir/llvm-project/build/lib"),
+    "runtime": os.path.join(package_root, "../../../runtime/build/lib"),
+    "enzyme": os.path.join(package_root, "../../../mlir/Enzyme/build/Enzyme"),
+}
+
+
+# The set of supported devices at runtime
+SUPPORTED_RT_DEVICES = {
+    "lightning.qubit": ("LightningSimulator", "librtd_lightning"),
+    "lightning.kokkos": ("LightningKokkosSimulator", "librtd_lightning"),
+    "braket.aws.qubit": ("OpenQasmDevice", "librtd_openqasm"),
+    "braket.local.qubit": ("OpenQasmDevice", "librtd_openqasm"),
+}
+
+
+def get_lib_path(project, env_var):
+    """Get the library path."""
+    if INSTALLED:
+        return os.path.join(package_root, "lib")  # pragma: no cover
+    return os.getenv(env_var, DEFAULT_LIB_PATHS.get(project, ""))
+
+
+def extract_backend_info(device):
+    """Extract the backend info as a tuple of (name, lib, kwargs)."""
+
+    dname = device.name
+    if isinstance(device, qml.Device):
+        dname = device.short_name
+
+    device_name = ""
+    device_lpath = ""
+    device_kwargs = {"device_type": dname}
+
+    if dname in SUPPORTED_RT_DEVICES:
+        device_name = SUPPORTED_RT_DEVICES[dname][0]
+        device_lpath = get_lib_path("runtime", "RUNTIME_LIB_DIR")
+        sys_platform = platform.system()
+
+        if sys_platform == "Linux":
+            device_lpath = os.path.join(device_lpath, SUPPORTED_RT_DEVICES[dname][1] + ".so")
+        elif sys_platform == "Darwin":
+            device_lpath = os.path.join(device_lpath, SUPPORTED_RT_DEVICES[dname][1] + ".dylib")
+        else:
+            raise NotImplementedError(f"Platform not supported: {sys_platform}")
+    elif hasattr(device, "get_c_interface"):
+        # Support for third party devices
+        device_name, device_lpath = device.get_c_interface()
+        if not pathlib.Path(device_lpath).is_file():
+            raise CompileError(f"Device at {device_lpath} cannot be found!")
+    else:
+        raise CompileError(f"The {dname} device is not supported for compilation at the moment.")
+
+    if hasattr(device, "shots"):
+        device_kwargs["shots"] = device.shots if device.shots else 0
+
+    if dname == "braket.local.qubit":  # pragma: no cover
+        device_kwargs["backend"] = device._device._delegate.DEVICE_ID
[…]
