### 2023-11-16 883d91f
Update README.md (#365)
Fixing sentence in docs
