### 2023-11-16 6ad7d5a
Adds very basic test.
