### 2023-11-16 2a6247c
Call make_jaxpr with abstracted_axes
