### 2023-11-16 a7a2d42
Pass arguments from Python to compiled function.
