+++ frontend/test/pytest/test_abstracted_axes.py
@@ -17,6 +17,7 @@ Tests for abstracted axes
 """
 from catalyst import qjit
 import jax
+import pennylane as qml
 from numpy.testing import assert_allclose
 
 class TestBasicInterface:
@@ -29,6 +30,22 @@ class TestBasicInterface:
         @qjit(abstracted_axes={0: "n"})
         def identity(a):
             return a
+
+        param = jax.numpy.array([1, 2, 3])
+        result = identity(param)
+        assert_allclose(param, result)
+        assert "tensor<?xi64>" in identity.mlir
+
+    def test_abstracted_axes_dictionary_2(self):
+        """This is a temporary test while dynamism is in development."""
+
+        @qml.qnode(qml.device("lightning.qubit", wires=1))
+        def foo(a):
+            return a
+
+        @qjit(abstracted_axes={0: "n"})
+        def identity(a):
+            return foo(a)
 
         param = jax.numpy.array([1, 2, 3])
         result = identity(param)
