+++ runtime/tests/Test_OpenQasmDevice.cpp
@@ -12,507 +12,555 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "ExecutionContext.hpp"
 #include "MemRefUtils.hpp"
 
+#include "ExecutionContext.hpp"
 #include "OpenQasmBuilder.hpp"
 #include "OpenQasmDevice.hpp"
 #include "OpenQasmRunner.hpp"
+#include "RuntimeCAPI.h"
+
+Catalyst::Runtime::PythonInterpreterGuard guard{};
 
 #include <catch2/catch.hpp>
 
 using namespace Catalyst::Runtime::Device;
 using BType = OpenQasm::BuilderType;
 
-Catalyst::Runtime::PythonInterpreterGuard guard{};
+// TEST_CASE("Test OpenQasmRunner base class", "[openqasm]")
+// {
+//     // check the coverage support
+//     OpenQasm::OpenQasmRunner runner{};
+//     REQUIRE_THROWS_WITH(runner.runCircuit("", "", 0),
+//                         Catch::Contains("[Function:runCircuit] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
 
-TEST_CASE("Test OpenQasmRunner base class", "[openqasm]")
+//     REQUIRE_THROWS_WITH(runner.Probs("", "", 0, 0),
+//                         Catch::Contains("[Function:Probs] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+
+//     REQUIRE_THROWS_WITH(runner.Sample("", "", 0, 0),
+//                         Catch::Contains("[Function:Sample] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+
+//     REQUIRE_THROWS_WITH(runner.Expval("", "", 0),
+//                         Catch::Contains("[Function:Expval] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+
+//     REQUIRE_THROWS_WITH(runner.Var("", "", 0),
+//                         Catch::Contains("[Function:Var] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+
+//     REQUIRE_THROWS_WITH(runner.State("", "", 0, 0),
+//                         Catch::Contains("[Function:State] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+
+//     REQUIRE_THROWS_WITH(runner.Gradient("", "", 0, 0),
+//                         Catch::Contains("[Function:Gradient] Error in Catalyst Runtime: "
+//                                         "Not implemented method"));
+// }
+
+// TEST_CASE("Test BraketRunner::runCircuit()", "[openqasm]")
+// {
+//     OpenQasm::BraketBuilder builder{};
+
+//     builder.Register(OpenQasm::RegisterType::Qubit, "q", 2);
+
+//     builder.Gate("Hadamard", {}, {}, {0}, false);
+//     builder.Gate("CNOT", {}, {}, {0, 1}, false);
+
+//     auto &&circuit = builder.toOpenQasm();
+
+//     OpenQasm::BraketRunner runner{};
+//     auto &&results = runner.runCircuit(circuit, "default", 100);
+//     CHECK(results.find("GateModelQuantumTaskResult") != std::string::npos);
+// }
+
+// TEST_CASE("Test the OpenQasmDevice constructor", "[openqasm]")
+// {
+//     SECTION("Common")
+//     {
+//         auto device = OpenQasmDevice(false, "{shots : 100}");
+//         CHECK(device.GetNumQubits() == 0);
+
+//         REQUIRE_THROWS_WITH(device.Circuit(),
[…]
