+++ runtime/tests/Test_DLManager.cpp
@@ -56,7 +56,7 @@ TEST_CASE("Test success of loading dummy device", "[Third Party]")
 }
 #endif
 
-TEST_CASE("Test __rt__device registering a custom device with shots=500 and device=lightning",
+TEST_CASE("Test __rt__device registering a custom device with shots=500 and device=lightning.qubit",
           "[CoreQIS]")
 {
     __quantum__rt__initialize();
@@ -83,49 +83,15 @@ TEST_CASE("Test __rt__device registering a custom device with shots=500 and devi
                         Catch::Contains("Invalid use of the global driver before initialization"));
 }
 
-TEST_CASE("Test __rt__device registering the OpenQasm device", "[CoreQIS]")
+#ifdef __device_lightning_kokkos
+TEST_CASE("Test __rt__device registering device=lightning.kokkos", "[CoreQIS]")
 {
     __quantum__rt__initialize();
 
     char dev[8] = "rtd_lib";
-    char dev_value[30] = "braket.aws.qubit";
-
-#if __has_include("OpenQasmDevice.hpp")
+    char dev_value[18] = "lightning.kokkos";
     __quantum__rt__device((int8_t *)dev, (int8_t *)dev_value);
-#else
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev, (int8_t *)dev_value),
-                        Catch::Contains("cannot open shared object file"));
-#endif
-
-    __quantum__rt__finalize();
-
-    __quantum__rt__initialize();
-
-    char dev_kwargs[20] = "rtd_kwargs";
-    char dev_value_kwargs[70] = "device_arn : arn:aws:braket:::device/quantum-simulator/amazon/sv1";
-
-    __quantum__rt__device((int8_t *)dev_kwargs, (int8_t *)dev_value_kwargs);
-
-#if __has_include("OpenQasmDevice.hpp")
-    __quantum__rt__device((int8_t *)dev, (int8_t *)dev_value);
-#else
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev, (int8_t *)dev_value),
-                        Catch::Contains("cannot open shared object file:"));
-#endif
-
-    __quantum__rt__finalize();
-
-    __quantum__rt__initialize();
-
-    char dev_lcl[8] = "rtd_lib";
-    char dev_value_lcl[30] = "braket.local.qubit";
-
-#if __has_include("OpenQasmDevice.hpp")
-    __quantum__rt__device((int8_t *)dev_lcl, (int8_t *)dev_value_lcl);
-#else
-    REQUIRE_THROWS_WITH(__quantum__rt__device((int8_t *)dev_lcl, (int8_t *)dev_value_lcl),
-                        Catch::Contains("cannot open shared object file"));
-#endif
 
     __quantum__rt__finalize();
 }
+#endif
