+++ runtime/tests/Test_OpenQasmDevice.cpp
@@ -27,496 +27,496 @@ Catalyst::Runtime::PythonInterpreterGuard guard{};
 using namespace Catalyst::Runtime::Device;
 using BType = OpenQasm::BuilderType;
 
-// TEST_CASE("Test OpenQasmRunner base class", "[openqasm]")
-// {
-//     // check the coverage support
-//     OpenQasm::OpenQasmRunner runner{};
-//     REQUIRE_THROWS_WITH(runner.runCircuit("", "", 0),
-//                         Catch::Contains("[Function:runCircuit] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.Probs("", "", 0, 0),
-//                         Catch::Contains("[Function:Probs] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.Sample("", "", 0, 0),
-//                         Catch::Contains("[Function:Sample] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.Expval("", "", 0),
-//                         Catch::Contains("[Function:Expval] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.Var("", "", 0),
-//                         Catch::Contains("[Function:Var] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.State("", "", 0, 0),
-//                         Catch::Contains("[Function:State] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-
-//     REQUIRE_THROWS_WITH(runner.Gradient("", "", 0, 0),
-//                         Catch::Contains("[Function:Gradient] Error in Catalyst Runtime: "
-//                                         "Not implemented method"));
-// }
-
-// TEST_CASE("Test BraketRunner::runCircuit()", "[openqasm]")
-// {
-//     OpenQasm::BraketBuilder builder{};
-
-//     builder.Register(OpenQasm::RegisterType::Qubit, "q", 2);
-
-//     builder.Gate("Hadamard", {}, {}, {0}, false);
-//     builder.Gate("CNOT", {}, {}, {0, 1}, false);
-
-//     auto &&circuit = builder.toOpenQasm();
-
-//     OpenQasm::BraketRunner runner{};
-//     auto &&results = runner.runCircuit(circuit, "default", 100);
-//     CHECK(results.find("GateModelQuantumTaskResult") != std::string::npos);
-// }
-
-// TEST_CASE("Test the OpenQasmDevice constructor", "[openqasm]")
-// {
-//     SECTION("Common")
-//     {
-//         auto device = OpenQasmDevice(false, "{shots : 100}");
-//         CHECK(device.GetNumQubits() == 0);
-
-//         REQUIRE_THROWS_WITH(device.Circuit(),
-//                             Catch::Contains("[Function:toOpenQasm] Error in Catalyst Runtime: "
-//                                             "Invalid number of quantum register"));
-//     }
-
-//     SECTION("Braket SV1")
-//     {
-//         auto device = OpenQasmDevice(
-//             false, "{shots: 100, device_type : braket.local.qubit, backend : default}");
-//         CHECK(device.GetNumQubits() == 0);
-
-//         REQUIRE_THROWS_WITH(device.Circuit(),
[…]
