+++ .github/workflows/check-catalyst.yaml
@@ -572,11 +572,12 @@ jobs:
       - name: Build Runtime test suite with both Lightning and Lightning-Kokkos simulators
         if: ${{ matrix.backend == 'lightning-kokkos' }}
         run: |
+          # ASAN fails w/ leaks, odr-violation, and double-free from Kokkos_Profiling.cpp
           C_COMPILER=$(which ${{ needs.constants.outputs[format('c_compiler.{0}', matrix.compiler)] }}) \
           CXX_COMPILER=$(which ${{ needs.constants.outputs[format('cxx_compiler.{0}', matrix.compiler)] }}) \
           COMPILER_LAUNCHER="" \
           ENABLE_LIGHTNING_KOKKOS=ON \
-          ENABLE_ASAN=ON \
+          ENABLE_ASAN=OFF \
           make test-runtime
 
       - name: Build Runtime test suite for OpenQasm device
