+++ mlir/lib/Catalyst/Transforms/RegisterAllPasses.cpp
@@ -32,4 +32,5 @@ void catalyst::registerAllCatalystPasses()
     mlir::registerPass(catalyst::createCopyGlobalMemRefPass);
     mlir::registerPass(catalyst::createCatalystConversionPass);
     mlir::registerPass(catalyst::createTestPass);
+    mlir::registerPass(catalyst::createHloCustomCallLoweringPass);
 }
