+++ mlir/include/Catalyst/Transforms/Passes.h
@@ -24,6 +24,7 @@ std::unique_ptr<mlir::Pass> createCatalystBufferizationPass();
 std::unique_ptr<mlir::Pass> createArrayListToMemRefPass();
 std::unique_ptr<mlir::Pass> createCatalystConversionPass();
 std::unique_ptr<mlir::Pass> createScatterLoweringPass();
+std::unique_ptr<mlir::Pass> createHloCustomCallLoweringPass();
 
 void registerAllCatalystPasses();
 
