### 2023-11-17 d37dfea
Tidy up code
