+++ frontend/catalyst/pennylane_extensions.py
@@ -29,6 +29,14 @@ import pennylane as qml
 from jax._src.api_util import shaped_abstractify
 from jax._src.lax.lax import _abstractify
 from jax._src.tree_util import PyTreeDef, tree_flatten, tree_unflatten, treedef_is_leaf
+from jax._src.interpreters.partial_eval import (DynamicJaxprTrace, JaxprStackFrame)
+
+
+from jax._src.util import unzip2
+from jax._src.pjit import _extract_implicit_args
+
+
+
 from jax.core import eval_jaxpr, get_aval
 from pennylane import QNode, QueuingManager
 from pennylane.measurements import MidMeasureMP
@@ -72,17 +80,19 @@ from catalyst.utils.jax_extras import (
     ShapedArray,
     _initial_style_jaxpr,
     _input_type_to_tracers,
-    abstractify,
+    # abstractify,
     convert_constvars_jaxpr,
     initial_style_jaxprs_with_common_consts1,
     initial_style_jaxprs_with_common_consts2,
     new_inner_tracer,
     tree_structure,
     wrap_init,
+    infer_lambda_input_type,
 )
 from catalyst.utils.patching import Patcher
 
 import jax._src.core as core
+from jax._src.core import DBIdx, InDBIdx, OutDBIdx, DShapedArray
 
 def _check_no_measurements(tape: QuantumTape) -> None:
     """Check the nested quantum tape for the absense of quantum measurements of any kind"""
@@ -169,10 +179,17 @@ class QFunc:
             device = self.device
 
         with EvaluationContext(EvaluationMode.QUANTUM_COMPILATION):
-            jaxpr, retval_tree = trace_quantum_function(self.func, device, args, kwargs, qnode)
+            closed_jaxpr, retval_tree = trace_quantum_function(self.func, device, args, kwargs, qnode)
 
         def _eval_jaxpr(*args):
-            return eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, *args)
+            print("=============+CJCJCJCJC")
+            print(closed_jaxpr.jaxpr)
+            print(closed_jaxpr.out_avals)
+            print("=============+CJCJCJCJC")
+            res = eval_jaxpr(closed_jaxpr.jaxpr, closed_jaxpr.consts, *args)
+            print("RRRRRRRRRRRR")
+            for t in res: print("- ", t)
+            return res
 
         args_data, _ = tree_flatten(args)
 
@@ -972,9 +989,13 @@ class WhileLoop(HybridOp):
             qreg_in = _input_type_to_tracers(body_trace.new_arg, [AbstractQreg()])[0]
             qrp_out = trace_quantum_tape(body_tape, device, qreg_in, ctx, body_trace)
             qreg_out = qrp_out.actualize()
-            body_jaxpr, _, body_consts = ctx.frames[body_trace].to_jaxpr2(
+            body_jaxpr, body_type, body_consts = ctx.frames[body_trace].to_jaxpr2(
                 res_classical_tracers + [qreg_out]
             )
+
+        print("BJBJBJBJBJBJ")
+        print(ClosedJaxpr(convert_constvars_jaxpr(body_jaxpr), ()))
+        print("BJBJBJBJBJBJ")
 
         qreg = qrp.actualize()
         qrp2 = QRegPromise(
@@ -1160,7 +1181,7 @@ class CondCallable:
         for branch in self.branch_fns + [self.otherwise_fn]:
             quantum_tape = QuantumTape()
             with EvaluationContext.frame_tracing_context(ctx) as inner_trace:
-                wffa, _, out_tree = deduce_avals(branch, [], {})
+                wffa, _, _, out_tree = deduce_avals(branch, [], {})
                 with QueuingManager.stop_recording(), quantum_tape:
[…]
