+++ frontend/test/pytest/test_jax_dynamic_api.py
@@ -100,6 +100,8 @@ def test_quantum_tracing():
             return (a, i)
 
         a2, _ = loop(a, i)
+        print("TEST_RET_VALUE")
+        print(a2)
         return a2
 
     result = f([2, 3])

[…]
