### 2023-11-17 9697443
Fixed named_shape in the caller.
