+++ doc/changelog.md
@@ -10,9 +10,10 @@
   [(#280)](https://github.com/PennyLaneAI/catalyst/pull/280)
 
 * The plugin system of Catalyst for backend devices. This system in runtime
-  is backed by `dlopen` to load libraries at runtime instead of linking them
-  at compile time. This provides further flexibility and extensibility to
-  Catalyst allowing users to add and load custom quantum devices dynamically.
+  is backed by `dlopen` and enables the devices to be packaged separately from
+  the runtime CAPI and loaded at run time instead of being linked at compile time.
+  This provides flexibility and extensibility to Catalyst allowing users to
+  load quantum devices dynamically.
   [(#343)](https://github.com/PennyLaneAI/catalyst/pull/343)
 
 <h3>Improvements</h3>
