### 2023-11-17 88d0166
Enable ASAN on macOS in our build scripts
Works with 'check-dialects', 'lit', and 'pytest' targets.

'test-demos' does not work due the preloading issue on macOS (whenever a
subprocess is invoked). The reason is that preloaded ASAN library is not
available to child processes, which manifests with Jupyter Notebooks
and pytest-xdist.
