+++ mlir/test/lit.cfg.py
@@ -21,9 +21,17 @@ python_executable = getattr(config, "python_executable", "python3.10")
 
 if "Address" in getattr(config, "llvm_use_sanitizer", ""):
     # With sanitized builds, Python tests require some preloading magic to run.
-    assert "Linux" in config.host_os, "Testing with sanitized builds requires Linux/Clang"
-
-    python_executable = f"LD_PRELOAD=$({config.host_cxx} -print-file-name=libclang_rt.asan-{config.host_arch}.so) {python_executable}"
+    if "Linux" in config.host_os:
+        python_executable = f"LD_PRELOAD=$({config.host_cxx} -print-file-name=libclang_rt.asan-{config.host_arch}.so) {python_executable}"
+    elif "Darwin" in config.host_os:
+        # It's important that the python executable discovered by CMake is the true interpreter
+        # binary. On macOS, it appears that there can be multiple "python" executables which are
+        # just wrappers around the actual interpreter. In that case, the DYLD_INSERT_LIBRARIES trick
+        # will not work, as it will not be forwarded to the final process.
+        # See https://stackoverflow.com/a/47853433 for more information.
+        python_executable = f"DYLD_INSERT_LIBRARIES=$({config.host_cxx} -print-file-name=libclang_rt.asan_osx_dynamic.dylib) {python_executable}"
+    else:
+        assert False, "Testing with sanitized builds requires Linux or MacOS"
 
 config.substitutions.append(("%PYTHON", python_executable))
 
