### 2023-11-17 742f1e1
Cleanup in-tree build files for `make frontend`
