### 2023-11-17 acf400b
Outline function to get implicit and explicit flat args.
