### 2023-11-17 d861aa7
Remove print statements from test.
