+++ frontend/catalyst/compilation_pipelines.py
@@ -48,6 +48,7 @@ from catalyst.utils.contexts import EvaluationContext
 from catalyst.utils.filesystem import WorkspaceManager
 from catalyst.utils.gen_mlir import inject_functions
 from catalyst.utils.patching import Patcher
+from catalyst.utils.jax_extras import get_implicit_and_explicit_flat_args
 
 # Required for JAX tracer objects as PennyLane wires.
 # pylint: disable=unnecessary-lambda
@@ -437,16 +438,6 @@ class CompiledFunction:
 
     def __call__(self, *args, **kwargs):
         if self.compile_options.abstracted_axes is not None:
-            def get_implicit_and_explicit_flat_args(options, *args, **kwargs):
-                from jax._src.pjit import _flat_axes_specs, _extract_implicit_args
-                from jax._src.interpreters import partial_eval as pe
-                axes_specs = _flat_axes_specs(options.abstracted_axes, *args, **kwargs)
-                explicit_args, in_tree = tree_flatten(args)
-                in_type = pe.infer_lambda_input_type(axes_specs, explicit_args)
-                in_avals = tuple(a for a, e in in_type if e)
-                implicit_args = _extract_implicit_args(in_type, explicit_args)
-                args_flat = [*implicit_args, *explicit_args]
-                return args_flat
 
             args = get_implicit_and_explicit_flat_args(self.compile_options, *args, **kwargs)
 
