### 2023-11-17 2cf2af8
Remove unused variables.
