+++ frontend/catalyst/utils/jax_extras.py
@@ -402,9 +402,8 @@ def new_inner_tracer(trace: DynamicJaxprTrace, aval) -> DynamicJaxprTracer:
 def get_implicit_and_explicit_flat_args(options, *args, **kwargs):
     """Get implicit arguments from explicit arguments and abstracted_axes."""
     axes_specs = _flat_axes_specs(options.abstracted_axes, *args, **kwargs)
-    explicit_args, in_tree = tree_flatten(args)
+    explicit_args, _ = tree_flatten(args)
     in_type = pe.infer_lambda_input_type(axes_specs, explicit_args)
-    in_avals = tuple(a for a, e in in_type if e)
     implicit_args = _extract_implicit_args(in_type, explicit_args)
     args_flat = [*implicit_args, *explicit_args]
     return args_flat
