### 2023-11-17 2b23715
Remove unused variables.
