### 2023-11-17 7dc5ca2
Remove comments in deduce_avals.
