+++ frontend/test/pytest/test_abstracted_axes.py
@@ -15,16 +15,17 @@
 """
 Tests for abstracted axes
 """
-from catalyst import qjit
 import jax
 import pennylane as qml
 from numpy.testing import assert_allclose
 
-class TestBasicInterface:
-    """Test thas abstracted_axes kwarg does not change any functionality for the time being
-    """
+from catalyst import qjit
 
-    def test_abstracted_axes_dictionary(self):
+
+class TestBasicInterface:
+    """Test thas abstracted_axes kwarg does not change any functionality for the time being"""
+
+    def test_identity(self):
         """This is a temporary test while dynamism is in development."""
 
         @qjit(abstracted_axes={0: "n"})
@@ -36,34 +37,36 @@ class TestBasicInterface:
         assert_allclose(param, result)
         assert "tensor<?xi64>" in identity.mlir
 
-    def test_abstracted_axes_dictionary_2(self):
+    def test_single_parameter_single_return_qnode(self):
         """This is a temporary test while dynamism is in development."""
 
         @qml.qnode(qml.device("lightning.qubit", wires=1))
-        def foo(a):
+        def circuit(a):
             return a
 
         @qjit(abstracted_axes={0: "n"})
         def identity(a):
-            return foo(a)
+            return circuit(a)
 
         param = jax.numpy.array([1, 2, 3])
         result = identity(param)
 
         assert_allclose(param, result)
+        assert "tensor<?xi64>" in identity.mlir
 
-    def test_abstracted_axes_dictionary_3(self):
+    def test_multiple_parameters_multiple_returns_qnode(self):
         """This is a temporary test while dynamism is in development."""
 
         @qml.qnode(qml.device("lightning.qubit", wires=1))
-        def foo(a, b):
+        def circuit(a, b):
             return a, b
 
         @qjit(abstracted_axes={0: "n"})
         def identity(a, b):
-            return foo(a, b)
+            return circuit(a, b)
 
         param = jax.numpy.array([1, 2, 3])
         result = identity(param, param)
 
         assert_allclose((param, param), result)
+        assert "tensor<?xi64>" in identity.mlir
