### 2023-11-17 b74da34
Use dlfcn.h on macOS
