+++ mlir/lib/Quantum/IR/QuantumOps.cpp
@@ -83,7 +83,6 @@ LogicalResult InsertOp::canonicalize(InsertOp insert, mlir::PatternRewriter &rew
         bool oneUse = extract.getResult().hasOneUse();
 
         if ((staticallyEqual || dynamicallyEqual) && oneUse) {
-            insert->getParentOp()->dump();
             rewriter.replaceOp(insert, extract.getQreg());
             rewriter.eraseOp(extract);
             return success();
