+++ mlir/lib/Catalyst/Transforms/HloCustomCallPatterns.cpp
@@ -1,6 +1,5 @@
 
 
-
 // Copyright 2023 Xanadu Quantum Technologies Inc.
 
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -32,10 +31,24 @@ struct HloCustomCallOpRewritePattern : public mlir::OpRewritePattern<mhlo::Custo
     mlir::LogicalResult matchAndRewrite(mhlo::CustomCallOp op,
                                         mlir::PatternRewriter &rewriter) const override
     {
+        auto *ctx = rewriter.getContext();
+        auto fnType = FunctionType::get(
+            ctx, /*inputs=*/
+            op.getOperandTypes(),
+            /*outputs=*/op.getResultTypes());
+        auto calleeName = op.getCallTargetName();
+        assert(calleeName=="lapack_dgesdd" && "This custom call is currently not supported.");
+        ModuleOp moduleOp = op->getParentOfType<ModuleOp>();
+        auto savedPoint = rewriter.saveInsertionPoint();
+        rewriter.setInsertionPointToStart(moduleOp.getBody());
+        auto declaration = rewriter.create<func::FuncOp>(op.getLoc(), calleeName, fnType);
+        declaration.setPrivate();
+        rewriter.restoreInsertionPoint(savedPoint);
+        auto operands = op.getOperands();
+        TypeRange resultsType = op.getResultTypes();
+        rewriter.replaceOpWithNewOp<func::CallOp>(op, declaration.getName(), resultsType, operands);
         return success();
     }
-
-
 };
 
 void populateHloCustomCallPatterns(RewritePatternSet &patterns)
