+++ frontend/catalyst/compiler.py
@@ -121,6 +121,7 @@ DEFAULT_PIPELINES = [
             "func.func(scalarize)",
             "canonicalize",
             "scatter-lowering",
+            "hlo-custom-call-lowering",
         ],
     ),
     (
