### 2023-11-18 87790e8
Add *runtime-all target to build all devices
