+++ Makefile
@@ -76,7 +76,10 @@ dialects:
 	$(MAKE) -C mlir dialects
 
 runtime:
-	$(MAKE) -C runtime all
+	$(MAKE) -C runtime runtime
+
+runtime-all:
+	$(MAKE) -C runtime runtime ENABLE_LIGHTNING_KOKKOS=ON ENABLE_OPENQASM=ON
 
 dummy_device:
 	$(MAKE) -C runtime dummy_device
@@ -86,6 +89,9 @@ test: test-runtime test-frontend test-demos
 
 test-runtime:
 	$(MAKE) -C runtime test
+
+test-runtime-all:
+	$(MAKE) -C runtime test ENABLE_LIGHTNING_KOKKOS=ON ENABLE_OPENQASM=ON
 
 test-frontend: lit pytest
 
