### 2023-11-18 dba14e9
[Frontend & MLIR] Last quantum register is dealloc: Quantum AllocOp and DeallOc use Value semantics (#360)
**Context:**

AllocOp has both memory semantics and value semantics. We want to be
able to use the last quantum register of quantum functions.

**Description of the Change:**

- Dealloc only the last quantum register
- AllocOp and DealloOp have Value semantics only

**Benefits:**

- Quantum passes can use the last register.

**Possible Drawbacks:**

- Cse eliminates two similar quantum allocs (because pure)
- It would not work with multiple allocation in the same function

---------

Co-a
[…]
