+++ doc/changelog.md
@@ -11,6 +11,11 @@
 
 <h3>Improvements</h3>
 
+* ``AllocOp``, ``DeallocOp`` have now (only) value semantics. In the frontend, the last 
+  quantum register is deallocated instead of the first one. This allows to return the quantum
+  register in functions and can be given to another function (useful for quantum transformation).
+  [(#360)](https://github.com/PennyLaneAI/catalyst/pull/360)
+
 <h3>Breaking changes</h3>
 
 <h3>Bug fixes</h3>
@@ -19,6 +24,7 @@
 
 This release contains contributions from (in alphabetical order):
 
+Romain Moyard,
 Erick Ochoa Lopez.
 
 # Release 0.3.2
