### 2023-11-18 7a4ce6b
Enable address sanitizers on macOS in our build scripts (#368)
Adjusted configuration in Makefiles and lit config files to successfully
run tests with ASAN on macOS.

Tested: `make dialects`, `make lit`, `make pytest`, `make test-demos`
(non-functional), `make runtime`, `make runtime-all` (new), `make
test-runtime`, `make test-runtime-all` (new), `make test-mlir` (new),
`make coverage-frontend`

The main issue on this platform is related to the need to preload the
ASAN runtime library for the process that will end up loading a
sanitized shared library. Similar to `LD_PRELOAD` on Linux, this is
[…]
