+++ doc/changelog.md
@@ -19,6 +19,10 @@
   register in functions and can be given to another function (useful for quantum transformation).
   [(#360)](https://github.com/PennyLaneAI/catalyst/pull/360)
 
+* The quantum adjoint compiler routine has been extended to support function calls (with single quantum register returned) 
+  within an adjoint region. By itself this provides no user-facing changes, but compiler pass developers may now generate quantum adjoint operations around a block of code containing function calls as well as quantum operations and control flow operations.
+  [(#353)](https://github.com/PennyLaneAI/catalyst/pull/353)
+
 <h3>Breaking changes</h3>
 
 <h3>Bug fixes</h3>
