### 2023-11-20 0fbfa96
Add assert for quantum classical
