### 2023-11-20 e498583
Apply suggestions from code review
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
