+++ doc/changelog.md
@@ -9,9 +9,6 @@
   `MeasurementProcess`es.
   [(#280)](https://github.com/PennyLaneAI/catalyst/pull/280)
 
-* The adjoint operation now supports call operations with single quantum register returned.
-  [(#353)](https://github.com/PennyLaneAI/catalyst/pull/353)
-
 <h3>Improvements</h3>
 
 * ``AllocOp``, ``DeallocOp`` have now (only) value semantics. In the frontend, the last 
@@ -19,8 +16,12 @@
   register in functions and can be given to another function (useful for quantum transformation).
   [(#360)](https://github.com/PennyLaneAI/catalyst/pull/360)
 
-* The quantum adjoint compiler routine has been extended to support function calls (with single quantum register returned) 
-  within an adjoint region. By itself this provides no user-facing changes, but compiler pass developers may now generate quantum adjoint operations around a block of code containing function calls as well as quantum operations and control flow operations.
+* The quantum adjoint compiler routine has been extended to support function calls that affect the
+  quantum state within an adjoint region. Note that the function may only provide a single result
+  consisting of the quantum register.
+  By itself this provides no user-facing changes, but compiler pass developers may now generate
+  quantum adjoint operations around a block of code containing function calls as well as quantum
+  operations and control flow operations.
   [(#353)](https://github.com/PennyLaneAI/catalyst/pull/353)
 
 <h3>Breaking changes</h3>
