### 2023-11-20 2af8dec
[MLIR] Adjoint for CallOp (#353)
**Context:**
The adjoint operation region is failing if the region contains call
operation.

**Description of the Change:**

- We add support for call operation that only return a quantum register.
- Create an adjoint func op (same as the original but the region has
adjoint applied on it.) and create a call op that replaces the original
one

**Benefits:**

- More support for adjoint

---------

Co-authored-by: David Ittah <dime10@users.noreply.github.com>
