### 2023-11-21 09a244b
Do not use shaped types in quantum tracing
