+++ frontend/catalyst/jax_tracer.py
@@ -69,6 +69,7 @@ from catalyst.utils.jax_extras import (
     sort_eqns,
     tree_structure,
     tree_unflatten,
+    tree_flatten,
     unzip2,
     wrap_init,
 )
@@ -350,7 +351,7 @@ def trace_to_mlir(func, abstracted_axes, *args, **kwargs):
         jaxpr, unflat_shape = jax.make_jaxpr(func, **make_jaxpr_kwargs)(*args, **kwargs)
 
     module, context, jaxpr, tree_structure = jaxpr_to_mlir(func.__name__, jaxpr, unflat_shape)
-    flat_shape, tree = jax.tree_util.tree_flatten(unflat_shape)
+    flat_shape, tree = tree_flatten(unflat_shape)
     return module, context, jaxpr, tree_structure, flat_shape
 
 
@@ -669,7 +670,7 @@ def split_tracers_and_measurements(flat_values):
     return classical, measurements
 
 
-def trace_post_processing(ctx, trace, post_processing, args_types, args):
+def trace_post_processing(ctx, trace, post_processing, pp_args):
     """Trace post processing function.
 
     Args:
@@ -694,27 +695,18 @@ def trace_post_processing(ctx, trace, post_processing, args_types, args):
 
         # We need to deduce the type/shape/tree of the post_processing.
 
-        _, tree = jax.tree_util.tree_flatten(args_types)
-        aval_flat_args = [arg.aval for arg in args]
-        args_to_deduce = tree_unflatten(tree, aval_flat_args)
-        wffa, _, _, out_tree_promise = deduce_avals(post_processing, (args_to_deduce,), {})
+        # _, tree = jax.tree_util.tree_flatten(args_types)
+        # aval_flat_args = [arg.aval for arg in args]
+        # args_to_deduce = tree_unflatten(tree, aval_flat_args)
+        in_tracers = [trace.full_raise(t) for t in tree_flatten(pp_args)[0]]
+        wffa, _, _, out_tree_promise = deduce_avals(post_processing, (in_tracers,), {})
 
         # wffa will take as an input a flatten tracers.
-        post_processing_retval_flat = wffa.call_wrapped(*args)
-
         # After wffa is called, then the shape becomes available in out_tree_promise.
-        post_processing_tracers = [trace.full_raise(t) for t in post_processing_retval_flat]
-        jaxpr, out_type, consts = ctx.frames[trace].to_jaxpr2(post_processing_tracers)
-        out_avals = [out_aval for out_aval, keep in out_type if keep]
+        out_tracers = [trace.full_raise(t) for t in wffa.call_wrapped(*in_tracers)]
+        jaxpr, out_type, consts = ctx.frames[trace].to_jaxpr2(out_tracers)
         closed_jaxpr = ClosedJaxpr(jaxpr, consts)
-        # pylint: disable=unnecessary-lambda-assignment
-        named_shape = lambda x: getattr(x, "named_shape", {})
-        post_processing_results = tree_unflatten(
-            out_tree_promise(),
-            [ShapeDtypeStruct(a.shape, a.dtype, named_shape(a)) for a in out_avals],
-        )
-
-        return closed_jaxpr, post_processing_results, out_type
+        return closed_jaxpr, out_type, out_tree_promise()
 
 
 def trace_quantum_function(
@@ -774,7 +766,7 @@ def trace_quantum_function(
             tapes, post_processing = apply_transform(qnode, quantum_tape, return_values_flat)
 
         # (2) - Quantum tracing
-        results_tracers, results_abstract = [], []
+        transformed_results = []
         is_program_transformed = qnode and qnode.transform_program
         for tape in tapes:
             # If the program is batched, that means that it was transformed.
[…]
