### 2023-11-21 8996101
Fix quantum transform post-processing type deduction
