### 2023-11-21 21d37bd
Monkeypatch Jax get_aval to stay compatible with the upstream Jax
