+++ frontend/catalyst/utils/jax_extras.py
@@ -15,29 +15,33 @@
 """
 from __future__ import annotations
 
-from contextlib import contextmanager
+from contextlib import ExitStack, contextmanager
 from typing import Any, Callable, Dict, Generator, List, Optional, Sequence, Set, Type
 
+import jax
 from jax import ShapeDtypeStruct
 from jax._src import state, util
-from jax._src.core import _update_thread_local_jit_state, eval_jaxpr, Var, DShapedArray
+from jax._src.core import _update_thread_local_jit_state, eval_jaxpr, concrete_aval
 from jax._src.dispatch import jaxpr_replicas
 from jax._src.effects import ordered_effects as jax_ordered_effects
 from jax._src.interpreters import partial_eval as pe
 from jax._src.interpreters.mlir import _module_name_regex
-from jax._src.interpreters.partial_eval import _input_type_to_tracers
+from jax._src.interpreters.partial_eval import (
+    _input_type_to_tracers, infer_lambda_input_type, trace_to_jaxpr_dynamic2, _arg_type,
+    _collect_implicit, _complete_specs, _canonicalize_specs
+)
 from jax._src.lax.control_flow import _initial_style_jaxpr, _initial_style_open_jaxpr
 from jax._src.lax.lax import _abstractify, xla
-from jax._src.linear_util import annotate
+from jax._src.linear_util import annotate, _check_input_type
 from jax._src.pjit import _extract_implicit_args, _flat_axes_specs
 from jax._src.sharding_impls import ReplicaAxisContext
 from jax._src.source_info_util import current as jax_current
 from jax._src.source_info_util import new_name_stack
-from jax._src.util import partition_list, safe_map, unzip2, unzip3, wrap_name
-from jax.api_util import flatten_fun
-from jax.core import ClosedJaxpr, Jaxpr, JaxprEqn, MainTrace
+from jax._src.util import partition_list, safe_map, unzip2, unzip3, wrap_name, wraps
+from jax.api_util import flatten_fun, shaped_abstractify
+from jax.core import ClosedJaxpr, Jaxpr, JaxprEqn, MainTrace, OutputType, AbstractValue, Tracer
 from jax.core import Primitive as JaxprPrimitive
-from jax.core import ShapedArray, Trace, gensym, thread_local_state
+from jax.core import Var, DShapedArray, ShapedArray, Trace, gensym, thread_local_state
 from jax.interpreters.mlir import (
     AxisContext,
     ModuleContext,
@@ -79,7 +83,9 @@ __all__ = (
     "_initial_style_jaxpr",
     "_input_type_to_tracers",
     "jaxpr_to_mlir",
+    "jaxpr_remove_implicit",
     "make_jaxpr_effects",
+    "make_jaxpr2"
     "new_dynamic_main2",
     "new_inner_tracer",
     "pytree",
@@ -270,7 +276,7 @@ def deduce_avals(f: Callable, args, kwargs):
     flat_args, in_tree = tree_flatten((args, kwargs))
     abstracted_axes = None
     axes_specs = _flat_axes_specs(abstracted_axes, *args, **kwargs)
-    in_type = pe.infer_lambda_input_type(axes_specs, flat_args)
+    in_type = infer_lambda_input_type(axes_specs, flat_args)
     in_avals, keep_inputs = unzip2(in_type)
     wf = wrap_init(f)
     wff, out_tree_promise = flatten_fun(wf, in_tree)
@@ -278,19 +284,17 @@ def deduce_avals(f: Callable, args, kwargs):
     return wffa, in_avals, keep_inputs, out_tree_promise
 
 
-def jaxpr_to_mlir(func_name, jaxpr, shape):
[…]
