### 2023-11-21 8c240b5
Simplify the quantum qfunc binding
