+++ mlir/lib/Driver/CompilerDriver.cpp
@@ -44,6 +44,8 @@
 #include "Gradient/Transforms/Passes.h"
 #include "Quantum/IR/QuantumDialect.h"
 #include "Quantum/Transforms/Passes.h"
+#include "Mitigation/IR/MitigationDialect.h"
+#include "Mitigation/Transforms/Passes.h"
 
 #include "Enzyme.h"
 
@@ -145,6 +147,7 @@ void registerAllCatalystDialects(DialectRegistry &registry)
     registry.insert<CatalystDialect>();
     registry.insert<quantum::QuantumDialect>();
     registry.insert<gradient::GradientDialect>();
+    registry.insert<mitigation::MitigationDialect>();
 }
 } // namespace
 
