### 2023-11-22 7836fad
Remove the old-style in_type inference, always infer implicit input arguments
