### 2023-11-22 0a279ee
Simplify make_jaxpr2
