+++ frontend/test/lit/test_jax_dynamic_api.py
@@ -18,18 +18,22 @@
 
 import jax.numpy as jnp
 import pennylane as qml
+
 from catalyst import qjit
+
 
 def print_attr(f, attr, *args, **kwargs):
     name = f"TEST {f.__name__}"
-    print("\n"+'-'*len(name))
+    print("\n" + "-" * len(name))
     print(f"{name}\n")
     res = f(*args, **kwargs)
     print(getattr(f, attr))
     return res
 
+
 def print_jaxpr(f, *args, **kwargs):
     return print_attr(f, "jaxpr", *args, **kwargs)
+
 
 def print_mlir(f, *args, **kwargs):
     return print_attr(f, "mlir", *args, **kwargs)
@@ -48,7 +52,6 @@ print_mlir(test_qjit_dynamic_argument, jnp.array([1, 2, 3]))
 # CHECK-LABEL: test_qnode_dynamic_arg
 @qjit(abstracted_axes={0: "n"})
 def test_qnode_dynamic_arg(a):
-
     # CHECK:       { lambda ; [[a:.]]:i64[] [[b:.]]:i64[[[a]]]. let
     # CHECK:         [[c:.]]:i64[[[a]]] = func[
     # CHECK:                                  ] [[a]] [[b]]
@@ -66,7 +69,6 @@ print_jaxpr(test_qnode_dynamic_arg, jnp.array([1, 2, 3]))
 # CHECK-LABEL: test_qjit_dynamic_result
 @qjit
 def test_qjit_dynamic_result(a):
-
     # CHECK:       { lambda ; [[a:.]]:i64[]. let
     # CHECK:         [[b:.]]:i64[] = add [[a]] 1
     # CHECK:         [[c:.]]:f64[[[b]]] = {{[a-z_0-9.]+\[[^]]*\]}} 1.0 [[b]]
@@ -77,11 +79,9 @@ def test_qjit_dynamic_result(a):
 print_jaxpr(test_qjit_dynamic_result, 3)
 
 
-
 # CHECK-LABEL: test_qnode_dynamic_result
 @qjit
 def test_qnode_dynamic_result(a):
-
     # CHECK:       { lambda ; [[a:.]]:i64[]. let
     # CHECK:         [[b:.]]:i64[] [[c:.]]:f64[[[b]]] = func[
     # CHECK:                                                ] [[a]]
@@ -94,4 +94,3 @@ def test_qnode_dynamic_result(a):
 
 
 print_jaxpr(test_qnode_dynamic_result, 3)
-
