+++ frontend/test/pytest/test_jax_dynamic_api.py
@@ -1,0 +1,248 @@
+# Copyright 2023 Xanadu Quantum Technologies Inc.
+
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+
+#     http://www.apache.org/licenses/LICENSE-2.0
+
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""This test fixes our expectations regarding the JAX dynamic API."""
+
+import numpy as np
+import pennylane as qml
+import pytest
+from jax import numpy as jnp
+from numpy import array_equal
+from numpy.testing import assert_allclose
+
+from catalyst import qjit, while_loop
+
+DTYPES = [float, int, jnp.float32, jnp.float64, jnp.int8, jnp.int16, "float32", np.float64]
+SHAPES = [3, (2, 3, 1), (), jnp.array([2, 1, 3], dtype=int)]
+
+
+def _assert_equal(a, b):
+    """Check that two arrays have exactly the same values and types"""
+
+    assert array_equal(a, b)
+    assert a.dtype == b.dtype
+
+
+def test_qjit_abstracted_axes():
+    """Test that qjit accepts dynamical arguments."""
+
+    @qjit(abstracted_axes={0: "n"})
+    def identity(a):
+        return a
+
+    param = jnp.array([1, 2, 3])
+    result = identity(param)
+    _assert_equal(param, result)
+    assert "tensor<?xi64>" in identity.mlir, identity.mlir
+
+
+def test_qnode_abstracted_axis():
+    """Test that qnode accepts dynamical arguments."""
+
+    @qml.qnode(qml.device("lightning.qubit", wires=1))
+    def circuit(a):
+        return a
+
+    @qjit(abstracted_axes={0: "n"})
+    def identity(a):
+        return circuit(a)
+
+    param = jnp.array([1, 2, 3])
+    result = identity(param)
+
+    _assert_equal(param, result)
+    assert "tensor<?xi64>" in identity.mlir, func.mlir
+
+
+def test_qnode_dynamic_structured_args():
+    """Test that qnode accepts dynamically-shaped structured args"""
+
+    @qml.qnode(qml.device("lightning.qubit", wires=1))
+    def circuit(a_b):
+        return a_b[0] + a_b[1]
+
+    @qjit(abstracted_axes={0: "n"})
+    def func(a_b):
+        return circuit(a_b)
+
+    param = jnp.array([1, 2, 3])
+    c = func((param, param))
+    _assert_equal(c, param + param)
+    assert "tensor<?xi64>" in func.mlir, func.mlir
+
+
+def test_qnode_dynamic_structured_results():
+    """Test that qnode returns dynamically-shaped results"""
+
+    @qml.qnode(qml.device("lightning.qubit", wires=1))
+    def circuit(a):
+        return jnp.ones((a + 1,)), jnp.ones(
+            (a + 2),
+        )
+
+    @qjit
+    def func(a):
+        return circuit(a)
+
+    a, b = func(3)
+    assert_allclose(a, jnp.ones((4,)))
+    assert_allclose(b, jnp.ones((5,)))
+    assert "tensor<?xf64>" in func.mlir, func.mlir
+
+
+@pytest.mark.parametrize("dtype", DTYPES)
[…]
