### 2023-11-22 3daa32b
Address pylint issues
