+++ frontend/catalyst/ag_primitives.py
@@ -68,7 +68,6 @@ __all__ = [
 
 def get_program_length(reference_tracers):
     """Get the current number of instructions of the quantum and classical program."""
-    # pylint: disable=unnecessary-dunder-call
 
     num_jaxpr_eqns, num_tape_ops = 0, 0
 
@@ -87,7 +86,6 @@ def get_program_length(reference_tracers):
 
 def reset_program_to_length(reference_tracers, num_jaxpr_eqns, num_tape_ops):
     """Reset the quantum and classical program back to a given length."""
-    # pylint: disable=unnecessary-dunder-call
 
     if EvaluationContext.is_tracing():  # pragma: no branch
         jaxpr_frame = EvaluationContext.find_jaxpr_frame(reference_tracers)
