### 2023-11-22 c433438
Address codefactor issue
