### 2023-11-22 4be27d9
Adds initial version of spec.
