+++ runtime/lib/backend/lightning/lightning-spec.toml
@@ -51,8 +51,14 @@ full = [
 	"MultiRZ",
 ]
 matrix = [
-	# "QubitUnitary",
+	# QubitUnitary should never be here.
+	# Gates in "matrix" are decomposed to QubitUnitary.
+	# Having a QubitUnitary here would mean an infinite loop.
 	"OrbitalRotation",
+	"MultiControlledX",
+	# TODO:
+	# Why is qml.ops.Controlled not in qml namespace?
+	# TODO:
 	# other gates supported via applying matrices
 ]
 
