### 2023-11-22 ef98028
Add spec for lightning.kokkos.
