+++ runtime/lib/backend/CMakeLists.txt
@@ -1,6 +1,7 @@
 if(ENABLE_LIGHTNING OR ENABLE_LIGHTNING_KOKKOS)
 add_subdirectory(lightning)
-configure_file(lightning/lightning.qubit.toml lightning.qubit.toml)
+configure_file(lightning/lightning_dynamic/lightning.qubit.toml lightning.qubit.toml)
+configure_file(lightning/lightning_kokkos/lightning.kokkos.toml lightning.kokkos.toml)
 endif()
 if(ENABLE_OPENQASM)
 add_subdirectory(openqasm)
