### 2023-11-22 f988643
Add support for custom devices.
