+++ frontend/test/lit/test_custom_ops.py
@@ -18,6 +18,7 @@ import pennylane as qml
 from pennylane.operation import Operation
 
 from catalyst import measure, qjit
+from catalyst.compiler import get_lib_path
 
 # This is used just for internal testing
 from catalyst.pennylane_extensions import qfunc
@@ -56,10 +57,16 @@ class CustomDeviceWithoutSupport(qml.QubitDevice):
         self.backend_name = backend_name if backend_name else "default"
         self.backend_lib = backend_lib if backend_lib else "default"
         self.backend_kwargs = backend_kwargs if backend_kwargs else ""
-        super().__init__(wires=wires, shots=shots)
+        self.backend_path = CustomDeviceWithoutSupport.get_c_interface()
+        super().__init__(shots=shots, wires=wires)
 
     def apply(self, operations, **kwargs):
         pass
+
+    @staticmethod
+    def get_c_interface():
+        """Location to shared object with C/C++ implementation"""
+        return get_lib_path("runtime", "RUNTIME_LIB_DIR") + "/libdummy_device.so"
 
 
 operations = lightning.operations.copy()
@@ -70,7 +77,7 @@ class CustomDeviceWithSupport(CustomDeviceWithoutSupport):
     operations = operations
     observables = lightning.observables.copy()
 
-    def __init__(self, shots=None, wires=None):
+    def __init__(self, shots=None, wires=None, backend_name=None, backend_path=None, backend_kwargs=None):
         super().__init__(wires=wires, shots=shots)
 
 
