### 2023-11-23 97e4664
Add openqasm device spec.
