### 2023-11-23 e33a0ff
Coverage for custom devices with location possibly different than get_lib_path.
