### 2023-11-23 c0ef87c
Style fixes for test_decomposition.
