### 2023-11-23 b8da704
Style changes to pennylane_extensions.py
