### 2023-11-23 43770cf
Style changes for test_custom_ops
