+++ mlir/lib/Catalyst/Transforms/catalyst_to_llvm.cpp
@@ -15,6 +15,7 @@
 #include <unordered_map>
 
 #include "mlir/Conversion/LLVMCommon/ConversionTarget.h"
+#include "mlir/Conversion/LLVMCommon/MemRefBuilder.h"
 #include "mlir/Conversion/LLVMCommon/TypeConverter.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
@@ -232,11 +233,6 @@ struct PrintOpPattern : public OpConversionPattern<PrintOp> {
     }
 };
 
-struct EncodedArguments {
-    LLVM::AllocaOp encoded; // `args` argument
-    SmallVector<LLVM::AllocaOp> values;
-};
-
 struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
     using OpConversionPattern::OpConversionPattern;
 
@@ -265,44 +261,46 @@ struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
                                            operandsConverted.end()};
 
         // Encode args
-        EncodedArguments arguments;
-
-        for (auto tuple : llvm::drop_begin(llvm::zip(operands, operandsConverted))) {
-            auto encoded_arg = encodeMemref(std::get<0>(tuple), std::get<1>(tuple));
-            encoded.push_back(*encoded_arg);
+        Value c1 = rewriter.create<LLVM::ConstantOp>(loc, rewriter.getI64IntegerAttr(1));
+        SmallVector<mlir::LLVM::AllocaOp> encoded;
+        for (auto tuple : llvm::drop_begin(llvm::zip(args, argsConverted))) {
+            auto memref_type = std::get<0>(tuple).getType().cast<MemRefType>();
+            auto encoded_arg = EncodeMemRef(loc, rewriter, memref_type, std::get<1>(tuple));
+            LLVM::AllocaOp alloca = rewriter.create<LLVM::AllocaOp>(loc, ptr, type, c1, 0);
+            // Start the lifetime of encoded value.
+            rewriter.create<LLVM::LifetimeStartOp>(loc, rewriter.getI64IntegerAttr(-1), alloca);
+            // Use volatile store to suppress expensive LLVM optimizations.
+            rewriter.create<LLVM::StoreOp>(loc, encoded_arg, alloca, /*alignment=*/0, /*isVolatile=*/true);
+            encoded.push_back(alloca);
         }
 
         // We store encoded arguments as `!llvm.array<ptr x len>`.
-        size_t len = 2 + encoded.size();
+        size_t len = encoded.size();
         Type ptr = LLVM::LLVMPointerType::get(rewriter.getContext());
         Type type = LLVM::LLVMArrayType::get(ptr, len);
 
         // Prepare an array for encoded arguments.
-        Value arr = rewriter.create<LLVM::UndefOp>(type);
+        Value arr = rewriter.create<LLVM::UndefOp>(loc, type);
         auto insert_value = [&](Value value, int64_t offset) {
-            arr = rewriter.create<LLVM::InsertValueOp>(arr, value, offset);
+            arr = rewriter.create<LLVM::InsertValueOp>(loc, arr, value, offset);
         };
-
         // Store pointer to encoded arguments into the allocated storage.
         for (const auto &pair : llvm::enumerate(encoded)) {
-            CustomCallArgEncoding::Encoded encoded = pair.value();
-            int64_t offset = 2 + pair.index();
-            insert_value(AsPtr(b, encoded.value), offset);
-            arguments.values.push_back(encoded.value);
+            int64_t offset = pair.index();
+            insert_value(pair.value(), offset);
         }
-
[…]
