### 2023-11-23 f8d6fef
Encode res
