### 2023-11-23 e4c5b29
Add test qnode validation.
