### 2023-11-23 b379ba6
Style test_decomposition.
