+++ mlir/lib/Catalyst/Transforms/catalyst_to_llvm.cpp
@@ -233,129 +233,6 @@ struct PrintOpPattern : public OpConversionPattern<PrintOp> {
     }
 };
 
-struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
-    using OpConversionPattern::OpConversionPattern;
-
-    LogicalResult matchAndRewrite(CustomCallOp op, CustomCallOpAdaptor adaptor,
-                                  ConversionPatternRewriter &rewriter) const override
-    {
-        MLIRContext *ctx = op.getContext();
-        Location loc = op.getLoc();
-        // Create function
-        auto ptr = LLVM::LLVMPointerType::get(ctx);
-        auto type = FunctionType::get(ctx, {/*args=*/ptr, /*rets=*/ptr}, {});
-
-        auto customCallFnOp = rewriter.create<func::FuncOp>(loc, op.getCallTargetName(), type);
-        customCallFnOp.setPrivate();
-
-        // Setup args and res
-        int32_t numberArg = op.getNumberOriginalArgAttr()[0];
-        SmallVector<Value> operands = op.getOperands();
-        SmallVector<Value> args = {operands.begin(), operands.end() - numberArg};
-        SmallVector<Value> res = {operands.begin() + numberArg, operands.end()};
-
-        SmallVector<Value> operandsConverted = adaptor.getOperands();
-        SmallVector<Value> argsConverted = {operandsConverted.begin(),
-                                            operandsConverted.end() - numberArg};
-        SmallVector<Value> resConverted = {operandsConverted.begin() + numberArg,
-                                           operandsConverted.end()};
-
-        // Encode args
-        Value c1 = rewriter.create<LLVM::ConstantOp>(loc, rewriter.getI64IntegerAttr(1));
-        SmallVector<mlir::LLVM::AllocaOp> encoded;
-        for (auto tuple : llvm::drop_begin(llvm::zip(args, argsConverted))) {
-            auto memref_type = std::get<0>(tuple).getType().cast<MemRefType>();
-            auto encoded_arg = EncodeMemRef(loc, rewriter, memref_type, std::get<1>(tuple));
-            LLVM::AllocaOp alloca = rewriter.create<LLVM::AllocaOp>(loc, ptr, type, c1, 0);
-            // Start the lifetime of encoded value.
-            rewriter.create<LLVM::LifetimeStartOp>(loc, rewriter.getI64IntegerAttr(-1), alloca);
-            // Use volatile store to suppress expensive LLVM optimizations.
-            rewriter.create<LLVM::StoreOp>(loc, encoded_arg, alloca, /*alignment=*/0,
-                                           /*isVolatile=*/true);
-            encoded.push_back(alloca);
-        }
-
-        // We store encoded arguments as `!llvm.array<ptr x len>`.
-        size_t len = encoded.size();
-        Type ptr = LLVM::LLVMPointerType::get(rewriter.getContext());
-        Type type = LLVM::LLVMArrayType::get(ptr, len);
-
-        // Prepare an array for encoded arguments.
-        Value arr = rewriter.create<LLVM::UndefOp>(loc, type);
-        auto insert_value = [&](Value value, int64_t offset) {
-            arr = rewriter.create<LLVM::InsertValueOp>(loc, arr, value, offset);
-        };
-        // Store pointer to encoded arguments into the allocated storage.
-        for (const auto &pair : llvm::enumerate(encoded)) {
-            int64_t offset = pair.index();
-            insert_value(pair.value(), offset);
-        }
-        // Get allocation for packed arguments pointers.
[…]
