### 2023-11-24 f6432d7
Outline GenerateSamples.
