### 2023-11-24 a2e6e02
Add support for mcmc and other kwargs.
