+++ frontend/test/lit/test_jax_dynamic_api.py
@@ -117,3 +117,19 @@ def test_qjit_aot(a: ShapedArray([1, 3, 1], dtype=float)):
 
 
 print_mlir(test_qjit_aot, aot=True)
+
+
+# CHECK-LABEL: test_qnode_dynamic_result_complex
+@qjit
+@qml.qnode(qml.device("lightning.qubit", wires=1))
+def test_qnode_dynamic_indexing(a):
+    """Test getting a dynamic result from qnode"""
+
+    r = jnp.ones((a + 1,), dtype=complex)
+
+    return r[0]
+
+
+print_mlir(test_qnode_dynamic_indexing, 3)
+
+
