+++ frontend/catalyst/utils/jax_extras.py
@@ -21,7 +21,7 @@ from typing import Any, Callable, Dict, Generator, List, Optional, Sequence, Set
 import jax
 from jax import ShapeDtypeStruct
 from jax._src import state, util
-from jax._src.core import _update_thread_local_jit_state
+from jax._src.core import (_update_thread_local_jit_state, DBIdx)
 from jax._src.dispatch import jaxpr_replicas
 from jax._src.effects import ordered_effects as jax_ordered_effects
 from jax._src.interpreters.mlir import _module_name_regex
@@ -39,16 +39,22 @@ from jax._src.source_info_util import current as jax_current
 from jax._src.source_info_util import new_name_stack
 from jax._src.util import partition_list, safe_map, unzip2, unzip3, wrap_name, wraps
 from jax.api_util import flatten_fun
-from jax.core import AbstractValue, ClosedJaxpr, Jaxpr, JaxprEqn, MainTrace, OutputType
-from jax.core import Primitive as JaxprPrimitive
 from jax.core import (
+    AbstractValue, ClosedJaxpr, Jaxpr, JaxprEqn, MainTrace, OutputType,
+    Primitive,
     ShapedArray,
+    DShapedArray,
+    InDBIdx,
+    OutDBIdx,
     Trace,
     Tracer,
     concrete_aval,
     eval_jaxpr,
     gensym,
     thread_local_state,
+    get_referent,
+    new_jaxpr_eqn,
+    find_top_trace
 )
 from jax.interpreters.mlir import (
     AxisContext,
@@ -64,7 +70,7 @@ from jax.interpreters.partial_eval import (
     make_jaxpr_effects,
 )
 from jax.lax import convert_element_type
-from jax.linear_util import wrap_init
+from jax.linear_util import wrap_init, transformation
 from jax.tree_util import (
     PyTreeDef,
     tree_flatten,
@@ -86,17 +92,25 @@ __all__ = (
     "PyTreeDef",
     "PyTreeRegistry",
     "ShapedArray",
+    "DShapedArray",
     "ShapeDtypeStruct",
+    "DynshapePrimitive",
     "convert_constvars_jaxpr",
     "convert_element_type",
+    "deduce_avals",
+    "deduce_avals2",
     "eval_jaxpr",
     "initial_style_jaxprs_with_common_consts1",
     "initial_style_jaxprs_with_common_consts2",
     "_abstractify",
     "_initial_style_jaxpr",
     "_input_type_to_tracers",
+    "_input_type_to_tracers2",
+    "_extract_implicit_args",
+    "_output_type_to_tracers",
     "jaxpr_to_mlir",
     "jaxpr_remove_implicit",
+    "jaxpr_force_outvars",
     "make_jaxpr_effects",
     "make_jaxpr2",
     "new_dynamic_main2",
@@ -180,7 +194,7 @@ def stable_toposort(end_nodes: list) -> list:
     return sorted_nodes
 
 
-def sort_eqns(eqns: List[JaxprEqn], forced_order_primitives: Set[JaxprPrimitive]) -> List[JaxprEqn]:
+def sort_eqns(eqns: List[JaxprEqn], forced_order_primitives: Set[Primitive]) -> List[JaxprEqn]:
     """Topologically sort JAXRR equations in a unsorted list of equations, based on their
     input/output variables and additional criterias."""
 
@@ -279,6 +293,29 @@ def initial_style_jaxprs_with_common_consts2(jaxprs, all_consts):
     jaxprs = tuple(pad_jaxpr_constvars(i, jaxpr) for i, jaxpr in enumerate(jaxprs))
[…]

[…]
