### 2023-11-24 0a39d26
Change default parameters for mcmc and num_burnin.
