### 2023-11-24 e14f253
Add test for mcmc.
