### 2023-11-24 26d5c6f
Improve style for test_compiler.
