+++ mlir/lib/Catalyst/Transforms/catalyst_to_llvm.cpp
@@ -13,11 +13,13 @@
 // limitations under the License.
 
 #include <unordered_map>
+#include <iostream>
 
 #include "mlir/Conversion/LLVMCommon/ConversionTarget.h"
 #include "mlir/Conversion/LLVMCommon/MemRefBuilder.h"
 #include "mlir/Conversion/LLVMCommon/TypeConverter.h"
 #include "mlir/Dialect/Func/IR/FuncOps.h"
+#include "mlir/Dialect/LLVMIR/FunctionCallUtils.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/Pass/Pass.h"
@@ -457,29 +459,34 @@ struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
         // Create function
         Type ptr = LLVM::LLVMPointerType::get(ctx);
 
-        FunctionType type = FunctionType::get(ctx, {/*args=*/ptr, /*rets=*/ptr}, {});
-        auto customCallFnOp = rewriter.create<func::FuncOp>(loc, op.getCallTargetName(), type);
+        Type voidType = LLVM::LLVMVoidType::get(ctx);
+        auto type = LLVM::LLVMFunctionType::get(voidType, {/*args=*/ptr, /*rets=*/ptr});
+        auto point = rewriter.saveInsertionPoint();
+        ModuleOp mod = op->getParentOfType<ModuleOp>();
+        rewriter.setInsertionPointToStart(mod.getBody());
+        LLVM::LLVMFuncOp customCallFnOp =
+            rewriter.create<LLVM::LLVMFuncOp>(loc, op.getCallTargetName(), type);
         customCallFnOp.setPrivate();
-
+        rewriter.restoreInsertionPoint(point);
         // Setup args and res
         int32_t numberArg = op.getNumberOriginalArgAttr()[0];
         SmallVector<Value> operands = op.getOperands();
-        SmallVector<Value> args = {operands.begin(), operands.end() - numberArg};
-        SmallVector<Value> res = {operands.begin() + numberArg, operands.end()};
+        SmallVector<Value> args = {operands.begin(), operands.begin() + numberArg};
+        SmallVector<Value> res = {operands.begin() + (numberArg+1), operands.end()};
 
         SmallVector<Value> operandsConverted = adaptor.getOperands();
         SmallVector<Value> argsConverted = {operandsConverted.begin(),
-                                            operandsConverted.end() - numberArg};
-        SmallVector<Value> resConverted = {operandsConverted.begin() + numberArg,
+                                            operandsConverted.begin() + numberArg};
+        SmallVector<Value> resConverted = {operandsConverted.begin() + (numberArg+1),
                                            operandsConverted.end()};
 
         // Encode args
         Value c1 = rewriter.create<LLVM::ConstantOp>(loc, rewriter.getI64IntegerAttr(1));
-        SmallVector<mlir::LLVM::AllocaOp> encoded;
+        SmallVector<LLVM::AllocaOp> encoded;
         for (auto tuple : llvm::drop_begin(llvm::zip(args, argsConverted))) {
             auto memref_type = std::get<0>(tuple).getType().cast<MemRefType>();
             auto encoded_arg = EncodeMemRef(loc, rewriter, memref_type, std::get<1>(tuple));
-            LLVM::AllocaOp alloca = rewriter.create<LLVM::AllocaOp>(loc, ptr, type, c1, 0);
+            LLVM::AllocaOp alloca = rewriter.create<LLVM::AllocaOp>(loc, ptr, encoded_arg.getType(), c1, 0);
             // Start the lifetime of encoded value.
[…]
