### 2023-11-27 5b8c4ef
Update frontend/catalyst/compiler.py
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
