### 2023-11-27 d138439
Update frontend/catalyst/utils/jax_extras.py
Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
