### 2023-11-27 3ad413e
Address review suggestions - remove code refs from comments
