### 2023-11-27 40dcd49
Address review suggestions: wffa -> flattened_fun
