+++ mlir/lib/Catalyst/Transforms/catalyst_to_llvm.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <unordered_map>
 #include <iostream>
+#include <unordered_map>
 
 #include "mlir/Conversion/LLVMCommon/ConversionTarget.h"
 #include "mlir/Conversion/LLVMCommon/MemRefBuilder.h"
@@ -404,49 +404,6 @@ Value EncodeMemRef(Location loc, PatternRewriter &rewriter, MemRefType memref_ty
     return memref;
 }
 
-// Convert EncodedMemRef back to llvm MemRef descriptor, e.g.,
-//   !llvm.struct<(i8, i8, ptr, array<2 x i64>)>
-//     --->>> (note that memref descriptor still uses typed LLVM pointers)
-//   !llvm.struct<(ptr<f32>, ptr<f32>, i64, array<1 x i64>, array<1 x i64>)>
-Value DecodeMemref(Location loc, PatternRewriter &rewriter, Type type, Type converted,
-                   LLVM::AllocaOp alloca)
-{
-    auto ctx = rewriter.getContext();
-    auto memref_type = cast<MemRefType>(type);
-    auto memref_desc = MemRefDescriptor::undef(rewriter, loc, converted);
-    Type ptr = LLVM::LLVMPointerType::get(ctx);
-    // Encode sizes together with strides as a single array.
-    int64_t sizes_and_strides_size = 2 * memref_type.getRank();
-
-    // Encoded memref type: !llvm.struct<(i8, i8, ptr<i8>, array<... x i64>)>.
-    Type i8 = rewriter.getI8Type();
-    Type arr = LLVM::LLVMArrayType::get(rewriter.getI64Type(), sizes_and_strides_size);
-    auto encoded = LLVM::LLVMStructType::getLiteral(ctx, {i8, i8, ptr, arr});
-
-    Value c0 = rewriter.create<LLVM::ConstantOp>(loc, rewriter.getI64IntegerAttr(0));
-    Value c2 = rewriter.create<LLVM::ConstantOp>(loc, rewriter.getI64IntegerAttr(2));
-
-    // Fill memref descriptor pointers and offset.
-    Value gep = rewriter.create<LLVM::GEPOp>(loc, ptr, encoded, alloca, ValueRange({c0, c2}));
-    Value data_ptr = rewriter.create<LLVM::LoadOp>(loc, ptr, gep);
-    memref_desc.setAllocatedPtr(rewriter, loc, data_ptr);
-    memref_desc.setAlignedPtr(rewriter, loc, data_ptr);
-
-    // Get the statically known strides and offset from the memref type.
-    SmallVector<int64_t> strides;
-    int64_t memref_offset;
-    getStridesAndOffset(memref_type, strides, memref_offset);
-    memref_desc.setConstantOffset(rewriter, loc, memref_offset);
-
-    // Fill memref descriptor dimensions and strides.
-    for (unsigned i = 0; i < memref_type.getRank(); ++i) {
-        memref_desc.setConstantSize(rewriter, loc, i, memref_type.getDimSize(i));
-        memref_desc.setConstantStride(rewriter, loc, i, strides[i]);
-    }
-
-    auto casted = rewriter.create<UnrealizedConversionCastOp>(loc, memref_type, Value(memref_desc));
-    return casted.getResult(0);
-}
 
 struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
     using OpConversionPattern::OpConversionPattern;
@@ -472,12 +429,12 @@ struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
         int32_t numberArg = op.getNumberOriginalArgAttr()[0];
[…]
