+++ frontend/catalyst/utils/lapack_wrapper.c
@@ -1,0 +1,1 @@
+void initlapack_wrapper(void) {}
