### 2023-11-27 266cf33
Update the custom_devices guideline
