### 2023-11-27 57ec460
Exclude codecov for Darwin as sys_platform
