### 2023-11-27 9042048
Address review suggestions: simple operation tests
