### 2023-11-27 80585f6
Update wheels recipes and fix the lib path for INSTALLED=True
