+++ frontend/catalyst/utils/runtime.py
@@ -47,7 +47,7 @@ SUPPORTED_RT_DEVICES = {
 def get_lib_path(project, env_var):
     """Get the library path."""
     if INSTALLED:
-        return os.path.join(package_root, "lib")  # pragma: no cover
+        return os.path.join(package_root, "..", "lib")  # pragma: no cover
     return os.getenv(env_var, DEFAULT_LIB_PATHS.get(project, ""))
 
 
