+++ .github/workflows/build-wheel-macos-arm64.yaml
@@ -250,7 +250,7 @@ jobs:
               -DENABLE_OPENQASM=ON \
               -ENABLE_OPENMP=OFF
 
-        cmake --build runtime-build --target rt_capi
+        cmake --build runtime-build --target rt_capi rtd_lightning rtd_openqasm
 
     - name: Test Catalyst-Runtime
       run: |
