+++ .github/workflows/build-wheel-macos-x86_64.yaml
@@ -253,7 +253,7 @@ jobs:
               -DENABLE_OPENQASM=ON \
               -DENABLE_OPENMP=OFF
 
-        cmake --build runtime-build --target rt_capi
+        cmake --build runtime-build --target rt_capi rtd_lightning rtd_openqasm
 
     - name: Test Catalyst-Runtime
       run: |
