### 2023-11-27 b89b3e3
Address review suggestions: working with shapes
