### 2023-11-27 351a12c
Address review suggestions: test array operations
