### 2023-11-28 a7f8227
Update the name of rtd libraries on macOS wheels
