### 2023-11-28 bf9626a
Address review suggestions: add xfail-marked tests
