### 2023-11-28 f04657b
Address codefactor issues
