### 2023-11-28 da14204
Apply the cleanup ded vars Jax workaround
