### 2023-11-28 1632c2d
Address review suggestions: Mark array indexing tests with xfail
