### 2023-11-28 7836e32
Remove unused code.
