+++ frontend/catalyst/utils/jax_extras.py
@@ -22,7 +22,7 @@ from copy import copy
 import jax
 from jax import ShapeDtypeStruct
 from jax._src import state, util
-from jax._src.core import (_update_thread_local_jit_state, DBIdx)
+from jax._src.core import (_update_thread_local_jit_state, DBIdx, same_referent)
 from jax._src.dispatch import jaxpr_replicas
 from jax._src.effects import ordered_effects as jax_ordered_effects
 from jax._src.interpreters.mlir import _module_name_regex
@@ -100,7 +100,10 @@ __all__ = (
     "convert_element_type",
     "deduce_avals",
     "deduce_avals2",
-    "deduce_output_type",
+    "infer_output_type",
+    "infer_lambda_input_type",
+    "expand_args",
+    "expand_results",
     "eval_jaxpr",
     "initial_style_jaxprs_with_common_consts1",
     "initial_style_jaxprs_with_common_consts2",
@@ -109,7 +112,7 @@ __all__ = (
     "_input_type_to_tracers",
     "_input_type_to_tracers2",
     "_extract_implicit_args",
-    "_output_type_to_tracers",
+    "output_type_to_tracers",
     "jaxpr_to_mlir",
     "jaxpr_remove_implicit",
     "jaxpr_force_outvars",
@@ -572,10 +575,10 @@ def get_referent_frame(self, frame):
     return self if val is None else get_referent_frame(val)
 
 
-def _output_type_to_tracers(out_type: OutputType,
-                            in_tracers: List[DynamicJaxprTracer],
-                            maker: Callable[[AbstractValue],DynamicJaxprTracer]
-                            ) -> List[DynamicJaxprTracer]:
+def output_type_to_tracers(out_type: OutputType,
+                           in_tracers: List[DynamicJaxprTracer],
+                           maker: Callable[[AbstractValue],DynamicJaxprTracer]
+                           ) -> List[DynamicJaxprTracer]:
     """ Creates an expanded list of tracers representing an output values of a Jaxpr program
     based on the ``out_type`` of this program. The resulting tracers might be nested as required by
     the Jax dynamic API and might contain ``in_tracers`` of the same Jaxpr program. """
@@ -590,23 +593,23 @@ def _output_type_to_tracers(out_type: OutputType,
     return out_tracers
 
 
-T = TypeVar("T")
+TracerLike = TypeVar("TracerLike")
 
-def deduce_output_type(inputs:List[T],
-                       outputs:List[T],
-                       eq_fun:Callable[[T,T],bool],
-                       allow_indbidx:bool = True,
-                       ) -> OutputType:
+def infer_output_type(inputs:List[TracerLike],
+                      outputs:List[TracerLike],
+                      # eq_fun:Callable[[TracerLike,TracerLike],bool],
+                      allow_indbidx:bool = True,
+                      ) -> OutputType:
     """ Deduce the Jax ``out_type`` given input and ouputs abstract entities. By abstract entities
     we mean either Jax tracers or Jaxpr variables. """
 
     def _is_tracer_like(x):
         return hasattr(x, "aval")
 
-    def _safe_index(l:List[T], x:T) -> Optional[int]:
+    def _safe_index(l:List[TracerLike], x:TracerLike) -> Optional[int]:
         assert isinstance(l, list)
         for i, t in enumerate(l):
-            if eq_fun(t, x):
+            if t is x:
                 return i
         return None
 
@@ -624,7 +627,7 @@ def deduce_output_type(inputs:List[T],
                     if i is not None:
                         d2 = InDBIdx(i)
                     else:
-                        i2 = _safe_index(acc, d)
+                        i2 = _safe_index(outputs, d)
                         if i2 is not None:
                             d2 = OutDBIdx(i2)
[…]
