+++ frontend/catalyst/compiler.py
@@ -22,6 +22,8 @@ import shutil
 import subprocess
 import sys
 import warnings
+import importlib
+from os import path
 from copy import deepcopy
 from dataclasses import dataclass
 from io import TextIOWrapper
@@ -97,6 +99,7 @@ default_lib_paths = {
     "llvm": os.path.join(package_root, "../../mlir/llvm-project/build/lib"),
     "runtime": os.path.join(package_root, "../../runtime/build/lib"),
     "enzyme": os.path.join(package_root, "../../mlir/Enzyme/build/Enzyme"),
+    "pyruntime": os.path.join(package_root, "../../frontend/catalyst/utils"),
 }
 
 
@@ -222,11 +225,28 @@ class LinkerDriver:
         """
         mlir_lib_path = get_lib_path("llvm", "MLIR_LIB_DIR")
         rt_lib_path = get_lib_path("runtime", "RUNTIME_LIB_DIR")
+        py_rt_lib_path = get_lib_path("pyruntime", "PYRUNTIME_LIB_DIR")
+        package_name = 'jaxlib'
+
+        jaxlib_package = importlib.util.find_spec(package_name)
+
+        if jaxlib_package is not None:
+            lapack_directory = path.dirname(jaxlib_package.origin)
+            jaxlib_so_path = lapack_directory + "/cpu"
 
         lib_path_flags = [
             f"-Wl,-rpath,{mlir_lib_path}",
             f"-L{mlir_lib_path}",
         ]
+        lib_path_flags += [
+                f"-Wl,-rpath,{jaxlib_so_path}",
+                f"-L{jaxlib_so_path}",
+            ]
+        
+        lib_path_flags += [
+                f"-Wl,-rpath,{py_rt_lib_path}",
+                f"-L{py_rt_lib_path}",
+            ]
 
         if rt_lib_path != mlir_lib_path:
             lib_path_flags += [
@@ -241,7 +261,7 @@ class LinkerDriver:
             system_flags += ["-Wl,-no-as-needed"]
         elif platform.system() == "Darwin":  # pragma: nocover
             system_flags += ["-Wl,-arch_errors_fatal"]
-
+        print(lib_path_flags)
         default_flags = [
             "-shared",
             "-rdynamic",
@@ -250,7 +270,9 @@ class LinkerDriver:
             "-lrt_backend",
             "-lrt_capi",
             "-lpthread",
-            "-lmlir_c_runner_utils",  # required for memref.copy
+            "-lmlir_c_runner_utils",
+            "-l:custom_calls.cpython-310-x86_64-linux-gnu.so",
+            "-l:_lapack.so",
         ]
 
         return default_flags
