+++ pyproject.toml
@@ -16,7 +16,7 @@ skip = ["frontend/test/pytest/conftest.py"]
 extend_skip_glob = ["mlir/llvm-project/*", "mlir/mlir-hlo/*", ".git/*", ".vscode/*"]
 
 [build-system]
-requires = ["setuptools>=62", "wheel", "pybind11>=2.7.0", "numpy>=1.22"]
+requires = ["setuptools>=62", "wheel", "pybind11>=2.7.0", "numpy>=1.22", "jaxlib"]
 build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
