+++ pyproject.toml
@@ -25,7 +25,7 @@ extend_skip_glob = [
 ]
 
 [build-system]
-requires = ["setuptools>=62", "wheel", "pybind11>=2.7.0", "numpy>=1.22"]
+requires = ["setuptools>=62", "wheel", "pybind11>=2.7.0", "numpy>=1.22", "jaxlib"]
 build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
