+++ mlir/lib/Driver/CMakeLists.txt
@@ -30,8 +30,6 @@ set(LIBS
     quantum-transforms
     MLIRGradient
     gradient-transforms
-    MLIRMitigation
-    mitigation-transforms
     MhloRegisterDialects
     StablehloRegister
     MLIRCatalystTest
