+++ frontend/catalyst/pennylane_extensions.py
@@ -71,7 +71,8 @@ from catalyst.utils.jax_extras import (
     ShapedArray,
     _initial_style_jaxpr,
     _input_type_to_tracers,
-    _input_type_to_tracers2,
+    collapse,
+    input_type_to_tracers,
     output_type_to_tracers,
     convert_constvars_jaxpr,
     infer_output_type,
@@ -80,6 +81,7 @@ from catalyst.utils.jax_extras import (
     jaxpr_force_outvars,
     deduce_avals,
     deduce_avals2,
+    deduce_avals3,
     get_implicit_and_explicit_flat_args,
     initial_style_jaxprs_with_common_consts1,
     initial_style_jaxprs_with_common_consts2,
@@ -970,27 +972,47 @@ class WhileLoop(HybridOp):
 
     def trace_quantum(self, ctx, device, trace, qrp) -> QRegPromise:
         cond_trace = self.regions[0].trace
-        res_classical_tracers = self.regions[0].res_classical_tracers
         with EvaluationContext.frame_tracing_context(ctx, cond_trace):
+            arg_expanded_classical_tracers = expand_args(
+                self.regions[0].arg_classical_tracers
+            )
+            res_expanded_classical_tracers, out_type = expand_results(
+                arg_expanded_classical_tracers,
+                self.regions[0].res_classical_tracers,
+                force_implicit_indbidx=True
+            )
             _input_type_to_tracers(cond_trace.new_arg, [AbstractQreg()])
-            cond_jaxpr, _, cond_consts = ctx.frames[cond_trace].to_jaxpr2(res_classical_tracers)
+            cond_jaxpr, out_type2, cond_consts = ctx.frames[cond_trace].to_jaxpr2(
+                res_expanded_classical_tracers
+            )
+            assert unzip2(out_type)[0] == unzip2(out_type2)[0], f"\n{out_type=}\n{out_type2=}"
 
         body_trace = self.regions[1].trace
         body_tape = self.regions[1].quantum_tape
-        arg_classical_tracers = self.regions[1].arg_classical_tracers
-        res_classical_tracers = self.regions[1].res_classical_tracers
         with EvaluationContext.frame_tracing_context(ctx, body_trace):
             qreg_in = _input_type_to_tracers(body_trace.new_arg, [AbstractQreg()])[0]
             qrp_out = trace_quantum_tape(body_tape, device, qreg_in, ctx, body_trace)
             qreg_out = qrp_out.actualize()
-
-
-            res_tracers = expand_results(arg_classical_tracers + [qreg_in],
-                                         res_classical_tracers + [qreg_out],
-                                         force_implicit_indbidx=True)
-            body_jaxpr, out_type, body_consts = ctx.frames[body_trace].to_jaxpr2(
-                res_tracers
+            arg_expanded_tracers = expand_args(
+                self.regions[1].arg_classical_tracers + [qreg_in]
             )
+            res_expanded_tracers, out_type = expand_results(
+                arg_expanded_tracers,
+                self.regions[1].res_classical_tracers + [qreg_out],
+                force_implicit_indbidx=True
+            )
+            body_jaxpr, out_type2, body_consts = ctx.frames[body_trace].to_jaxpr2(
+                res_expanded_tracers
+            )
+            assert unzip2(out_type)[0] == unzip2(out_type2)[0], f"\n{out_type=}\n{out_type2=}"
+            print("QBIND_RES_TRACERS")
+            for t in self.regions[1].res_classical_tracers + [qreg_out]: print("-", t)
+            print("QBIND_RES_EXP_TRACERS")
+            for t in res_expanded_tracers: print("-", t)
+            print("BBBBBBBBBBBBBB")
[…]
