### 2023-11-29 ba57ebc
Apply suggestions from code review
Co-authored-by: Josh Izaac <josh146@gmail.com>
