+++ frontend/test/pytest/test_transform.py
@@ -35,7 +35,10 @@ import pytest
 from jax import numpy as jnp
 from numpy.testing import assert_allclose
 from pennylane import numpy as pnp
-from pennylane import qcut
+try:
+    from pennylane import qcut
+except:
+    from pennylane.transforms import qcut
 from pennylane.transforms import hamiltonian_expand, merge_rotations, sum_expand
 
 from catalyst import measure, qjit
