+++ frontend/test/pytest/test_transform.py
@@ -37,7 +37,7 @@ from numpy.testing import assert_allclose
 from pennylane import numpy as pnp
 try:
     from pennylane import qcut
-except:
+except:  #pylint: disable=bare-except
     from pennylane.transforms import qcut
 from pennylane.transforms import hamiltonian_expand, merge_rotations, sum_expand
 
