### 2023-11-29 e425745
Apply code review suggestions: Update docstrings, simplify the QuantumDevice constructor and extract_backend_info
