### 2023-11-29 86ca6fe
Replace pennylane.transforms.qcut with pennylane.qcut (#371)
**Context:**

Qcut has its own module in PennyLane 0.34.

**Description of the Change:**

Update the import path in the tests.

---------

Co-authored-by: erick-xanadu <110487834+erick-xanadu@users.noreply.github.com>
