### 2023-11-29 bb32710
Add test to avoid recompilation when using abstracted_axes.
