### 2023-11-29 bd46085
Adds documentation for recompilation.
