### 2023-11-29 33e32ff
Remove dead code.
