### 2023-11-29 f95e0e1
Add mcmc test.
