+++ runtime/lib/CMakeLists.txt
@@ -1,2 +1,3 @@
 add_subdirectory(capi)
 add_subdirectory(backend)
+add_subdirectory(registry)
