+++ runtime/lib/registry/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_library(registry SHARED Registry.cpp)
