### 2023-11-29 ca19af6
Add example python module.
