+++ runtime/lib/registry/CMakeLists.txt
@@ -1,1 +1,4 @@
-add_library(registry SHARED Registry.cpp)
+find_package(pybind11 REQUIRED)
+
+pybind11_add_module(pyregistry PyRegistry.cpp)
+
