### 2023-11-29 88d7bcf
Add example python module.
