+++ runtime/lib/registry/PyRegistry.cpp
@@ -1,0 +1,12 @@
+#include <pybind11/pybind11.h>
+
+namespace py = pybind11;
+
+int add(int i, int j) {
+	return i + j;
+}
+
+PYBIND11_MODULE(PyRegistry, m) {
+	m.doc() = "pybind11 example plugin"; // optional module docstring.
+	m.def("add", &add, "A function that adds two numbers");
+}
