+++ runtime/lib/registry/PyRegistry.cpp
@@ -6,7 +6,7 @@ int add(int i, int j) {
 	return i + j;
 }
 
-PYBIND11_MODULE(PyRegistry, m) {
+PYBIND11_MODULE(pyregistry, m) {
 	m.doc() = "pybind11 example plugin"; // optional module docstring.
 	m.def("add", &add, "A function that adds two numbers");
 }
