### 2023-11-29 9ea1f5d
Add build commands.
