### 2023-11-29 7f11485
More general.
