### 2023-11-29 91b7294
Simple example of registering a function.
