### 2023-11-29 d33c608
Adding support for dynamic API to while-loop
