+++ frontend/catalyst/jax_tracer.py
@@ -72,6 +72,10 @@ from catalyst.utils.jax_extras import (
     tree_structure,
     tree_unflatten,
     wrap_init,
+    deduce_avals3,
+    input_type_to_tracers,
+    unzip2,
+    convert_constvars_jaxpr
 )
 
 
@@ -350,6 +354,33 @@ def has_nested_tapes(op: Operation) -> bool:
         and len(op.regions) > 0
         and any(r.quantum_tape is not None for r in op.regions)
     )
+
+
+def trace_function(ctx, fun, *args, force_implicit_indbidx=False, **kwargs):
+
+    with EvaluationContext.frame_tracing_context(ctx) as trace:
+        wfun, in_sig, out_sig = deduce_avals3(
+            fun, args, kwargs, force_implicit_indbidx=force_implicit_indbidx
+        )
+        arg_classical_tracers = input_type_to_tracers(in_sig.in_type, trace.new_arg)
+        res_classical_tracers = [
+            trace.full_raise(t)
+            for t in wfun.call_wrapped(*arg_classical_tracers)
+        ]
+        out_aval, out_keep = unzip2(out_sig.out_type())
+        jaxpr, out_type2, consts = ctx.frames[trace].to_jaxpr2(res_classical_tracers)
+        assert len(out_type2) == len(out_sig.out_type())
+        out_aval2, _ = unzip2(out_type2)
+
+        return (
+            ClosedJaxpr(convert_constvars_jaxpr(jaxpr), ()),
+            tuple(zip(out_aval2, out_keep)),
+            out_sig.out_tree(),
+            consts
+        )
+
+
+
 
 
 def trace_to_mlir(func, abstracted_axes, *args, **kwargs):
