+++ frontend/test/pytest/test_jax_dynamic_api.py
@@ -345,15 +345,36 @@ def test_qnode_while_indbidx_outdbidx():
             return (a, b, i)
 
         a2, b2, _ = loop(a, b, 0)
-        return a2, b2
+        return a+a2, b2
 
     res_a, res_b = f(3)
-    assert_array_and_dtype_equal(res_a, jnp.ones(3))
+    assert_array_and_dtype_equal(res_a, 2*jnp.ones(3))
     assert_array_and_dtype_equal(res_b, jnp.ones(4))
 
 
 
 def test_qjit_while_1():
+    """Test that catalyst tensor primitive is compatible with quantum while"""
+
+    @qjit
+    def f(sz):
+        a = jnp.ones([sz+1], dtype=float)
+
+        @while_loop(lambda _, i: i < 3)
+        def loop(_, i):
+            b = jnp.ones([sz+1], dtype=float)
+            i += 1
+            return (b, i)
+
+        a2, _ = loop(a, 0)
+        return a2
+
+    result = f(3)
+    expected = jnp.ones(4)
+    assert_array_and_dtype_equal(result, expected)
+
+
+def test_qjit_while_2():
     """Test that catalyst tensor primitive is compatible with quantum while"""
 
     @qjit()
@@ -372,6 +393,69 @@ def test_qjit_while_1():
     result = f(3)
     expected = jnp.ones(4)
     assert_array_and_dtype_equal(result, expected)
+
+
+def test_qjit_while_shared_indbidx():
+    """Test that catalyst tensor primitive is compatible with quantum while"""
+
+    @qjit()
+    def f(sz):
+        a = jnp.ones([sz], dtype=float)
+        b = jnp.ones([sz], dtype=float)
+
+        @while_loop(lambda _a, _b, i: i < 3)
+        def loop(a, b, i):
+            i += 1
+            return (a, b, i)
+
+        a2, b2, _ = loop(a, b, 0)
+        return a2 + b2
+
+    result = f(3)
+    expected = 2*jnp.ones(3)
+    assert_array_and_dtype_equal(result, expected)
+
+
+def test_qjit_while_indbidx_outdbidx():
+    """Test that catalyst tensor primitive is compatible with quantum while"""
+
+    @qjit()
+    def f(sz):
+        a0 = jnp.ones([sz], dtype=float)
+        b0 = jnp.ones([sz+1], dtype=float)
+
+        @while_loop(lambda _a, _b, i: i < 3)
+        def loop(a, _, i):
+            b = jnp.ones([sz+1], dtype=float)
+            i += 1
+            return (a, b, i)
+
+        a2, b2, _ = loop(a0, b0, 0)
+        return a0+a2, b2
+
+    res_a, res_b = f(3)
+    assert_array_and_dtype_equal(res_a, 2*jnp.ones(3))
+    assert_array_and_dtype_equal(res_b, jnp.ones(4))
+
+
+def test_qjit_while_outer():
+    """Test that catalyst tensor primitive is compatible with quantum while"""
+
+    @qjit
+    def f(sz):
+        a0 = jnp.ones([sz], dtype=float)
+
+        @while_loop(lambda _a, i: i < 3)
+        def loop(a, i):
+            i += 1
+            return (a0, i)
+
+        a2, _ = loop(a0, 0)
+        return a0+a2
+
+    res_a = f(3)
+    print(f.jaxpr)
+    assert_array_and_dtype_equal(res_a, 2*jnp.ones(3))
 
 
 @pytest.mark.skip("Dynamic arrays support in quantum control flow is not implemented")
