### 2023-11-29 6db724c
Update custom_devices guideline
