### 2023-11-29 21e62ab
[Frontend][Runtime] A plugin system for the backend devices (#343)
**Description of the Change:**
- [X] Update the device info communication API. We pass a tuple of
(`rtd_name`, `rtd_lib`, `rtd_kwargs`) representing the device name, the
shared library path, and other specifications (number of shots, S3, ARN,
etc.) from the frontend to the backend.
- [X] Better encapsulating the runtime utility functions in the frontend
by adding `runtime.py`.
- [X] Re-structure the runtime to fully split the backend devices build
system from the runtime CAPI and execution context manager.
- [X] Update the init
[…]
