### 2023-11-30 bea496f
Adding support for dynamic API to while-loop
