+++ .github/workflows/build-wheel-macos-arm64.yaml
@@ -297,7 +297,7 @@ jobs:
         retention-days: 14
 
   test-wheels:
-    needs: [constants, catalyst-macos-wheels-x86-64]
+    needs: [constants, catalyst-macos-wheels-arm64]
     strategy:
       fail-fast: false
       matrix:
