### 2023-11-30 d34f584
Restore device_type
