### 2023-11-30 dc25d3b
Remove parameter from constructor.
