### 2023-11-30 6e293da
Adding support for dynamic API to for-loop
