+++ frontend/catalyst/jax_tracer.py
@@ -277,13 +277,10 @@ class HybridOp(Operation):
     num_wires = AnyWires
     binder: Callable = _no_binder
 
-    def __init__(self, in_classical_tracers, out_classical_tracers, regions: List[HybridOpRegion],
-                 in_type = None, out_type = None):
+    def __init__(self, in_classical_tracers, out_classical_tracers, regions: List[HybridOpRegion]):
         self.in_classical_tracers = in_classical_tracers
         self.out_classical_tracers = out_classical_tracers
         self.regions: List[HybridOpRegion] = regions
-        self.in_type = in_type
-        self.out_type = out_type
         super().__init__(wires=Wires(HybridOp.num_wires))
 
     def __repr__(self):
