+++ .github/workflows/build-wheel-macos-arm64.yaml
@@ -248,7 +248,7 @@ jobs:
               -DKokkos_ENABLE_COMPLEX_ALIGN=OFF \
               -DENABLE_WARNINGS=OFF \
               -DENABLE_OPENQASM=ON \
-              -ENABLE_OPENMP=OFF
+              -DENABLE_OPENMP=OFF
 
         cmake --build runtime-build --target rt_capi rtd_lightning rtd_openqasm
 
