### 2023-11-30 c972eb7
[CI] Fix the Wheel macOS (arm64) script with name of the job dependency (#372)
The recipe has been failing after merging PR #363. This PR should
address the issue.

Wheels are built and tested on macOS
https://github.com/PennyLaneAI/catalyst/actions/runs/7049734894
