### 2023-11-30 970bca6
Lower callback to CustomCallOp.
