### 2023-12-01 06233e7
Apply suggestions from code review
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
