### 2023-12-01 345620c
Proof of concept.
