+++ mlir/lib/Catalyst/Transforms/catalyst_to_llvm.cpp
@@ -307,9 +307,9 @@ struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
     LogicalResult matchAndRewrite(CustomCallOp op, CustomCallOpAdaptor adaptor,
                                   ConversionPatternRewriter &rewriter) const override
     {
-        MLIRContext *ctx = op.getContext();
         Location loc = op.getLoc();
         // Create function
+        MLIRContext *ctx = op.getContext();
         Type ptr = LLVM::LLVMPointerType::get(ctx);
 
         Type voidType = LLVM::LLVMVoidType::get(ctx);
@@ -317,15 +317,29 @@ struct CustomCallOpPattern : public OpConversionPattern<CustomCallOp> {
         ModuleOp mod = op->getParentOfType<ModuleOp>();
         rewriter.setInsertionPointToStart(mod.getBody());
 
+        if (op.getCallTargetName().equals("pyregistry")) {
+            SmallVector<Type> argTypes(adaptor.getOperands().getTypes().begin(),
+                                       adaptor.getOperands().getTypes().end());
+            auto type2 = LLVM::LLVMFunctionType::get(ptr, argTypes);
+            LLVM::LLVMFuncOp customCallFnOp =
+                rewriter.create<LLVM::LLVMFuncOp>(loc, op.getCallTargetName(), type2);
+            LLVM::LLVMFuncOp customCallFnOp = mlir::LLVM::lookupOrCreateFn(
+                mod, op.getCallTargetName(), {/*args=*/ptr, /*rets=*/ptr}, /*ret_type=*/voidType);
+            rewriter.restoreInsertionPoint(point);
+            rewriter.replaceOpWithNewOp<LLVM::CallOp>(op, customCallFnOp, adaptor.getOperands());
+            return success();
+        }
         LLVM::LLVMFuncOp customCallFnOp = mlir::LLVM::lookupOrCreateFn(
             mod, op.getCallTargetName(), {/*args=*/ptr, /*rets=*/ptr}, /*ret_type=*/voidType);
-
         customCallFnOp.setPrivate();
         rewriter.restoreInsertionPoint(point);
 
         // Setup args and res
         int32_t numberArg = op.getNumberOriginalArgAttr()[0];
         SmallVector<Value> operands = op.getOperands();
+        // What if it h as no arguments and no results?
+        size_t how_many_args_and_results = operands.size();
+        bool has_results = (how_many_args_and_results - numberArg) > 0;
         SmallVector<Value> args = {operands.begin(), operands.begin() + numberArg};
         SmallVector<Value> res = {operands.begin() + numberArg, operands.end()};
 
