### 2023-12-01 667e6e8
Proof of concept.
