+++ runtime/lib/capi/CMakeLists.txt
@@ -73,7 +73,7 @@ set_property(TARGET catalyst_qir_qis_obj PROPERTY POSITION_INDEPENDENT_CODE ON)
 
 add_library(rt_capi SHARED)
 
-target_link_libraries(rt_capi ${CMAKE_DL_LIBS} catalyst_qir_qis_obj)
+target_link_libraries(rt_capi ${CMAKE_DL_LIBS} catalyst_qir_qis_obj registry)
 
 target_include_directories(rt_capi PUBLIC .
     ${CMAKE_CURRENT_SOURCE_DIR}
