+++ frontend/test/pytest/test_jax_dynamic_api.py
@@ -371,6 +371,7 @@ def test_accessing_shapes():
 
 def test_no_recompilation():
     """Test that the function is not recompiled when changing the argument shape across invocations."""
+
     @qjit(abstracted_axes={0: "n"})
     def i(x):
         return x
