### 2023-12-01 d3df4de
Address review suggestions: comment for the shape-conversion test
