### 2023-12-01 e83553d
Black.
