+++ mlir/include/Catalyst/Transforms/Patterns.h
@@ -23,4 +23,7 @@ void populateBufferizationPatterns(mlir::TypeConverter &, mlir::RewritePatternSe
 
 void populateScatterPatterns(mlir::RewritePatternSet &);
 
+void populateQnodeToAsyncPatterns(mlir::RewritePatternSet &);
+
+
 } // namespace catalyst
