+++ mlir/lib/Catalyst/Transforms/RegisterAllPasses.cpp
@@ -31,5 +31,6 @@ void catalyst::registerAllCatalystPasses()
     mlir::registerPass(catalyst::createEmitCatalystPyInterfacePass);
     mlir::registerPass(catalyst::createCopyGlobalMemRefPass);
     mlir::registerPass(catalyst::createCatalystConversionPass);
+    mlir::registerPass(catalyst::createQnodeToAsyncLoweringPass);
     mlir::registerPass(catalyst::createTestPass);
 }
