+++ mlir/lib/Catalyst/Transforms/CMakeLists.txt
@@ -4,6 +4,8 @@ file(GLOB SRC
     ArrayListToMemRefPass.cpp
     scatter_lowering.cpp
     ScatterPatterns.cpp
+    qnode_to_async_lowering.cpp
+    QnodeToAsyncPatterns.cpp
     RegisterAllPasses.cpp
     BufferizationPatterns.cpp
     catalyst_bufferize.cpp
