### 2023-12-01 344935f
Allow gradient functions to be called outside qjit
