+++ doc/changelog.md
@@ -42,6 +42,12 @@
 
 <h3>Improvements</h3>
 
+* Catalyst gradient functions ``grad``, ``jacobian``, ``jvp``, and ``vjp`` can now be invoked from
+  outside a ``@qjit`` context. This simplifies the process of writing functions where compilation
+  can be turned on and off easily by adding or removing the decorator. The functions dispatch to
+  their JAX equivalents when the compilation is turned off.
+  [(#375)](https://github.com/PennyLaneAI/catalyst/pull/375)
+
 * ``AllocOp``, ``DeallocOp`` have now (only) value semantics. In the frontend, the last
   quantum register is deallocated instead of the first one. This allows to return the quantum
   register in functions and can be given to another function (useful for quantum transformation).
