### 2023-12-02 77b1659
Adding support for dynamic API to for-loop; Debug prints
