+++ frontend/catalyst/jax_primitives.py
@@ -1229,6 +1229,7 @@ def _while_loop_lowering(
     cond_nconsts: int,
     body_nconsts: int,
 ):
+    from pdb import set_trace; set_trace()
     loop_carry_types_plus_consts = [mlir.aval_to_ir_types(a)[0] for a in jax_ctx.avals_in]
     flat_args_plus_consts = mlir.flatten_lowering_ir_args(iter_args_plus_consts)
     assert [val.type for val in flat_args_plus_consts] == loop_carry_types_plus_consts
@@ -1292,21 +1293,25 @@ def _while_loop_lowering(
 # for loop
 #
 @for_p.def_abstract_eval
-def _for_loop_abstract_eval(*args, body_jaxpr, body_iter_index, **kwargs):
+def _for_loop_abstract_eval(*args, body_jaxpr,
+                            nimplicit = 0,
+                            **kwargs):
     _assert_jaxpr_without_constants(body_jaxpr)
     out_type = infer_output_type(body_jaxpr.jaxpr.invars,
                                  body_jaxpr.jaxpr.outvars,
                                  force_implicit_indbidx=False)
 
-    out_type2 = out_type_force_outdbidx(out_type, body_iter_index, body_jaxpr.jaxpr.invars, body_jaxpr.jaxpr.outvars)
-    out_type3 = out_type_shift_indbidx(out_type2, 3)
-    return out_type3
+    body_loop_arg_idx = nimplicit
+    out_type = out_type_force_outdbidx(out_type, body_loop_arg_idx, [], body_jaxpr.jaxpr.invars, body_jaxpr.jaxpr.outvars)
+    return out_type
 
 
 # pylint: disable=too-many-arguments
 @for_p.def_impl
 def _for_loop_def_impl(
-    ctx, lower_bound, upper_bound, step, *iter_args_plus_consts, body_jaxpr, body_iter_index, body_nconsts
+    ctx, lower_bound, upper_bound, step, *iter_args_plus_consts, body_jaxpr,
+    nimplicit = 0,
+    body_nconsts
 ):  # pragma: no cover
     raise NotImplementedError()
 
@@ -1314,51 +1319,67 @@ def _for_loop_def_impl(
 # pylint: disable=too-many-statements, too-many-arguments
 def _for_loop_lowering(
     jax_ctx: mlir.LoweringRuleContext,
-    lower_bound: ir.Value,
-    upper_bound: ir.Value,
-    step: ir.Value,
+    # lower_bound: ir.Value,
+    # upper_bound: ir.Value,
+    # step: ir.Value,
     *iter_args_plus_consts: tuple,
     body_jaxpr: core.ClosedJaxpr,
-    body_iter_index : int,
     body_nconsts: int,
     apply_reverse_transform: bool,
+    nimplicit : int = 0
 ):
+    # print("IIIIIIIIIII", [v.type for v in [lower_bound, upper_bound, step, *iter_args_plus_consts]])
     # Separate constants from iteration arguments.
     # The MLIR value provided by JAX for the iteration index is not needed
     # (as it's identical to the lower bound value).
-    body_consts = iter_args_plus_consts[:body_nconsts]
-    loop_index = iter_args_plus_consts[body_nconsts]
-    loop_args = iter_args_plus_consts[body_nconsts + 1 :]
 
-    all_param_types_plus_consts = [mlir.aval_to_ir_types(a)[0] for a in jax_ctx.avals_in]
+    body_consts    = iter_args_plus_consts[:body_nconsts]
+    body_implicits = iter_args_plus_consts[body_nconsts : body_nconsts + nimplicit]
+    lower_bound    = iter_args_plus_consts[body_nconsts + nimplicit + 0]
[…]
