+++ frontend/test/pytest/test_loops.py
@@ -529,7 +529,25 @@ class TestInterpretationControlFlow:
         mulc = qjit(muli)
         assert mulc(1, 2) == muli(1, 2)
 
-    def test_for_loop(self):
+
+    def test_for_loop_identity(self, backend):
+        """Test simple for loop."""
+
+        @qjit
+        @qml.qnode(qml.device(backend, wires=1))
+        def fun(x):
+            @for_loop(0, 10, 1)
+            def loop(_, agg):
+                return agg
+
+            res = loop(x)
+            return res
+
+        assert fun(0) == 0
+        assert False
+
+
+    def test_for_loop_acc(self):
         """Test simple for loop."""
 
         def muli(x: int, n: int):
