+++ frontend/catalyst/pennylane_extensions.py
@@ -890,10 +890,9 @@ class ForLoop(HybridOp):
 
     binder = for_p.bind
 
-    def __init__(self, *args, apply_reverse_transform:bool, body_iter_index:int, **kwargs):
+    def __init__(self, *args, apply_reverse_transform:bool, **kwargs):
         super().__init__(*args, **kwargs)
         self.apply_reverse_transform = apply_reverse_transform
-        self.body_iter_index = body_iter_index
 
 
     def trace_quantum(self, ctx, device, trace, qrp) -> QRegPromise:
@@ -907,7 +906,8 @@ class ForLoop(HybridOp):
             qreg_out = qrp_out.actualize()
 
             arg_expanded_tracers = expand_args(
-                self.regions[0].arg_classical_tracers + [qreg_in]
+                self.regions[0].arg_classical_tracers + [qreg_in],
+                force_implicit_indbidx=True
             )[0]
             res_expanded_tracers, _ = expand_results(
                 arg_expanded_tracers,
@@ -921,40 +921,44 @@ class ForLoop(HybridOp):
         print(consts)
         print("JJJJJJJJJJJJ")
 
-        in_expanded_tracers = (
+        in_expanded_tracers = expand_args(op.in_classical_tracers,
+                                          force_implicit_indbidx=True)[0]
+
+        nimplicit = len(in_expanded_tracers) - len(op.in_classical_tracers)
+
+        in_c_expanded_tracers = (
             [trace.full_raise(c) for c in consts] +
-            op.in_classical_tracers[0:3] +
-            expand_args(op.in_classical_tracers[3:])[0] +
+            in_expanded_tracers +
             [qrp.actualize()]
         )
 
-        print("BIND_IN_EXPANDED_TRACERS")
-        for t in in_expanded_tracers: print("I", t)
+        print("IN_C_EXPANDED_TRACERS")
+        for t in in_c_expanded_tracers: print("I", t)
 
         print("BIND_OUT_CLASSICAL_RESULTS")
         for t in self.out_classical_tracers: print("O", t)
 
-        out_expanded_classical_tracers = (
+        out_c_expanded_classical_tracers = (
             expand_results(
-                in_expanded_tracers,
+                in_c_expanded_tracers,
                 self.out_classical_tracers,
                 force_implicit_indbidx=True,
             )[0]
         )
 
-        print("BIND_OUT_EXPANDED_CLASSICAL_TRACERS")
-        for t in out_expanded_classical_tracers: print("O", t)
+        print("OUT_C_EXPANDED_CLASSICAL_TRACERS")
+        for t in out_c_expanded_classical_tracers: print("O", t)
 
         qrp2 = QRegPromise(
             op.bind_overwrite_classical_tracers2(
                 ctx,
                 trace,
-                in_expanded_tracers=in_expanded_tracers,
-                out_expanded_tracers=out_expanded_classical_tracers,
+                in_expanded_tracers=in_c_expanded_tracers,
+                out_expanded_tracers=out_c_expanded_classical_tracers,
                 body_jaxpr=ClosedJaxpr(convert_constvars_jaxpr(jaxpr), ()),
                 body_nconsts=len(consts),
                 apply_reverse_transform=self.apply_reverse_transform,
-                body_iter_index=self.body_iter_index,
+                nimplicit=nimplicit,
             )
         )
         return qrp2
@@ -1558,11 +1562,6 @@ def for_loop(lower_bound, upper_bound, step):
                     force_implicit_indbidx=True
                 )
                 in_type = in_sig.in_type
-                in_expanded_classical_tracers = (
-                    aux_classical_tracers +
[…]

[…]
