### 2023-12-04 d2f2e8d
Update installation.rst (#376)
Fix broken rst link
