### 2023-12-04 efc6441
Apply suggestions from code review
Co-authored-by: Ali Asadi <ali@xanadu.ai>
