### 2023-12-04 143fb01
Update doc/dev/custom_devices.rst
