+++ runtime/lib/backend/openqasm/braket.aws.qubit.toml
@@ -21,33 +21,18 @@ full = [
 	"Hadamard",
 	"S",
 	"T",
-	"PhaseShift",
-	"RX",
-	"RY",
-	"RZ",
-	"Rot",
 	"CNOT",
 	"CY",
 	"CZ",
 	"SWAP",
-	"IsingXX",
-	"IsingXY",
-	"IsingYY",
-	"IsingZZ",
-	"ControlledPhaseShift",
-	"CRX",
-	"CRY",
-	"CRZ",
-	"CRot",
-	"SingleExcitation",
-	"SingleExcitationMinus",
-	"SingleExcitationPlus",
-	"Toffoli",
+	"PhaseShift",
+	"RX",
+	"RY",
+	"RZ",
 	"CSWAP",
-	"DoubleExcitation",
-	"DoubleExcitationMinus",
-	"DoubleExcitationPlus",
-	"MultiRZ",
+	"PSWAP",
+	"ISWAP",
+	"Toffoli",
 ]
 matrix = [
 	# QubitUnitary should never be here.
