+++ runtime/tests/Test_DevicePool.cpp
@@ -1,0 +1,234 @@
+
+// Copyright 2023 Xanadu Quantum Technologies Inc.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+//     http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include <future>
+#include <thread>
+
+#include "ExecutionContext.hpp"
+#include "QuantumDevice.hpp"
+#include "QuantumDeviceInterface.hpp"
+#include "RuntimeCAPI.h"
+
+#include "TestUtils.hpp"
+
+using namespace Catalyst::Runtime;
+
+inline auto get_dylib_ext() -> std::string
+{
+#ifdef __linux__
+    return ".so";
+#elif defined(__APPLE__)
+    return ".dylib";
+#endif
+}
+
+TEST_CASE("Test getter/setter methods in RTDeviceInfoT", "[device_pool]")
+{
+    std::unique_ptr<RTDeviceInfoT> device = std::make_unique<RTDeviceInfoT>(
+        "librtd_lightning" + get_dylib_ext(), "LightningSimulator", "{shots: 100}");
+    std::cerr << *device << std::endl;
+    auto &&[other_lib, other_name, other_kwargs] = device->getDeviceInfo();
+    CHECK(*device == RTDeviceInfoT(other_lib, other_name, other_kwargs));
+    CHECK(device->getQuantumDevicePtr() != nullptr);
+    CHECK(device->getQuantumDevicePtr() != nullptr);
+    device->setTapeRecorderStatus(true);
+    CHECK(device->getTapeRecorderStatus());
+
+    std::unique_ptr<RTDeviceInfoT> dev_oq3 = std::make_unique<RTDeviceInfoT>("braket.aws.qubit");
+    CHECK(*dev_oq3 == RTDeviceInfoT("librtd_openqasm" + get_dylib_ext(), "OpenQasmDevice"));
+}
+
+TEST_CASE("Test the device pool in ExecutionContext (single-thread)", "device_pool")
+{
+    std::unique_ptr<ExecutionContext> driver = std::make_unique<ExecutionContext>();
+
+    auto dev0 = driver->addDevice("lightning.qubit");
+    auto dev1 = driver->addDevice("lightning.qubit");
+    auto dev2 = driver->addDevice("lightning.qubit");
+    auto dev3 = driver->addDevice("lightning.qubit");
+
+    CHECK(dev0.first == 0);
+    CHECK(dev1.first == 1);
+    CHECK(dev2.first == 2);
+    CHECK(dev3.first == 3);
+
+    CHECK(driver->getDevice(dev0.first) != nullptr);
+    CHECK(driver->getDevice(dev3.first) != nullptr);
+    CHECK(driver->getDevice(50) == nullptr);
+
+    driver->removeDevice(dev1.first);
+    driver->removeDevice(dev3.first);
+
+    CHECK(driver->getDevice(dev1.first) == nullptr);
+    CHECK(driver->getDevice(dev3.first) == nullptr);
+}
+
+void addDeviceToDriver(std::unique_ptr<ExecutionContext> &driver, std::string name)
+{
+    auto result = driver->addDevice(std::move(name));
+    CHECK(result.first < driver->getPoolSize());
+}
+
[…]
