+++ runtime/lib/capi/ExecutionContext.hpp
@@ -255,8 +255,8 @@ class RTDeviceInfoT {
         void *f_ptr = rtd_dylib->getSymbol(factory_name);
         QuantumDevice *impl =
             f_ptr ? reinterpret_cast<decltype(GenericDeviceFactory) *>(f_ptr)(rtd_kwargs) : nullptr;
-        rtd_qdevice.reset(impl);
 
+        rtd_qdevice.reset(impl);
         return rtd_qdevice;
     }
 
@@ -268,6 +268,8 @@ class RTDeviceInfoT {
     {
         return {rtd_lib, rtd_name, rtd_kwargs};
     }
+
+    [[nodiscard]] auto getDeviceName() const -> const std::string & { return rtd_name; }
 
     friend std::ostream &operator<<(std::ostream &os, const RTDeviceInfoT &device)
     {
@@ -285,26 +287,16 @@ class ExecutionContext final {
     size_t pool_counter{0}; // Counter for generating unique keys for devices
     std::mutex pool_mutex;  // Mutex to guard the device pool
 
-    // Device specifications
+    // Device info before initialization to support the current frontend-backend pipeline
+    // TODO: remove this after the end-to-end adaptation of the async support in Catalyst
     std::string zero_rtd_kwargs{};
     std::string zero_rtd_name{};
     bool zero_rtd_tape_recorder_status;
 
     // ExecutionContext pointers
-    std::unique_ptr<QuantumDevice> active_rtd_ptr{nullptr};
+    std::shared_ptr<QuantumDevice> active_rtd_ptr{nullptr};
     std::unique_ptr<MemoryManager> memory_man_ptr{nullptr};
-    std::unique_ptr<SharedLibraryManager> init_rtd_dylib{nullptr};
     std::unique_ptr<PythonInterpreterGuard> py_guard{nullptr};
-
-    // Helper methods
-    void update_device_pl2runtime(std::string &pl_name, const std::string &rtd_name)
-    {
-#ifdef __linux__
-        pl_name = "librtd_" + rtd_name + ".so";
-#elif defined(__APPLE__)
-        pl_name = "librtd_" + rtd_name + ".dylib";
-#endif
-    }
 
   public:
     explicit ExecutionContext() : zero_rtd_tape_recorder_status(false)
@@ -314,12 +306,9 @@ class ExecutionContext final {
 
     ~ExecutionContext()
     {
-        active_rtd_ptr.reset(nullptr);
         memory_man_ptr.reset(nullptr);
-        init_rtd_dylib.reset(nullptr);
         py_guard.reset(nullptr);
 
-        RT_ASSERT(getDevice() == nullptr);
         RT_ASSERT(getMemoryManager() == nullptr);
     };
 
@@ -329,17 +318,8 @@ class ExecutionContext final {
 
     void setDeviceName(std::string_view name) noexcept { zero_rtd_name = name; }
 
-    [[nodiscard]] QuantumDevice *loadDevice(std::string filename)
-    {
-        init_rtd_dylib = std::make_unique<SharedLibraryManager>(filename);
-        std::string factory_name{zero_rtd_name + "Factory"};
-        void *f_ptr = init_rtd_dylib->getSymbol(factory_name);
-        return f_ptr ? reinterpret_cast<decltype(GenericDeviceFactory) *>(f_ptr)(zero_rtd_kwargs)
-                     : nullptr;
-    }
-
-    [[nodiscard]] auto addDevice(std::string rtd_lib, std::string rtd_name = {},
-                                 std::string rtd_kwargs = {})
+    [[nodiscard]] auto addDevice(const std::string &rtd_lib, const std::string &rtd_name = {},
+                                 const std::string &rtd_kwargs = {})
         -> std::pair<size_t, std::shared_ptr<QuantumDevice>>
     {
[…]
