### 2023-12-05 23bd286
Apply suggestions from code review
Co-authored-by: Josh Izaac <josh146@gmail.com>
