### 2023-12-05 b0f9a3f
[Frontend] Enable Jax dynamic API and use `abstracted_axes` in qjit (#366)
**Context:** Enable Jax Dynamic API and add `abstracted_axes` top-level
argument

**Description of the Change:** 
* Enable `dynamic_shapes` option in JAX.
* Automatically deduce implicit parameters when calling the compiled
function
* Remove implicit parameters from wrapped function arguments.

**Benefits:**
* Jax dynamic API allows to compile tensors with variable dimension
sizes.
* `jax.numpy.ones/zeros/empy` etc could now be used in the compiled
programs.

**Drawbacks:**
* Accessing dynamically-shaped tensors inside 
[…]
