+++ frontend/catalyst/jax_tracer.py
@@ -57,19 +57,20 @@ from catalyst.utils.jax_extras import (
     DynamicJaxprTrace,
     DynamicJaxprTracer,
     PyTreeDef,
+    PyTreeRegistry,
     ShapedArray,
-    ShapeDtypeStruct,
     _abstractify,
     _input_type_to_tracers,
     convert_element_type,
     deduce_avals,
     eval_jaxpr,
+    jaxpr_remove_implicit,
     jaxpr_to_mlir,
-    pytree,
+    make_jaxpr2,
     sort_eqns,
+    tree_flatten,
     tree_structure,
     tree_unflatten,
-    unzip2,
     wrap_init,
 )
 
@@ -92,14 +93,13 @@ class Function:
         self.__name__ = fn.__name__
 
     def __call__(self, *args, **kwargs):
-        jaxpr, shape = jax.make_jaxpr(self.fn, return_shape=True)(*args)
-        shape_tree = tree_structure(shape)
+        jaxpr, _, out_tree = make_jaxpr2(self.fn)(*args)
 
         def _eval_jaxpr(*args):
             return jax.core.eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, *args)
 
         retval = func_p.bind(wrap_init(_eval_jaxpr), *args, fn=self)
-        return tree_unflatten(shape_tree, retval)
+        return tree_unflatten(out_tree, retval)
 
 
 KNOWN_NAMED_OBS = (qml.Identity, qml.PauliX, qml.PauliY, qml.PauliZ, qml.Hadamard)
@@ -322,31 +322,39 @@ def has_nested_tapes(op: Operation) -> bool:
     )
 
 
-def trace_to_mlir(func, *args, **kwargs):
+def trace_to_mlir(func, abstracted_axes, *args, **kwargs):
     """Lower a Python function into an MLIR module.
 
     Args:
         func: python function to be lowered
+        abstracted_axes: abstracted axes specification. Necessary for JAX to use dynamic tensor
+            sizes.
         args: arguments to ``func``
         kwargs: keyword arguments to ``func``
 
     Returns:
-        module: the MLIR module corresponding to ``func``
-        context: the MLIR context corresponding
-        jaxpr: the jaxpr corresponding to ``func``
-        shape: the shape of the return values in ``PyTreeDef``
+        MLIR module: the MLIR module corresponding to ``func``
+        MLIR context: the MLIR context
+        ClosedJaxpr: the Jaxpr program corresponding to ``func``
+        jax.OutputType: Jaxpr output type (a list of abstract values paired with
+                        explicintess flags).
+        PyTreeDef: PyTree-shape of the return values in ``PyTreeDef``
     """
 
-    # The compilation cache must be clear for each translation unit.
-    # Otherwise, MLIR functions which do not exist in the current translation unit will be assumed
-    # to exist if an equivalent python function is seen in the cache. This happens during testing or
-    # if we wanted to compile a single python function multiple times with different options.
+    # The compilation cache must be clear for each translation unit. Otherwise, MLIR functions
+    # which do not exist in the current translation unit will be assumed to exist if an equivalent
+    # python function is seen in the cache. This happens during testing or if we wanted to compile a
+    # single python function multiple times with different options.
     mlir_fn_cache.clear()
 
     with EvaluationContext(EvaluationMode.CLASSICAL_COMPILATION):
-        jaxpr, shape = jax.make_jaxpr(func, return_shape=True)(*args, **kwargs)
+        make_jaxpr_kwargs = {"abstracted_axes": abstracted_axes}
[…]
