### 2023-12-05 32197c1
[Frontend] `typing` in Python3.9 does not support `|` operator (#385)
**Context:** Python3.9 does not support `|` operator for types in the
`typing` module.

**Description of the Change:** Use `Union`.

**Benefits:** Tests for 3.9 succeed.
