### 2023-12-05 e58cc03
Factorizing ExpansionStrategy
