+++ frontend/catalyst/utils/jax_extras.py
@@ -16,7 +16,7 @@
 from __future__ import annotations
 
 from contextlib import ExitStack, contextmanager
-from typing import Any, Callable, Dict, Generator, List, Optional, Sequence, Set, Type, TypeVar
+from typing import (Any, Callable, Dict, Generator, List, Optional, Sequence, Set, Type, TypeVar)
 from copy import copy
 from dataclasses import dataclass
 
@@ -28,6 +28,7 @@ from jax._src.dispatch import jaxpr_replicas
 from jax._src.effects import ordered_effects as jax_ordered_effects
 from jax._src.interpreters.mlir import _module_name_regex
 from jax._src.interpreters.partial_eval import (
+    AbstractedAxesSpec,
     _input_type_to_tracers,
     infer_lambda_input_type,
     trace_to_jaxpr_dynamic2,
@@ -90,6 +91,10 @@ __all__ = (
     "ClosedJaxpr",
     "DynamicJaxprTrace",
     "DynamicJaxprTracer",
+    "ExpansionStrategy",
+    "for_loop_expansion_strategy",
+    "while_loop_expansion_strategy",
+    "default_expansion_strategy",
     "Jaxpr",
     "PyTreeDef",
     "PyTreeRegistry",
@@ -312,10 +317,13 @@ def expanded_fun(in_type, *args_expanded):
 
 @transformation_with_aux
 def expanded_fun2(static_args, *args_expanded):
-    (in_type, force_implicit_indbidx) = static_args
+    (in_type, expansion_strategy) = static_args
     args_collapsed = [a for a,(_,k) in zip(args_expanded, in_type) if k]
     res_flat = yield args_collapsed, {}
-    all_outs, out_sig = infer_output_type3(args_expanded, res_flat, force_implicit_indbidx=force_implicit_indbidx)
+    num_implicit_inputs = len([() for _,k in in_type if not k])
+    all_outs, out_sig = infer_output_type3(
+        args_expanded, res_flat, expansion_strategy, num_implicit_inputs
+    )
     yield all_outs, out_sig
 
 @dataclass
@@ -333,17 +341,17 @@ class OutputSignature:
     out_tree: Callable[[], PyTreeDef]
 
 
-def deduce_avals3(f: Callable, args, kwargs, abstracted_axes=None, force_implicit_indbidx=False):
+def deduce_avals3(f: Callable, args, kwargs, expansion_strategy):
     """Wraps the callable ``f`` into a WrappedFun container accepting collapsed flatten arguments
     and returning expanded flatten results. Calculate input abstract values and output_tree promise.
     The promise must be called after the resulting wrapped function is evaluated."""
     flat_args, in_tree = tree_flatten((args, kwargs))
-    flat_axes_specs = _flat_axes_specs(abstracted_axes, *args, **kwargs)
-    in_expanded_args, in_type = expand_args(flat_args, axes_specs=flat_axes_specs,
-                                            force_implicit_indbidx=force_implicit_indbidx)
+    # flat_axes_specs = _flat_axes_specs(abstracted_axes, *args, **kwargs)
+    in_expanded_args, in_type = expand_args(flat_args, #axes_specs=flat_axes_specs,
+                                            expansion_strategy=expansion_strategy)
     wf = wrap_init(f)
     wf, out_tree_promise = flatten_fun(wf, in_tree)
-    wf, out_sig_promise = expanded_fun2(wf, (in_type, force_implicit_indbidx))
+    wf, out_sig_promise = expanded_fun2(wf, (in_type, expansion_strategy))
     wf = annotate(wf, in_type)
     return (
         wf,
[…]
