### 2023-12-05 bd3e25f
[Tests] Fix extension in test case for apple targets. (#390)
**Context:** Apple uses dylib

**Description of the Change:** Use dylib
