### 2023-12-05 ed54237
Factorizing ExpansionStrategy, loops work
