+++ frontend/catalyst/pennylane_extensions.py
@@ -913,45 +913,42 @@ class ForLoop(HybridOp):
                 self.regions[0].arg_classical_tracers + [qreg_in],
                 expansion_strategy=for_loop_expansion_strategy()
             )[0]
+
+            nimplicit = len(arg_expanded_tracers) - len(self.regions[0].arg_classical_tracers) - 1
+
             res_expanded_tracers, _ = expand_results(
+                [],
                 arg_expanded_tracers,
                 self.regions[0].res_classical_tracers + [qreg_out],
-                expansion_strategy=for_loop_expansion_strategy()
+                expansion_strategy=for_loop_expansion_strategy(),
+                num_implicit_inputs=nimplicit
             )
             jaxpr, _, consts = ctx.frames[inner_trace].to_jaxpr2(res_expanded_tracers)
 
-        in_expanded_tracers = expand_args(op.in_classical_tracers,
-                                          expansion_strategy=for_loop_expansion_strategy())[0]
 
-        nimplicit = len(in_expanded_tracers) - len(op.in_classical_tracers)
+        in_expanded_tracers = [
+            *[trace.full_raise(c) for c in consts],
+            *expand_args(op.in_classical_tracers,
+                        expansion_strategy=for_loop_expansion_strategy())[0],
+            qrp.actualize()
+        ]
 
-        in_c_expanded_tracers = (
-            [trace.full_raise(c) for c in consts] +
-            in_expanded_tracers +
-            [qrp.actualize()]
-        )
-
-        print("BBBBBBBBBB")
-        print(jaxpr)
-        print("BBBBBBBBBB")
-
-        out_c_expanded_classical_tracers = (
+        out_expanded_classical_tracers = (
             expand_results(
-                in_c_expanded_tracers,
+                [],
+                in_expanded_tracers,
                 self.out_classical_tracers,
                 expansion_strategy=for_loop_expansion_strategy(),
+                num_implicit_inputs=nimplicit
             )[0]
         )
-
-        print("OUT_C_EXPANDED_CLASSICAL_TRACERS")
-        for t in out_c_expanded_classical_tracers: print(t)
 
         qrp2 = QRegPromise(
             op.bind_overwrite_classical_tracers2(
                 ctx,
                 trace,
-                in_expanded_tracers=in_c_expanded_tracers,
-                out_expanded_tracers=out_c_expanded_classical_tracers,
+                in_expanded_tracers=in_expanded_tracers,
+                out_expanded_tracers=out_expanded_classical_tracers,
                 body_jaxpr=ClosedJaxpr(convert_constvars_jaxpr(jaxpr), ()),
                 body_nconsts=len(consts),
                 apply_reverse_transform=self.apply_reverse_transform,
@@ -1018,9 +1015,11 @@ class WhileLoop(HybridOp):
         cond_trace = self.regions[0].trace
         with EvaluationContext.frame_tracing_context(ctx, cond_trace):
             arg_expanded_classical_tracers = expand_args(
-                self.regions[0].arg_classical_tracers
+                self.regions[0].arg_classical_tracers,
+                expansion_strategy=while_loop_expansion_strategy()
             )[0]
             res_expanded_classical_tracers, out_type = expand_results(
+                [],
                 arg_expanded_classical_tracers,
                 self.regions[0].res_classical_tracers,
                 expansion_strategy=while_loop_expansion_strategy()
@@ -1031,6 +1030,7 @@ class WhileLoop(HybridOp):
             )
             assert unzip2(out_type)[0] == unzip2(out_type2)[0], f"\n{out_type=}\n{out_type2=}"
 
+        nimplicit = len(arg_expanded_classical_tracers) - len(self.regions[0].arg_classical_tracers)
         body_trace = self.regions[1].trace
[…]
