### 2023-12-05 3f52c5c
add expval shots
