+++ runtime/lib/backend/lightning/lightning_dynamic/LightningSimulator.cpp
@@ -172,12 +172,10 @@ auto LightningSimulator::HamiltonianObservable(const std::vector<double> &coeffs
     return this->obs_manager.createHamiltonianObs(coeffs, obs);
 }
 
-auto LightningSimulator::Expval(ObsIdType obsKey, size_t shots=0, std::vector<size_t>& shots_range = {}) -> double
+auto LightningSimulator::Expval(ObsIdType obsKey) -> double
 {
     RT_FAIL_IF(!this->obs_manager.isValidObservables({obsKey}),
                "Invalid key for cached observables");
-    RT_FAIL_IF(short_range.size()>shots,
-               "Short range size is larger than number of shots");
     auto &&obs = this->obs_manager.getObservable(obsKey);
 
     // update tape caching
@@ -187,11 +185,25 @@ auto LightningSimulator::Expval(ObsIdType obsKey, size_t shots=0, std::vector<si
 
     Pennylane::LightningQubit::Measures::Measurements<StateVectorT> m{*(this->device_sv)};
 
-    if(shots == 0){
-        return m.expval(*obs);
+    return m.expval(*obs);
+}
+
+auto LightningSimulator::Expval(ObsIdType obsKey, const size_t shots,
+                                const std::vector<size_t> &shot_range = {}) -> double
+{
+    RT_FAIL_IF(!this->obs_manager.isValidObservables({obsKey}),
+               "Invalid key for cached observables");
+    RT_FAIL_IF(shot_range.size() > shots, "Shot range size is larger than number of shots");
+    auto &&obs = this->obs_manager.getObservable(obsKey);
+
+    // update tape caching
+    if (this->tape_recording) {
+        this->cache_manager.addObservable(obsKey, MeasurementsT::Expval);
     }
-    return m.expval(*obs, shots, shots_range);
-        
+
+    Pennylane::LightningQubit::Measures::Measurements<StateVectorT> m{*(this->device_sv)};
+
+    return m.expval(*obs, shots, shot_range);
 }
 
 auto LightningSimulator::Var(ObsIdType obsKey) -> double
