### 2023-12-05 ec47014
add expval(namedObs, shot)
