### 2023-12-06 fd97ece
Improve removal of quantum ops in gradients
