+++ mlir/lib/Gradient/Transforms/GradMethods/PS_QuantumGradient.cpp
@@ -290,13 +290,18 @@ func::FuncOp ParameterShiftLowering::genQGradFunction(PatternRewriter &rewriter,
 
         // Finally erase all quantum operations.
         gradientFn.walk([&](Operation *op) {
-            if (auto gate = dyn_cast<quantum::QuantumGate>(op)) {
+            if (isa<quantum::DeviceOp>(op)) {
+                rewriter.eraseOp(op);
+            }
+            else if (auto gate = dyn_cast<quantum::QuantumGate>(op)) {
                 // We are undoing the def-use chains of this gate's return values
                 // so that we can safely delete it (all quantum ops must be eliminated).
                 rewriter.replaceOp(gate, gate.getQubitOperands());
             }
-            else if (isa<quantum::DeviceOp>(op)) {
-                // Erase redundant device specifications.
+            else if (auto region = dyn_cast<quantum::QuantumRegion>(op)) {
+                rewriter.replaceOp(op, region.getRegisterOperand());
+            }
+            else if (isa<quantum::DeallocOp>(op)) {
                 rewriter.eraseOp(op);
             }
         });
