### 2023-12-06 5434802
Eliminate compiler warning about unused variable
