### 2023-12-06 39d02ca
Simplify while-loop classical tracing
