### 2023-12-06 da16051
Adding support for dynamic API to for-loop
