### 2023-12-06 23effad
add more unit tests
