### 2023-12-06 6359f1c
Add dimension sharing user option in while-loops
