+++ frontend/catalyst/pennylane_extensions.py
@@ -80,7 +80,7 @@ from catalyst.utils.jax_extras import (
     unzip2,
 )
 from catalyst.utils.patching import Patcher
-from catalyst.utils.runtime import extract_backend_info
+from catalyst.utils.runtime import extract_backend_info, get_lib_path
 from catalyst.utils.toml import toml_load
 
 
@@ -116,10 +116,29 @@ class QFunc:
         self.device = device
         update_wrapper(self, fn)
 
+    @staticmethod
+    def _add_toml_file(device):
+        """Temporary function. This function adds the config field to devices.
+        TODO: Remove this function when `qml.Device`s are guaranteed to have their own
+        config file field."""
+        device_lpath = pathlib.Path(get_lib_path("runtime", "RUNTIME_LIB_DIR"))
+        # We need this replace because the toml file uses
+        # lightning_kokkos.toml and lightning_qubit.toml
+        # instead of lightning.qubit.toml and lightning.kokkos.toml
+        # We should just make all of them similar.
+        name = device.name
+        if isinstance(device, qml.Device):
+            name = device.short_name
+
+        toml_file_name = name.replace(".", "_") + ".toml"
+        toml_file = device_lpath.parent / "lib" / "backend" / toml_file_name
+        device.config = toml_file
+
     def __call__(self, *args, **kwargs):
         qnode = None
         if isinstance(self, qml.QNode):
             qnode = self
+            QFunc._add_toml_file(self.device)
             device = QJITDevice(
                 self.device.shots, self.device.wires, *extract_backend_info(self.device)
             )
@@ -164,54 +183,101 @@ class QJITDevice(qml.QubitDevice):
     pennylane_requires = "0.1.0"
     version = "0.0.1"
     author = ""
-    operations = [
-        "MidCircuitMeasure",
-        "Cond",
-        "WhileLoop",
-        "ForLoop",
-        "PauliX",
-        "PauliY",
-        "PauliZ",
-        "Hadamard",
-        "Identity",
-        "S",
-        "T",
-        "PhaseShift",
-        "RX",
-        "RY",
-        "RZ",
-        "CNOT",
-        "CY",
-        "CZ",
-        "SWAP",
-        "IsingXX",
-        "IsingYY",
-        "IsingXY",
-        "IsingZZ",
-        "ControlledPhaseShift",
-        "CRX",
-        "CRY",
-        "CRZ",
-        "CRot",
-        "CSWAP",
-        "MultiRZ",
-        "QubitUnitary",
-        "Adjoint",
-    ]
-    observables = [
-        "Identity",
-        "PauliX",
-        "PauliY",
-        "PauliZ",
-        "Hadamard",
-        "Hermitian",
-        "Hamiltonian",
-    ]
+    # Even though these are empty, these are needed.
+    # Otherwise, the following error is raised:
+    #
+    #   TypeError: Can't instantiate abstract class QJITDevice with abstract method operations
+    operations = []
+    observables = []
+
+    @staticmethod
+    def _check_qjit_compatible(config):
+        """Check the qjit_compatible record in the configuration file."""
+        with open(config, "rb") as f:
+            spec = toml_load(f)
+
+        if not spec["compilation"]["qjit_compatible"]:
+            msg = "Attempting to compile to device without qjit support."
+            raise CompileError(msg)
+
+    @staticmethod
+    def _check_mid_circuit_measurement(config):
+        with open(config, "rb") as f:
+            spec = toml_load(f)
+
+        return spec["compilation"]["mid_circuit_measurement"]
+
+    @staticmethod
+    def _check_control_flow(config):
+        with open(config, "rb") as f:
+            spec = toml_load(f)
+
[…]
