+++ runtime/lib/backend/lightning/lightning_kokkos/lightning_kokkos.toml
@@ -43,11 +43,7 @@ native = [
         "IsingYY",
         "IsingZZ",
         "IsingXY",
-        # Adjoint'ed gates
-        # which just need the inverse flag set
-        # are considered natively supported gates.
-        "Adjoint(S)",
-        "Adjoint(T)",
+        "Identity",
 ]
 # decomp gates are those which are neither supported natively
 # and are not executed via the applyMatrix operation. Technically
@@ -59,13 +55,6 @@ decomp = [
         "SISWAP",
         "SQISW",
         "CPhase",
-        # Adjoint'ed gates
-        # which are not in native set must either be decomposed
-        # or applied via matrix.
-        "Adjoint(SX)",
-        "Adjoint(ISWAP)",
-        "Adjoint(SISWAP)",
-        "Identity",
         "BasisState",
         "QubitStateVector",
         "StatePrep",
@@ -82,6 +71,11 @@ decomp = [
         "OrbitalRotation",
         "QFT",
         "ECR",
+        "Adjoint(SX)",
+        "Adjoint(ISWAP)",
+        "Adjoint(SISWAP)",
+        "Adjoint(S)",
+        "Adjoint(T)",
 ]
 # QubitUnitary should never be here.
 # Gates in "matrix" are decomposed to QubitUnitary.
@@ -116,5 +110,7 @@ finiteshots = [
 [compilation]
 qjit_compatible = true
 control_flow = true
+adjoint = true
+quantum_control = false
 mid_circuit_measurement = true
 dynamic_qubit_management = false # true after merging `StateVectorLQubitDynamic` to the monorepo
