### 2023-12-06 f9d24b4
Apply code review suggestion
