### 2023-12-06 b9cd726
Add verification fn for a "QuantumFree" attribute
