### 2023-12-06 0858a6f
Enforce "QuantumFree" on cloned QNode functions
