+++ runtime/lib/backend/openqasm/braket.local.qubit.toml
@@ -9,54 +9,114 @@ observables = [
 	"NamedObs",
 	"HermitianObs",
 	"TensorObs",
-	"HamiltonianObs",
 ]
 
+# These are gates which are defined in Quantum Instruction Set.
 [[operations.gates]]
-full = [
-	"Identity",
-	"PauliX",
-	"PauliY",
-	"PauliZ",
-	"Hadamard",
-	"S",
-	"T",
-	"CNOT",
-	"CY",
-	"CZ",
-	"SWAP",
-	"PhaseShift",
-	"RX",
-	"RY",
-	"RZ",
-	"CSWAP",
-	"PSWAP",
-	"ISWAP",
-	"Toffoli",
+native = [
+        "PauliX",
+        "PauliY",
+        "PauliZ",
+        "S",
+        "T",
+        "CNOT",
+        "CZ",
+        "SWAP",
+        "PhaseShift",
+        "RX",
+        "RY",
+        "RZ",
+        "CSWAP",
 ]
+# decomp gates are those which are neither supported natively
+# and are not executed via the applyMatrix operation. Technically
+# they could also be decomposed via applyMatrix.
+decomp = [
+	# TODO: These ones should be moved to Native once
+	# this issue is fixed: https://github.com/PennyLaneAI/catalyst/issues/395
+        "ISWAP",
+        "PSWAP",
+
+	# These ones are fine here.
+        "QubitUnitary",
+        "MultiRZ",
+        "Hadamard",
+        "Toffoli",
+        "CY",
+        "ControlledPhaseShift",
+        "Rot",
+        "CRX",
+        "CRY",
+        "CRZ",
+        "CRot",
+        "Identity",
+        "IsingXX",
+        "IsingYY",
+        "IsingZZ",
+        "IsingXY",
+        "SX",
+        "SISWAP",
+        "SQISW",
+        "CPhase",
+        # Adjoint'ed gates
+        # which are not in native set must either be decomposed
+        # or applied via matrix.
+        "Adjoint(SX)",
+        "Adjoint(ISWAP)",
+        "Adjoint(SISWAP)",
+        "BasisState",
+        "QubitStateVector",
+        "StatePrep",
+        "ControlledQubitUnitary",
+        "DiagonalQubitUnitary",
+        "SingleExcitation",
+        "SingleExcitationPlus",
+        "SingleExcitationMinus",
+        "DoubleExcitation",
+        "DoubleExcitationPlus",
+        "DoubleExcitationMinus",
+        "QubitCarry",
+        "QubitSum",
+        "OrbitalRotation",
+        "QFT",
+        "ECR",
+        "Adjoint(S)",
+        "Adjoint(T)",
+]
+# QubitUnitary should never be here.
+# Gates in "matrix" are decomposed to QubitUnitary.
+# Having a QubitUnitary here would mean an infinite loop.
+
+# These are gates that appear on the list of supported gates
+# but are not in the list of Quantum Gate Set Instructions in
+# the Catalyst repository.
+# One disadvantage of matrix gates is that they will be lowered to
+# QubitUnitary and QubitUnitary is not differentiable.
+# TODO: Move some of these to native once we add support for them
+# in Catalyst runtime.
 matrix = [
-	# QubitUnitary should never be here.
-	# Gates in "matrix" are decomposed to QubitUnitary.
-	# Having a QubitUnitary here would mean an infinite loop.
-	"OrbitalRotation",
-	"MultiControlledX",
-	# TODO:
-	# Why is qml.ops.Controlled not in qml namespace?
-	# TODO:
-	# other gates supported via applying matrices
+        "MultiControlledX",
 ]
 
 [measurements]
[…]
