+++ runtime/tests/Test_OpenQasmDevice.cpp
@@ -517,13 +517,12 @@ TEST_CASE("Test __quantum__rt__device_init registering the OpenQasm device", "[C
 {
     __quantum__rt__initialize();
 
-    char dev[8] = "rtd_lib";
-    char dev_value[30] = "braket.aws.qubit";
+    char device_aws[30] = "braket.aws.qubit";
 
 #if __has_include("OpenQasmDevice.hpp")
-    __quantum__rt__device_init((int8_t *)dev, (int8_t *)dev_value);
+    __quantum__rt__device_init((int8_t *)device_aws, nullptr, nullptr);
 #else
-    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)dev, (int8_t *)dev_value),
+    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)device_aws, nullptr, nullptr),
                         Catch::Contains("cannot open shared object file"));
 #endif
 
@@ -531,29 +530,12 @@ TEST_CASE("Test __quantum__rt__device_init registering the OpenQasm device", "[C
 
     __quantum__rt__initialize();
 
-    char dev_kwargs[20] = "rtd_kwargs";
-    char dev_value_kwargs[70] = "device_arn : arn:aws:braket:::device/quantum-simulator/amazon/sv1";
-
-    __quantum__rt__device_init((int8_t *)dev_kwargs, (int8_t *)dev_value_kwargs);
+    char device_local[30] = "braket.local.qubit";
 
 #if __has_include("OpenQasmDevice.hpp")
-    __quantum__rt__device_init((int8_t *)dev, (int8_t *)dev_value);
+    __quantum__rt__device_init((int8_t *)device_local, nullptr, nullptr);
 #else
-    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)dev, (int8_t *)dev_value),
-                        Catch::Contains("cannot open shared object file:"));
-#endif
-
-    __quantum__rt__finalize();
-
-    __quantum__rt__initialize();
-
-    char dev_lcl[8] = "rtd_lib";
-    char dev_value_lcl[30] = "braket.local.qubit";
-
-#if __has_include("OpenQasmDevice.hpp")
-    __quantum__rt__device_init((int8_t *)dev_lcl, (int8_t *)dev_value_lcl);
-#else
-    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)dev_lcl, (int8_t *)dev_value_lcl),
+    REQUIRE_THROWS_WITH(__quantum__rt__device_init((int8_t *)(int8_t *), nullptr, nullptr),
                         Catch::Contains("cannot open shared object file"));
 #endif
 
