### 2023-12-06 9fa70e0
Add qml.QubitUnitary to local. Remove unnecessary test.
