+++ setup.py
@@ -31,7 +31,7 @@ requirements = [
     "pennylane>=0.32",
     f"jax=={jax_version}",
     f"jaxlib=={jax_version}",
-    "tomlkit",
+    "tomlkit;python_version<'3.11'",
 ]
 
 classifiers = [
