+++ frontend/catalyst/utils/runtime.py
@@ -106,8 +106,9 @@ def extract_backend_info(device):
     with open(device.config, "rb") as f:
         spec = toml_load(f)
 
-    for k, v in spec["options"].items():
-        if hasattr(device, v):
-            device_kwargs[k] = getattr(device, v)
+    if "options" in spec.keys():
+        for k, v in spec["options"].items():
+            if hasattr(device, v):
+                device_kwargs[k] = getattr(device, v)
 
     return device_name, device_lpath, device_kwargs, device.config
