### 2023-12-07 7d39067
Fix minor failures in the existing tests
