### 2023-12-07 94b757e
Refactor call tree verification for Grad
The primitives for Function and Grad have been changed to bind the
callee, instead of the object instance itself.
Additionally, the call tree traversal of the callee has been improved to
be more robust.

No (external) functional changes.
