### 2023-12-07 5b0caed
tensorprod namedobs
