### 2023-12-07 88bd0e2
[Frontend] Refactor call tree verification for Grad (#399)
The primitives for Function and Grad have been changed to bind the
callee, instead of the object instance itself.
Additionally, the call tree traversal of the callee has been improved to
be more robust.

No (external) functional changes.

---------

Co-authored-by: Erick <erick.ochoalopez@xanadu.ai>
