### 2023-12-07 761f68c
Apply suggestions from code review
Co-authored-by: David Ittah <dime10@users.noreply.github.com>
