### 2023-12-08 3778174
Fix changelog formatting
