### 2023-12-08 e928054
QubitUnitary is a native gate for some devices.
