+++ runtime/lib/backend/lightning/lightning_dynamic/lightning_qubit.toml
@@ -6,93 +6,93 @@ precision = ['float32', 'float64']
 
 [operations]
 observables = [
-	"NamedObs",
-	"HermitianObs",
-	"TensorObs",
-	"HamiltonianObs",
-]
-
-# These are gates which are defined in Quantum Instruction Set.
-[[operations.gates]]
-native = [
-        "QubitUnitary",
         "PauliX",
         "PauliY",
         "PauliZ",
-        "MultiRZ",
+        "Hadamard",
+        "Hermitian",
+        "Identity",
+        "Projector",
+        "SparseHamiltonian",
+        "Hamiltonian",
+        "Sum",
+        "SProd",
+        "Prod",
+        "Exp",
+]
+
+# These are gates that do not need decomposition.
+# Also gates which are natively supported by the device.
+[[operations.gates]]
+native = [
+        "Identity",
+        "PauliX",
+        "PauliY",
+        "PauliZ",
         "Hadamard",
         "S",
         "T",
-        "CNOT",
-        "SWAP",
-        "CSWAP",
-        "Toffoli",
-        "CY",
-        "CZ",
         "PhaseShift",
-        "ControlledPhaseShift",
         "RX",
         "RY",
         "RZ",
         "Rot",
+        "CNOT",
+        "CY",
+        "CZ",
+        "SWAP",
+        "CSWAP",
+        "Toffoli",
+        "IsingXX",
+        "IsingXY",
+        "IsingYY",
+        "IsingZZ",
+        "ControlledPhaseShift",
         "CRX",
         "CRY",
         "CRZ",
         "CRot",
-        "Identity",
-        "IsingXX",
-        "IsingYY",
-        "IsingZZ",
-        "IsingXY",
-]
-# decomp gates are those which are neither supported natively
-# and are not executed via the applyMatrix operation. Technically
-# they could also be decomposed via applyMatrix.
-decomp = [
-        "SX",
-        "ISWAP",
-        "PSWAP",
-        "SISWAP",
-        "SQISW",
-        "CPhase",
-        # Adjoint'ed gates
-        # which are not in native set must either be decomposed
-        # or applied via matrix.
-        "Adjoint(SX)",
-        "Adjoint(ISWAP)",
-        "Adjoint(SISWAP)",
-        "BasisState",
-        "QubitStateVector",
-        "StatePrep",
-        "ControlledQubitUnitary",
-        "DiagonalQubitUnitary",
         "SingleExcitation",
         "SingleExcitationPlus",
         "SingleExcitationMinus",
         "DoubleExcitation",
         "DoubleExcitationPlus",
         "DoubleExcitationMinus",
+        "MultiRZ",
+	# QubitUnitary is technically
+	# "applyMatrix", but it doesn't
+	# belong in matrix gates.
+        "QubitUnitary",
+]
+
+# These are gates which need to be decomposed to QubitUnitary.
+matrix = [
+	# All adjoint operations in lightning_qubit
+	# are applied via applyMatrix.
+        "Adjoint(S)",
+        "Adjoint(T)",
+        "Adjoint(SX)",
+        "Adjoint(ISWAP)",
+        "Adjoint(SISWAP)",
+
+	# The rest.
+        "BasisState",
+        "ECR",
+        "QubitStateVector",
+        "StatePrep",
+        "ControlledQubitUnitary",
+        "MultiControlledX",
+        "DiagonalQubitUnitary",
+        "SX",
+        "ISWAP",
+        "PSWAP",
+        "SISWAP",
+        "SQISW",
+        "CPhase",
         "QubitCarry",
         "QubitSum",
         "OrbitalRotation",
         "QFT",
-        "ECR",
-        "Adjoint(S)",
-        "Adjoint(T)",
-]
[…]

[…]
