### 2023-12-08 a6efe74
Apply code review suggestions
