+++ mlir/lib/Quantum/Transforms/ConversionPatterns.cpp
@@ -106,36 +106,29 @@ struct DeviceOpPattern : public OpConversionPattern<DeviceOp> {
         MLIRContext *ctx = this->getContext();
         ModuleOp mod = op->getParentOfType<ModuleOp>();
 
-        auto specs = op.getSpecs(); // maybe {} for using the default backend device
-        if (specs && !specs->empty()) {
-            const auto args = specs.value();
-            StringRef qirName =
-                "__quantum__rt__device_init"; // (int8_t *, int8_t *, int8_t *) -> void
+        StringRef qirName = "__quantum__rt__device_init"; // (int8_t *, int8_t *, int8_t *) -> void
 
-            Type charPtrType = LLVM::LLVMPointerType::get(IntegerType::get(ctx, 8));
-            Type qirSignature = LLVM::LLVMFunctionType::get(LLVM::LLVMVoidType::get(ctx),
-                                                            {/* rtd_lib = */ charPtrType,
-                                                             /* rtd_name = */ charPtrType,
-                                                             /* rtd_kwargs = */ charPtrType});
-            LLVM::LLVMFuncOp fnDecl =
-                ensureFunctionDeclaration(rewriter, op, qirName, qirSignature);
+        Type charPtrType = LLVM::LLVMPointerType::get(IntegerType::get(ctx, 8));
+        Type qirSignature = LLVM::LLVMFunctionType::get(LLVM::LLVMVoidType::get(ctx),
+                                                        {/* rtd_lib = */ charPtrType,
+                                                         /* rtd_name = */ charPtrType,
+                                                         /* rtd_kwargs = */ charPtrType});
+        LLVM::LLVMFuncOp fnDecl = ensureFunctionDeclaration(rewriter, op, qirName, qirSignature);
 
-            auto rtd_lib = args[0].cast<StringAttr>().getValue().str();
-            auto rtd_name = args[1].cast<StringAttr>().getValue().str();
-            auto rtd_kwargs = args[2].cast<StringAttr>().getValue().str();
+        auto rtd_lib = op.getLib().str();
+        auto rtd_name = op.getName().str();
+        auto rtd_kwargs = op.getKwargs().str();
 
-            auto rtd_lib_gs = getGlobalString(
-                loc, rewriter, rtd_lib, StringRef(rtd_lib.c_str(), rtd_lib.length() + 1), mod);
-            auto rtd_name_gs = getGlobalString(
-                loc, rewriter, rtd_name, StringRef(rtd_name.c_str(), rtd_name.length() + 1), mod);
-            auto rtd_kwargs_gs =
-                getGlobalString(loc, rewriter, rtd_kwargs,
-                                StringRef(rtd_kwargs.c_str(), rtd_kwargs.length() + 1), mod);
+        auto rtd_lib_gs = getGlobalString(loc, rewriter, rtd_lib,
+                                          StringRef(rtd_lib.c_str(), rtd_lib.length() + 1), mod);
+        auto rtd_name_gs = getGlobalString(loc, rewriter, rtd_name,
+                                           StringRef(rtd_name.c_str(), rtd_name.length() + 1), mod);
+        auto rtd_kwargs_gs = getGlobalString(
+            loc, rewriter, rtd_kwargs, StringRef(rtd_kwargs.c_str(), rtd_kwargs.length() + 1), mod);
 
-            SmallVector<Value> operands = {rtd_lib_gs, rtd_name_gs, rtd_kwargs_gs};
+        SmallVector<Value> operands = {rtd_lib_gs, rtd_name_gs, rtd_kwargs_gs};
 
-            rewriter.create<LLVM::CallOp>(loc, fnDecl, operands);
-        }
+        rewriter.create<LLVM::CallOp>(loc, fnDecl, operands);
 
         rewriter.eraseOp(op);
 
