+++ frontend/catalyst/jax_primitives.py
@@ -502,11 +502,7 @@ def _qdevice_lowering(jax_ctx: mlir.LoweringRuleContext, rtd_lib, rtd_name, rtd_
     ctx = jax_ctx.module_context.context
     ctx.allow_unregistered_dialects = True
 
-    DeviceOp(
-        specs=ir.ArrayAttr.get(
-            [ir.StringAttr.get(rtd_lib), ir.StringAttr.get(rtd_name), ir.StringAttr.get(rtd_kwargs)]
-        )
-    )
+    DeviceOp(ir.StringAttr.get(rtd_lib), ir.StringAttr.get(rtd_name), ir.StringAttr.get(rtd_kwargs))
 
     return ()
 
