### 2023-12-08 fc848a3
Add a templated macro for generating the device entry-points
