### 2024-05-01 #6012 @danielskatz
👋 @nathanshammah, @travishumble, @AndrewLitteken - would a couple of you be willing to review this submission for [JOSS](https://joss.theoj.org)? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
