### 2024-05-16 #6720 @danielskatz
👋 @pmcao - I know it's just been two weeks at this point, but it would be great if you could use the command `@editorialbot generate my checklist` to create your review checklist. @editorialbot commands need to be the first thing in a new comment.

Once you do this, you should be able to check off the first two items fairly quickly.
