### 2024-06-05 #6720 @otbrown
Hi @josh146,

Sorry you've been waiting so long for this review! Catalyst itself looks great, but as above, I have some minor comments relating to the write-up for JOSS. I've ordered my comments from most to least trivial, but I don't think any of them entail significant work to address. Please let me know if you have any questions.

- A space is missing before the citation on line 49.
- The reference on line 131 (Lattner & Adve 2004) lists the conference as 'Cgo' rather than 'International Symposium on Code Generation and Optimization (CGO)' as Lattner et al. 2021.
- Code examples -- while there are many good example and tutorial codes provided in the documentation and in the repository, the code snippets provided in the paper feel incomplete.
  - The `hybrid_function(x)` example does not include code to execute the function, or an example of inputs and expected outputs.
  - The `workflow(init_params)` example _does_ include code to execute the function, and an example of inputs, but does not include expected outputs.
  - The `workflow(init_params)` example does not include all required imports, as it relies on you having already executed the `hybrid_function(x)` code snippet, and retained those imports.
  - The `workflow(init_params)` example seems to only run when a GPU is present, though I could believe this is just an issue with my local environment! If it _should_ only run on GPU, I would propose explicitly saying this.
  - In the interests of reproducibility, I would recommend changing the example code snippets in the paper to be complete such that they could be copied and pasted into a python script, and include inputs and expected outputs.
