### 2024-06-12 #6720 @josh146
Thanks @danielskatz @otbrown! I've implemented all code review recommendations here: https://github.com/PennyLaneAI/catalyst/pull/807. Will comment again as soon as it is merged 🙂

I've also directly updated the paper for the following two bullet points:

> - A space is missing before the citation on line 49.
> - The reference on line 131 (Lattner & Adve 2004) lists the conference as 'Cgo' rather than 'International Symposium on Code Generation and Optimization (CGO)' as Lattner et al. 2021.
