### 2024-06-20 77c2ee7
Review suggestions for the Catalyst paper (#807)
See
https://github.com/openjournals/joss-reviews/issues/6720#issuecomment-2150538647
