### 2024-06-26 #6720 @pmcao
@josh146

This is a great effort to bridge classical and quantum runtime. Here are the comments:

Write up:

- I am curious how Catalyst handles faults in quantum hardware? If there are examples of faults and the code like exception handling that'd be great. Since the paper claims "advanced quantum programming features essential for fault-tolerant hardware support" more descriptions in fault handling is needed.

- The Catalyst documentation seems extensive. There are four examples, however, I couldn't get the taxonomy of how those examples related together. Coming from signal processing background, can you add an example of an HMM implemented in Catalyst with quantum hardware, since HMM is widely used for inference in many of the domains.

Misc:
- Link to catalyst is broken ({[https://github.com/PennyLaneAI/catalyst}](https://github.com/PennyLaneAI/catalyst%7D),)

I have a few more that I'd like to add before the end of this week. Hope those help with the documentation clarity and adoption of Catalyst!
