### 2024-06-26 #6720 @josh146
Hi @pmcao, thanks for your review!

> I am curious how Catalyst handles faults in quantum hardware? If there are examples of faults and the code like exception handling that'd be great. Since the paper claims "advanced quantum programming features essential for fault-tolerant hardware support" more descriptions in fault handling is needed.

There is a slight subtlety in this phrase here I think. With the Catalyst framework, we are more focused (at the moment) on enabling research that is focused on exploring, building, analyzing/executing algorithms *designed for* fault-tolerant quantum hardware. That is, the programming approach and features you need for algorithms targeting fault-tolerant quantum hardware is different to the features you need for targeting near-term quantum hardware!

What we don't currently support in Catalyst is implementing direct support for error correction (that is, compiling the algorithm down to physical hardware using error correction schemes that would correct against hardware faults). This is a very active field, and something we are likely to explore in the future :)

> The Catalyst documentation seems extensive. There are four examples, however, I couldn't get the taxonomy of how those examples are related together.

I can briefly run through some of these examples:

- [Magic State Distillation](https://docs.pennylane.ai/projects/catalyst/en/latest/demos/magic_distillation_demo.html): This is an example of a workflow/algorithm that is common for researchers who are exploring error correction and fault-tolerant quantum computing.

- [Variational Quantum Eigensolver](https://docs.pennylane.ai/projects/catalyst/en/latest/demos/adaptive_circuits_demo.html): This is a quantum chemistry-focused example (using quantum computing to find the ground state of a molecule). Unlike the example above, this is an example of a hybrid algorithm executing on near-term (non-error corrected) hardware. That is, it involves a classical optimization loop around quantum computation.

- [QML Optimization with Optax](https://docs.pennylane.ai/projects/catalyst/en/latest/demos/qml_optimization.html): Similar to the example above (a hybrid, near term algorithm), however this one is more quantum-machine learning focused, and showcases Catalyst's integration with the JAX machine learning framework.

So the examples aren't thematically related, apart from being different (and common!) workflows you see across different research areas in quantum computing.

>  Coming from signal processing background, can you add an example of an HMM implemented in Catalyst with quantum hardware, since HMMs are widely used for inference in many of the domains.

Unfortunately, myself and the team are not familiar with HMMs (or their quantum algorithm equivalents), so likely would not be able to provide this example!
