### 2024-07-01 #6720 @danielskatz
@pmcao - can you respond to @josh146 ☝️ ?  Also, you previously mentioned

> I have a few minor comments that I'd like to add before the end of this week. Hope those help with the documentation clarity and adoption of Catalyst!

Has this happened?
