### 2024-07-01 PennyLaneAI/catalyst#887 @pmcao
openjournals/joss-reviews#6720
openjournals/joss-reviews#6720

Comment:

On balance, the paper has motivated the need, described three critical pillars of the JIT compiler infrastructure, and made significant open-source contributions.

My takeaway from reviewing this paper is that fault-tolerant quantum computing has to be supported end-to-end in both hardware, mid-circuit, and classical processing because the fault can manifest at any part of the pipeline, for example, Silent Data Corruption in the classical processing part. As the author mentioned, "becoming clear that we cannot separate classical and quantum parts of the program".

This JOSS paper would benefit from a Discussion / Future Work in which the authors, with their expertise, could point to future directions in this area.
