### 2024-07-01 PennyLaneAI/catalyst#887 @pmcao
This is a minor issue that would need a few sentences to be resolved.

After that, the submission can be published.

This comment finalizes my review.
