### 2024-07-01 #6720 @pmcao
This minor issue would need a few sentences to be resolved.

After that, the submission can be published.

This comment finalizes my review.
