### 2024-07-01 #6720 @danielskatz
@josh146 - somehow, the discussion of https://github.com/PennyLaneAI/catalyst/issues/887 seems to have moved into this issue rather than that one. Can you make your comment there, and if you think that issue is resolved, close it.  Then you can use the command `@editorialbot generate pdf` here to make a new, up-to-date PDF. (@editorialbot commands need to the first thing in a comment.)
