### 2024-07-01 #6720 @danielskatz
At this point could you:

At the same time, I will proofread the submission and might suggest changes (noting that the changes will not need to be included in the archived version, where we focus on the code, since we will also publish the paper itself built from the final source)
