### 2024-07-01 #6720 @josh146
Thanks @danielskatz!

> Make a tagged release of your software, and list the version tag of the archived version here.

Do you mean a GitHub release, or simply a git tag? A github release is non-trivial at the moment, as we have a lot of continuous integration that would be triggered.

We have a new release of Catalyst (0.7.0) scheduled on July 10th, so another option is to wait for then.
