### 2024-07-01 #6720 @danielskatz
I meant a tag to the state of the software repo at the time that the review has finished, which could just be a tag, or could be a release. The reason for a release might be to use it to automatically make a deposit on Zenodo using the GitHub - Zenodo linkage for this, but you can make the archive however you want, including manually. Also, waiting until July 10 would be ok too.
