### 2024-07-01 #6720 @danielskatz
Actually, I went ahead and fixed the DOI in https://github.com/PennyLaneAI/catalyst/pull/889, as I saw another one to add too.
