### 2024-07-09 #6720 @josh146
Hi @danielskatz, our release is now done so I can complete the process:

> Make a tagged release of your software, and list the version tag of the archived version here.

The tag is [`v0.7.0`](https://github.com/PennyLaneAI/catalyst/tree/v0.7.0), and it also has an associated [release](https://github.com/PennyLaneAI/catalyst/releases/tag/v0.7.0).

> Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository). Please list the DOI of the archived version here.

The archived reviewed software is here: [10.5281/zenodo.12696447](https://zenodo.org/doi/10.5281/zenodo.12696447).
