### 2024-07-09 #6720 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
