### 2024-07-09 #6720 @danielskatz
Congratulations to @josh146 (Joshua Izaac) and co-authors on your publication!!

And thanks to @pmcao and @otbrown for reviewing!
JOSS depends on volunteers and couldn't succeed without you
